Atualizações nas configurações de busca, geração de links automatizados
