+++ Exemplos de Codigo-fonte/CSharp/IntegracaoInvoicy/ConsoleUI/Models/Documentos/Descarte.cs
@@ -1,0 +1,11 @@
+﻿namespace ConsoleUI.Models.Documentos
+{
+    public class Descarte
+    {
+        public string ModeloDocumento { get; set; }
+        public int? tpAmb { get; set; }
+        public string CNPJ_emit { get; set; }
+        public int Numero { get; set; }
+        public string Serie { get; set; }
+    }
+}
