+++ Exemplos de Codigo-fonte/CSharp/IntegracaoInvoicy/ConsoleUI/Services/Source.cs
@@ -5,75 +5,76 @@ namespace ConsoleUI.Services
     public class Source
     {
         public Dictionary<string, string> Empresa { get; set; }
-        public Dictionary<string, string> NFe { get; set; }
-        public Dictionary<string, string> NFCe { get; set; }
-        public Dictionary<string, string> MDFe { get; set; }
-        public Dictionary<string, string> NFSe { get; set; }
-        public Dictionary<string, string> CTe { get; set; }
+        public Dictionary<string, Dictionary<string, string>> Dados;
+        private Dictionary<string, string> NFe { get; set; }
+        private Dictionary<string, string> NFCe { get; set; }
+        private Dictionary<string, string> MDFe { get; set; }
+        private Dictionary<string, string> NFSe { get; set; }
+        private Dictionary<string, string> CTe { get; set; }
 
-        public string NFeEnvio = "[\r\n    {\r\n        \"Documento\": {\r\n            \"ModeloDocumento\": \"NFe\",\r\n            \"Versao\": 4,\r\n            \"Parametros\": {\r\n                \"ApelidoLogomarca\": \"\"\r\n            },\r\n            \"ide\": {\r\n                \"cNF\": 1512,\r\n                \"cUF\": 43,\r\n                \"natOp\": \"NFE CST 00\",\r\n                \"mod\": \"55\",\r\n                \"serie\": \"245\",\r\n                \"nNF\": 26,\r\n                \"dhEmi\": \"2021-02-22T15:14:21\",\r\n                \"fusoHorario\": \"-03:00\",\r\n                \"dhSaiEnt\": \"0000-00-00T00:00:00\",\r\n                \"tpNf\": 1,\r\n                \"idDest\": 1,\r\n                \"indFinal\": 0,\r\n                \"indPres\": 0,\r\n                \"cMunFg\": 4321808,\r\n                \"tpImp\": 1,\r\n                \"tpEmis\": 1,\r\n                \"tpAmb\": 2,\r\n                \"xJust\": \"\",\r\n                \"dhCont\": \"0000-00-00T00:00:00\",\r\n                \"finNFe\": 1,\r\n                \"EmailArquivos\": \"email@gmail.com\",\r\n                \"NumeroPedido\": \"\"\r\n            },\r\n            \"emit\": {\r\n                \"CNPJ_emit\": \"06354976000149\",\r\n                \"CPF_emit\": \"\",\r\n                \"xNome\": \"Migrate\",\r\n                \"xFant\": \"EMPRESA DE TESTE\",\r\n                \"IM\": \"\",\r\n                \"CNAE\": \"\",\r\n                \"IE\": \"1470049241\",\r\n                \"IEST\": \"\",\r\n                \"CRT\": 3,\r\n                \"enderEmit\": {\r\n                    \"xLgr\": \"LOGRADOURO\",\r\n                    \"nro\": \"1700\",\r\n                    \"xCpl\": \"ANDAR 02\",\r\n                    \"xBairro\": \"BAIRRO\",\r\n                    \"cMun\": 4321808,\r\n                    \"xMun\": \"ENCANTADO\",\r\n                    \"UF\": \"RS\",\r\n                    \"CEP\": 95960000,\r\n                    \"cPais\": 1058,\r\n                    \"xPais\": \"BRASIL\",\r\n                    \"fone\": 5137513322,\r\n                    \"fax\": 3335354830,\r\n                    \"Email\": \"\"\r\n                }\r\n            },\r\n            \"dest\": {\r\n                \"CNPJ_dest\": \"11635726000117\",\r\n                \"CPF_dest\": \"\",\r\n                \"idEstrangeiro\": \"\",\r\n                \"xNome_dest\": \"NF-E EMITIDA EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"IE_dest\": \"1060127838\",\r\n                \"ISUF\": \"\",\r\n                \"indIEDest\": 1,\r\n                \"IM_dest\": \"\",\r\n                \"enderDest\": {\r\n                    \"nro_dest\": \"67\",\r\n                    \"xCpl_dest\": \"CENTRO\",\r\n                    \"xBairro_dest\": \"CENTRO\",\r\n                    \"xEmail_dest\": \"email@gmail.com\",\r\n                    \"xLgr_dest\": \"LOGRADOURO\",\r\n                    \"xPais_dest\": \"Brasil\",\r\n                    \"cMun_dest\": 4309605,\r\n                    \"xMun_dest\": \"CONTAGEM\",\r\n                    \"UF_dest\": \"RS\",\r\n                    \"CEP_dest\": 32145900,\r\n                    \"cPais_dest\": 1058,\r\n                    \"fone_dest\": 5535485105\r\n                }\r\n            },\r\n            \"retirada\": {\r\n                \"CNPJ_ret\": \"\",\r\n                \"CPF_ret\": \"\",\r\n                \"xNome_ret\": \"\",\r\n                \"xLgr_ret\": \"\",\r\n                \"nro_ret\": \"\",\r\n                \"xCpl_ret\": \"\",\r\n                \"xBairro_ret\": \"\",\r\n                \"xMun_ret\": \"\",\r\n                \"cMun_ret\": 0,\r\n                \"UF_ret\": \"\",\r\n                \"CEP_ret\": 0,\r\n                \"cPais_ret\": 0,\r\n                \"xPais_ret\": \"\",\r\n                \"fone_ret\": 0,\r\n                \"email_ret\": \"\",\r\n                \"IE_ret\": \"\"\r\n            },\r\n            \"entrega\": {\r\n                \"CNPJ_entr\": \"\",\r\n                \"CPF_entr\": \"\",\r\n                \"xLgr_entr\": \"\",\r\n                \"nro_entr\": \"\",\r\n                \"xCpl_entr\": \"\",\r\n                \"xBairro_entr\": \"\",\r\n                \"cMun_entr\": 0,\r\n                \"xMun_entr\": \"\",\r\n                \"UF_entr\": \"\",\r\n                \"xNome_entr\": \"\",\r\n                \"CEP_entr\": 0,\r\n                \"cPais_entr\": 0,\r\n                \"xPais_entr\": \"\",\r\n                \"fone_entr\": 0,\r\n                \"email_entr\": \"\",\r\n                \"IE_entr\": \"\"\r\n            },\r\n            \"autXML\": [],\r\n            \"det\": [\r\n                {\r\n                    \"infADProd\": \"\",\r\n                    \"prod\": {\r\n                        \"cProd\": \"45093812504843\",\r\n                        \"cEAN\": \"SEM GTIN\",\r\n                        \"xProd\": \"MOTOCICLETA MARCA HONDA, TIPO CB 300 R ANO DE FABRICACAO 2009 ANO MODELO 2010\",\r\n                        \"NCM\": \"87112020\",\r\n                        \"EXTIPI\": \"87\",\r\n                        \"CFOP\": 5102,\r\n                        \"uCOM\": \"BOT\",\r\n                        \"qCOM\": \"1.0000\",\r\n                        \"vUnCom\": \"9975.2100000000\",\r\n                        \"vProd\": \"9975.21\",\r\n                        \"cEANTrib\": \"SEM GTIN\",\r\n                        \"uTrib\": \"BOT\",\r\n                        \"qTrib\": \"1.0000\",\r\n                        \"vUnTrib\": \"9975.2100000000\",\r\n                        \"vFrete\": \"0.00\",\r\n                        \"vSeg\": \"0.00\",\r\n                        \"vDesc\": \"0.00\",\r\n                        \"vOutro_item\": \"0.00\",\r\n                        \"indTot\": 1,\r\n                        \"nTipoItem\": 1,\r\n                        \"dProd\": 0,\r\n                        \"xPed_item\": \"\",\r\n                        \"nItemPed\": \"\",\r\n                        \"nFCI\": \"\",\r\n                        \"nRECOPI\": \"123456\",\r\n                        \"CEST\": \"\",\r\n                        \"cBenef\": \"RS051011\",\r\n                        \"indEscala\": \"\",\r\n                        \"CNPJFab\": \"\",\r\n                        \"NVEs\": [],\r\n                        \"detDI\": [],\r\n                        \"detExport\": [],\r\n                        \"veicProd\": {\r\n                            \"tpOp\": 1,\r\n                            \"chassi\": \"9C2NC4310AR033234\",\r\n                            \"cCor\": \"PRME\",\r\n                            \"xCor\": \"PRATA METALICO\",\r\n                            \"pot\": \"2600\",\r\n                            \"cilin\": \"0291\",\r\n                            \"PesoL\": \"000000143\",\r\n                            \"PesoB\": \"000000175\",\r\n                            \"nSerie\": \"0AR033234\",\r\n                            \"tpComb\": \"02\",\r\n                            \"nMotor\": \"NC43E1A033234\",\r\n                            \"CMT\": \"00179.000\",\r\n                            \"dist\": \"1402\",\r\n                            \"RENAVAM\": \"\",\r\n                            \"anoMod\": 2010,\r\n                            \"anoFab\": 2009,\r\n                            \"tpPint\": \"M\",\r\n                            \"tpVeic\": \"04\",\r\n                            \"espVeic\": 1,\r\n                            \"VIN\": \"N\",\r\n                            \"condVeic\": 1,\r\n                            \"cMod\": 2710,\r\n                            \"cCorDENATRAN\": \"04\",\r\n                            \"lota\": 2,\r\n                            \"tpRest\": 0\r\n                        },\r\n                        \"med\": [],\r\n                        \"arma\": [],\r\n                        \"comb\": {\r\n                            \"cProdANP\": 0,\r\n                            \"pMixGN\": 0,\r\n                            \"CODIF\": \"\",\r\n                            \"qTemp\": \"0.0000\",\r\n                            \"UFcons\": \"\",\r\n                            \"descANP\": \"\",\r\n                            \"pGLP\": 0,\r\n                            \"pGNn\": 0,\r\n                            \"pGNi\": 0,\r\n                            \"vPart\": \"0.00\",\r\n                            \"CIDE\": {\r\n                                \"qBCprod\": \"0.0000\",\r\n                                \"vAliqProd\": \"0.0000\",\r\n                                \"vCIDE\": \"0.00\"\r\n                            },\r\n                            \"Encerrante\": {\r\n                                \"nBico\": 0,\r\n                                \"nBomba\": 0,\r\n                                \"nTanque\": 0,\r\n                                \"vEncerranteIni\": 0,\r\n                                \"vEncerranteFim\": 0\r\n                            }\r\n                        },\r\n                        \"Rastro\": []\r\n                    },\r\n                    \"imposto\": {\r\n                        \"ICMS\": {\r\n                            \"orig\": 0,\r\n                            \"CST\": \"40\",\r\n                            \"modBC\": \"0\",\r\n                            \"vBC\": \"0.00\",\r\n                            \"pICMS\": 0,\r\n                            \"vICMS_icms\": \"0.00\",\r\n                            \"modBCST\": 0,\r\n                            \"pMVAST\": 0,\r\n                            \"pRedBCST\": 0,\r\n                            \"vBCST\": \"0.00\",\r\n                            \"vBCSTRet\": \"0.00\",\r\n                            \"pICMSST\": 0,\r\n                            \"vICMSST_icms\": \"0.00\",\r\n                            \"vICMSSTRet\": \"0.00\",\r\n                            \"pRedBC\": 0,\r\n                            \"motDesICMS\": 0,\r\n                            \"vICMSDeson\": \"0.00\",\r\n                            \"vICMSOp\": \"0.00\",\r\n                            \"pDif\": 0,\r\n                            \"vICMSDif\": \"0.00\",\r\n                            \"pBCOp\": 0,\r\n                            \"UFST\": \"\",\r\n                            \"vBCSTDest\": \"0.00\",\r\n                            \"vICMSSTDest_icms\": \"0.00\",\r\n                            \"pCredSN\": 0,\r\n                            \"vCredICMSSN\": \"0.00\",\r\n                            \"GerarTagStRet\": \"\",\r\n                            \"pFCP\": 0,\r\n                            \"vFCP\": \"0.00\",\r\n                            \"vBCFCP\": \"0.00\",\r\n                            \"vBCFCPST\": \"0.00\",\r\n                            \"pFCPST\": 0,\r\n                            \"vFCPST\": \"0.00\",\r\n                            \"pST\": 0,\r\n                            \"vICMSSubstituto\": \"0.00\",\r\n                            \"vBCFCPSTRet\": \"0.00\",\r\n                            \"pFCPSTRet\": 0,\r\n                            \"vFCPSTRet\": \"0.00\",\r\n                            \"GerarICMSST\": \"\",\r\n                            \"pRedBCEfet\": 0,\r\n                            \"vBCEfet\": \"0.00\",\r\n                            \"pICMSEfet\": 0,\r\n                            \"vICMSEfet\": \"0.00\"\r\n                        },\r\n                        \"IPI\": {\r\n                            \"clEnq\": \"\",\r\n                            \"CNPJProd\": \"\",\r\n                            \"cSelo\": \"\",\r\n                            \"qSelo\": 0,\r\n                            \"cEnq\": \"\",\r\n                            \"CSTIPI\": {\r\n                                \"CST_IPI\": \"\",\r\n                                \"vBC_IPI\": \"0.00\",\r\n                                \"qUnid_IPI\": \"0.0000\",\r\n                                \"vUnid_IPI\": \"0.0000\",\r\n                                \"pIPI\": 0,\r\n                                \"vIPI\": \"0.00\"\r\n                            }\r\n                        },\r\n                        \"II\": {\r\n                            \"vBC_imp\": \"0.00\",\r\n                            \"vDespAdu\": \"0.00\",\r\n                            \"vII\": \"0.00\",\r\n                            \"vIOF\": \"0.00\"\r\n                        },\r\n                        \"PIS\": {\r\n                            \"CST_pis\": \"07\",\r\n                            \"vBC_pis\": \"0.00\",\r\n                            \"pPIS\": 0,\r\n                            \"vPIS\": \"0.00\",\r\n                            \"qBCprod_pis\": \"0.0000\",\r\n                            \"vAliqProd_pis\": \"0.0000\"\r\n                        },\r\n                        \"PISST\": {\r\n                            \"vBC_pis_ST\": \"0.00\",\r\n                            \"pPIS_ST\": 0,\r\n                            \"qBCprod_pis_ST\": \"0.0000\",\r\n                            \"vAliqProd_pis_ST\": \"0.0000\",\r\n                            \"vPIS_ST\": \"0.00\"\r\n                        },\r\n                        \"COFINS\": {\r\n                            \"CST_cofins\": \"07\",\r\n                            \"vBC_cofins\": \"0.00\",\r\n                            \"pCOFINS\": 0,\r\n                            \"vCOFINS\": \"0.00\",\r\n                            \"qBCProd_cofins\": \"0.0000\",\r\n                            \"vAliqProd_cofins\": \"0.0000\"\r\n                        },\r\n                        \"COFINSST\": {\r\n                            \"vBC_cofins_ST\": \"0.00\",\r\n                            \"pCOFINS_cofins_ST\": 0,\r\n                            \"qBCProd_cofins_ST\": \"0.0000\",\r\n                            \"vAliqProd_cofins_ST\": \"0.0000\",\r\n                            \"vCOFINS_cofins_ST\": \"0.00\"\r\n                        },\r\n                        \"ISSQN\": {\r\n                            \"vBC_issqn\": \"0.00\",\r\n                            \"vAliq\": 0,\r\n                            \"vISSQN\": \"0.00\",\r\n                            \"cMunFg_issqn\": 0,\r\n                            \"cListServ\": \"\",\r\n                            \"cSitTrib\": \"\",\r\n                            \"vDeducao\": \"0.00\",\r\n                            \"vOutro_issqn\": \"0.00\",\r\n                            \"vDescIncond\": \"0.00\",\r\n                            \"vDescCond\": \"0.00\",\r\n                            \"indISSRet\": 0,\r\n                            \"vISSRet\": \"0.00\",\r\n                            \"indISS\": 0,\r\n                            \"cServico\": \"\",\r\n                            \"cMun_issqn\": 0,\r\n                            \"cPais_issqn\": 0,\r\n                            \"nProcesso\": \"\",\r\n                            \"indIncentivo\": 0\r\n                        },\r\n                        \"ICMSUFDest\": {\r\n                            \"vBCUFDest\": \"0.00\",\r\n                            \"pFCPUFDest\": 0,\r\n                            \"pICMSUFDest\": 0,\r\n                            \"pICMSInter\": 0,\r\n                            \"pICMSInterPart\": 0,\r\n                            \"vFCPUFDest\": \"0.00\",\r\n                            \"vICMSUFDest\": \"0.00\",\r\n                            \"vICMSUFRemet\": \"0.00\",\r\n                            \"vBCFCPUFDest\": \"0.00\"\r\n                        }\r\n                    },\r\n                    \"impostoDevol\": {\r\n                        \"pDevol\": 0,\r\n                        \"IPIDevol\": {\r\n                            \"vIPIDevol\": \"0.00\"\r\n                        }\r\n                    }\r\n                }\r\n            ],\r\n            \"total\": {\r\n                \"ICMStot\": {\r\n                    \"vBC_ttlnfe\": \"0.00\",\r\n                    \"vICMS_ttlnfe\": \"0.00\",\r\n                    \"vICMSDeson_ttlnfe\": \"0.00\",\r\n                    \"vBCST_ttlnfe\": \"0.00\",\r\n                    \"vST_ttlnfe\": \"0.00\",\r\n                    \"vProd_ttlnfe\": \"9975.21\",\r\n                    \"vFrete_ttlnfe\": \"0.00\",\r\n                    \"vSeg_ttlnfe\": \"0.00\",\r\n                    \"vDesc_ttlnfe\": \"0.00\",\r\n                    \"vII_ttlnfe\": \"0.00\",\r\n                    \"vIPI_ttlnfe\": \"0.00\",\r\n                    \"vPIS_ttlnfe\": \"0.00\",\r\n                    \"vCOFINS_ttlnfe\": \"0.00\",\r\n                    \"vOutro\": \"0.00\",\r\n                    \"vNF\": \"9975.21\",\r\n                    \"vTotTrib_ttlnfe\": \"\",\r\n                    \"vFCPUFDest_ttlnfe\": \"0.00\",\r\n                    \"vICMSUFDest_ttlnfe\": \"0.00\",\r\n                    \"vICMSUFRemet_ttlnfe\": \"0.00\",\r\n                    \"vFCP_ttlnfe\": \"0.00\",\r\n                    \"vFCPST_ttlnfe\": \"0.00\",\r\n                    \"vFCPSTRet_ttlnfe\": \"0.00\",\r\n                    \"vIPIDevol_ttlnfe\": \"0.00\",\r\n                    \"vAFRMM_ttlnfe\": \"0.00\"\r\n                },\r\n                \"ISSQNtot\": {\r\n                    \"vServ\": \"0.00\",\r\n                    \"vBC_ttlnfe_iss\": \"0.00\",\r\n                    \"vISS\": \"0.00\",\r\n                    \"vPIS_servttlnfe\": \"0.00\",\r\n                    \"vCOFINS_servttlnfe\": \"0.00\",\r\n                    \"dCompet\": \"0000-00-00\",\r\n                    \"vDeducao_servttlnfe\": \"0.00\",\r\n                    \"vOutro_servttlnfe\": \"0.00\",\r\n                    \"vDescIncond_servttlnfe\": \"0.00\",\r\n                    \"vDescCond_servttlnfe\": \"0.00\",\r\n                    \"vISSRet_servttlnfe\": \"0.00\",\r\n                    \"cRegTrib\": 0\r\n                },\r\n                \"retTrib\": {\r\n                    \"vRetPIS\": \"0.00\",\r\n                    \"vRetCOFINS_servttlnfe\": \"0.00\",\r\n                    \"vRetCSLL\": \"0.00\",\r\n                    \"vBCIRRF\": \"0.00\",\r\n                    \"vIRRF\": \"0.00\",\r\n                    \"vBCRetPrev\": \"0.00\",\r\n                    \"vRetPrev\": \"0.00\"\r\n                }\r\n            },\r\n            \"transp\": {\r\n                \"modFrete\": 9,\r\n                \"balsa\": \"\",\r\n                \"vagao\": \"\",\r\n                \"transporta\": {\r\n                    \"CNPJ_transp\": \"\",\r\n                    \"CPF_transp\": \"\",\r\n                    \"xNome_transp\": \"\",\r\n                    \"IE_transp\": \"\",\r\n                    \"xEnder\": \"\",\r\n                    \"xMun_transp\": \"\",\r\n                    \"UF_transp\": \"\"\r\n                },\r\n                \"retTransp\": {\r\n                    \"vServ_transp\": \"0.00\",\r\n                    \"vBCRet\": \"0.00\",\r\n                    \"pICMSRet\": 0,\r\n                    \"vICMSRet\": \"0.00\",\r\n                    \"CFOP_transp\": 0,\r\n                    \"cMunFG_transp\": 0\r\n                },\r\n                \"veicTransp\": {\r\n                    \"placa\": \"IDN1565\",\r\n                    \"UF_veictransp\": \"RS\",\r\n                    \"RNTC\": \"00811004\"\r\n                },\r\n                \"reboque\": [\r\n                    {\r\n                        \"placa_rebtransp\": \"CHI488\",\r\n                        \"UF_rebtransp\": \"RS\",\r\n                        \"RNTC_rebtransp\": \"00811004\"\r\n                    },\r\n                    {\r\n                        \"placa_rebtransp\": \"CH4985\",\r\n                        \"UF_rebtransp\": \"RS\",\r\n                        \"RNTC_rebtransp\": \"00811004\"\r\n                    }\r\n                ],\r\n                \"vol\": []\r\n            },\r\n            \"cobr\": {\r\n                \"fat\": {\r\n                    \"nFat\": \"\",\r\n                    \"vOrig\": \"0.00\",\r\n                    \"vDesc_cob\": \"0.00\",\r\n                    \"vLiq\": \"0.00\"\r\n                },\r\n                \"dup\": []\r\n            },\r\n            \"pag\": [\r\n                {\r\n                    \"indPag_pag\": \"\",\r\n                    \"tPag\": \"01\",\r\n                    \"vPag\": \"9975.21\",\r\n                    \"card\": {\r\n                        \"tipoIntegracao\": 0,\r\n                        \"CNPJ_card\": \"\",\r\n                        \"tBand\": \"\",\r\n                        \"cAut\": \"\"\r\n                    }\r\n                }\r\n            ],\r\n            \"infAdic\": {\r\n                \"infAdFisco\": \"INFORMACOES DE INTERESSE DO FISCO\",\r\n                \"infCpl\": \"INFORMACOES COMPLEMENTARES\",\r\n                \"obsCont\": [],\r\n                \"procRef\": []\r\n            },\r\n            \"exporta\": {\r\n                \"UFEmbarq\": \"\",\r\n                \"xLocEmbarq\": \"\",\r\n                \"xLocDespacho\": \"\"\r\n            },\r\n            \"compra\": {\r\n                \"xNEmp\": \"\",\r\n                \"xPed\": \"\",\r\n                \"xCont\": \"\"\r\n            },\r\n            \"cana\": {\r\n                \"safra\": \"\",\r\n                \"ref\": \"\",\r\n                \"qTotMes\": \"0.0000000000\",\r\n                \"qTotAnt\": \"0.0000000000\",\r\n                \"qTotGer\": \"0.0000000000\",\r\n                \"vFor\": \"0.00\",\r\n                \"vTotDed\": \"0.00\",\r\n                \"vLiqFor\": \"0.00\",\r\n                \"canaDiario\": [],\r\n                \"canaDeducoes\": []\r\n            }\r\n        }\r\n    }\r\n]";
-        public string NFeSefaz = "{\n    \"Documento\":\"<NFe xmlns=\\\"http://www.portalfiscal.inf.br/nfe\\\"><infNFe versao=\\\"4.00\\\" Id=\\\"NFe43210206354976000149552450000000221000005000\\\"><ide><cUF>43</cUF><cNF>00000500</cNF><natOp>NFE CST 00</natOp><mod>55</mod><serie>245</serie><nNF>22</nNF><dhEmi>2021-02-18T13:50:27-03:00</dhEmi><tpNF>1</tpNF><idDest>2</idDest><cMunFG>4321808</cMunFG><tpImp>2</tpImp><tpEmis>1</tpEmis><cDV>2</cDV><tpAmb>2</tpAmb><finNFe>1</finNFe><indFinal>0</indFinal><indPres>9</indPres><indIntermed>0</indIntermed><procEmi>0</procEmi><verProc>MigrateInvoiCy</verProc></ide><emit><CNPJ>06354976000149</CNPJ><xNome>MIGRATE</xNome><xFant>MIGRATE</xFant><enderEmit><xLgr>RUA PE CACIQUE</xLgr><nro>985</nro><xCpl>2 ANDAR</xCpl><xBairro>CENTRO</xBairro><cMun>4321808</cMun><xMun>TRÊS DE MAIO</xMun><UF>RS</UF><CEP>98910000</CEP><fone>5535354830</fone></enderEmit><IE>1470049241</IE><IM>111</IM><CRT>3</CRT></emit><dest><CNPJ>48662175000352</CNPJ><xNome>NF-E EMITIDA EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL</xNome><enderDest><xLgr>AV JOSE MAZZI</xLgr><nro>1328</nro><xBairro>CENTRO</xBairro><cMun>3518602</cMun><xMun>GUARIBA</xMun><UF>SP</UF><CEP>14840000</CEP><cPais>1058</cPais><xPais>Brasil</xPais></enderDest><indIEDest>1</indIEDest><IE>334005675110</IE></dest><det nItem=\\\"1\\\"><prod><cProd>06011B60D4000</cProd><cEAN>7891009866031</cEAN><xProd>Furad Impac GSB 550 RE 550W 127V14 acc</xProd><NCM>84672100</NCM><CEST>2110000</CEST><cBenef>RS051011</cBenef><CFOP>6401</CFOP><uCom>PC</uCom><qCom>10.0000</qCom><vUnCom>188.4500000000</vUnCom><vProd>1884.50</vProd><cEANTrib>7891009866031</cEANTrib><uTrib>PC</uTrib><qTrib>10.0000</qTrib><vUnTrib>188.4500000000</vUnTrib><indTot>1</indTot><xPed>194101</xPed><nItemPed>000009</nItemPed><nFCI>47366CA3-981D-476E-B479-5C500787E175</nFCI></prod><imposto><ICMS><ICMS10><orig>5</orig><CST>10</CST><modBC>3</modBC><vBC>1884.50</vBC><pICMS>12.0000</pICMS><vICMS>226.14</vICMS><modBCST>4</modBCST><pMVAST>0.0000</pMVAST><vBCST>3101.53</vBCST><pICMSST>18.0000</pICMSST><vICMSST>332.14</vICMSST></ICMS10></ICMS><IPI><cEnq>999</cEnq><IPITrib><CST>50</CST><vBC>1884.50</vBC><pIPI>8.0000</pIPI><vIPI>150.76</vIPI></IPITrib></IPI><PIS><PISAliq><CST>01</CST><vBC>1884.50</vBC><pPIS>1.6500</pPIS><vPIS>31.09</vPIS></PISAliq></PIS><COFINS><COFINSAliq><CST>01</CST><vBC>1884.50</vBC><pCOFINS>7.6000</pCOFINS><vCOFINS>143.22</vCOFINS></COFINSAliq></COFINS></imposto><infAdProd>PC 194101</infAdProd></det><total><ICMSTot><vBC>1884.50</vBC><vICMS>226.14</vICMS><vICMSDeson>0.00</vICMSDeson><vFCP>0.00</vFCP><vBCST>3101.53</vBCST><vST>332.14</vST><vFCPST>0.00</vFCPST><vFCPSTRet>0.00</vFCPSTRet><vProd>1884.50</vProd><vFrete>0.00</vFrete><vSeg>0.00</vSeg><vDesc>0.00</vDesc><vII>0.00</vII><vIPI>150.76</vIPI><vIPIDevol>0.00</vIPIDevol><vPIS>31.09</vPIS><vCOFINS>143.22</vCOFINS><vOutro>0.00</vOutro><vNF>2367.40</vNF></ICMSTot></total><transp><modFrete>0</modFrete><transporta><CNPJ>48740351014204</CNPJ><xNome>BRASPRESS TRANSPORTES URGENTES LTDA</xNome><IE>9073946899</IE><xEnder>ROD REGIS BITTENCOURT - BR 116</xEnder><xMun>CAMPINA GRANDE DO SUL</xMun><UF>PR</UF></transporta><vol><qVol>1</qVol><esp>VOLUMES</esp><pesoL>18.225</pesoL><pesoB>19.100</pesoB></vol></transp><cobr><fat><nFat>0000477031</nFat><vOrig>2367.40</vOrig><vDesc>0.00</vDesc><vLiq>2367.40</vLiq></fat><dup><nDup>001</nDup><dVenc>2022-01-04</dVenc><vDup>2367.40</vDup></dup></cobr><pag><detPag><indPag>1</indPag><tPag>15</tPag><vPag>2367.40</vPag></detPag></pag><infAdic><infAdFisco>ICMS SUBST. TRIBUTARIA</infAdFisco><infCpl>R: 1715906838 O: 0203158316 Numero do Cliente: 0003589702 DVE: BR60 VENDA EFETUADA ATRAVES DE NDISTRIBUIDOR VDBN REPRESENTACOES COMERCIAIS FONE DISTRIB.: 41 2109 8284 PORTARIA CAT No 010 DE 31 DE JANEIRO DE 2020 Usuario: UC4CPIC DUVIDAS: VDBN REPRESENTACOES COMERCIAIS LTDA FONE: 41 2109 8284 EMAIL: posvendasbosch@negrao.com.br .</infCpl></infAdic><infRespTec><CNPJ>45990181003285</CNPJ><xContato>RESPONSAVEL TECNICO BOSCH</xContato><email>REMESSAS.DIVERSAS@BR.BOSCH.COM</email><fone>4738022693</fone><idCSRT>01</idCSRT><hashCSRT>2gfQWXWmGUeAKRzU2Gk/7IhOb+c=</hashCSRT></infRespTec></infNFe></NFe>\" \n}\n";
-        public string NFeDescarte = "[\r\n    {\r\n        \"ModeloDocumento\": \"NFe\",\r\n        \"tpAmb\": 2,\r\n        \"CNPJ_emit\": \"06354976000149\",\r\n        \"Numero\": 25,\r\n        \"Serie\": \"245\",\r\n        \"NumeroSubstituta\": 26,\r\n        \"SerieSubstituta\": \"245\"\r\n    }\r\n]";
-        public string NFeCancelamento = "[\r\n    {\r\n        \"ModeloDocumento\": \"NFe\",\r\n        \"Versao\": 4.00,\r\n        \"Evento\": {\r\n            \"NtfCnpjEmissor\": \"06354976000149\",\r\n            \"NtfNumero\": 26,\r\n            \"NtfSerie\": \"245\",\r\n            \"tpAmb\": 2,\r\n            \"EveInf\": {\r\n                \"EveDh\": \"2021-02-23T14:30:00\",\r\n                \"EveFusoHorario\": \"-03:00\",\r\n                \"EveTp\": 110111,\r\n                \"EvenSeq\": \"1\",\r\n                \"Evedet\": {\r\n                    \"EveDesc\": \"Cancelamento\",\r\n                    \"EvexJust\": \"estou cancelando este documento para testes\"\r\n                }\r\n            }\r\n        }\r\n    }\r\n]";
-        public string NFeCartaCorrecao = "[\r\n    {\r\n        \"ModeloDocumento\": \"NFe\",\r\n        \"Versao\": 4.00,\r\n        \"Evento\": {\r\n            \"NtfCnpjEmissor\": \"06354976000149\",\r\n            \"NtfNumero\": 26,\r\n            \"NtfSerie\": \"245\",\r\n            \"tpAmb\": 2,\r\n            \"EveInf\": {\r\n                \"EveDh\": \"2021-02-23T09:45:00\",\r\n                \"EveFusoHorario\": \"-03:00\",\r\n                \"EveTp\": 110110,\r\n                \"EvenSeq\": \"1\",\r\n                \"Evedet\": {\r\n                    \"EveDesc\": \"Carta de Correcao\",\r\n                    \"EveCorrecao\": \"No campo descrição do produto, Onde se lê: ANO MODELO 2010 Leia-se: ANO MODELO 2015.\",\r\n                    \"EveCondUso\": \"A Carta de Correcao e disciplinada pelo paragrafo 1o-A do art. 7o do Convenio S/N, de 15 de dezembro de 1970 e pode ser utilizada para regularizacao de erro ocorrido na emissao de documento fiscal, desde que o erro nao esteja relacionado com: I - as variaveis que determinam o valor do imposto tais como: base de calculo, aliquota, diferenca de preco, quantidade, valor da operacao ou da prestacao; II - a correcao de dados cadastrais que implique mudanca do remetente ou do destinatario; III - a data de emissao ou de saida.\"\r\n                }\r\n            }\r\n        }\r\n    }\r\n]";
-        public string NFeInutilizacao = "[\r\n    {\r\n        \"ModeloDocumento\": \"NFe\",\r\n        \"Versao\": 4.00,\r\n        \"CnpjEmissor\": \"06354976000149\",\r\n        \"tpAmb\": 2,\r\n        \"UfEmissor\": 43,\r\n        \"NumeroInicial\": 27,\r\n        \"NumeroFinal\": 27,\r\n        \"Serie\": \"245\",\r\n        \"Justificativa\": \"Pulos de numeração.\"\r\n    }\r\n]";
-        public string NFeExport = "";
-        public string NFeExportV2 = "";
-        public string NFeSubstituicao = "";
-        public string NFeEvento = "";
-        public string NFeImportacao = "";
+        private string NFeEnvio = "[\r\n    {\r\n        \"Documento\": {\r\n            \"ModeloDocumento\": \"NFe\",\r\n            \"Versao\": 4,\r\n            \"Parametros\": {\r\n                \"ApelidoLogomarca\": \"\"\r\n            },\r\n            \"ide\": {\r\n                \"cNF\": 1512,\r\n                \"cUF\": 43,\r\n                \"natOp\": \"NFE CST 00\",\r\n                \"mod\": \"55\",\r\n                \"serie\": \"245\",\r\n                \"nNF\": 26,\r\n                \"dhEmi\": \"2021-02-22T15:14:21\",\r\n                \"fusoHorario\": \"-03:00\",\r\n                \"dhSaiEnt\": \"0000-00-00T00:00:00\",\r\n                \"tpNf\": 1,\r\n                \"idDest\": 1,\r\n                \"indFinal\": 0,\r\n                \"indPres\": 0,\r\n                \"cMunFg\": 4321808,\r\n                \"tpImp\": 1,\r\n                \"tpEmis\": 1,\r\n                \"tpAmb\": 2,\r\n                \"xJust\": \"\",\r\n                \"dhCont\": \"0000-00-00T00:00:00\",\r\n                \"finNFe\": 1,\r\n                \"EmailArquivos\": \"email@gmail.com\",\r\n                \"NumeroPedido\": \"\"\r\n            },\r\n            \"emit\": {\r\n                \"CNPJ_emit\": \"06354976000149\",\r\n                \"CPF_emit\": \"\",\r\n                \"xNome\": \"Migrate\",\r\n                \"xFant\": \"EMPRESA DE TESTE\",\r\n                \"IM\": \"\",\r\n                \"CNAE\": \"\",\r\n                \"IE\": \"1470049241\",\r\n                \"IEST\": \"\",\r\n                \"CRT\": 3,\r\n                \"enderEmit\": {\r\n                    \"xLgr\": \"LOGRADOURO\",\r\n                    \"nro\": \"1700\",\r\n                    \"xCpl\": \"ANDAR 02\",\r\n                    \"xBairro\": \"BAIRRO\",\r\n                    \"cMun\": 4321808,\r\n                    \"xMun\": \"ENCANTADO\",\r\n                    \"UF\": \"RS\",\r\n                    \"CEP\": 95960000,\r\n                    \"cPais\": 1058,\r\n                    \"xPais\": \"BRASIL\",\r\n                    \"fone\": 5137513322,\r\n                    \"fax\": 3335354830,\r\n                    \"Email\": \"\"\r\n                }\r\n            },\r\n            \"dest\": {\r\n                \"CNPJ_dest\": \"11635726000117\",\r\n                \"CPF_dest\": \"\",\r\n                \"idEstrangeiro\": \"\",\r\n                \"xNome_dest\": \"NF-E EMITIDA EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"IE_dest\": \"1060127838\",\r\n                \"ISUF\": \"\",\r\n                \"indIEDest\": 1,\r\n                \"IM_dest\": \"\",\r\n                \"enderDest\": {\r\n                    \"nro_dest\": \"67\",\r\n                    \"xCpl_dest\": \"CENTRO\",\r\n                    \"xBairro_dest\": \"CENTRO\",\r\n                    \"xEmail_dest\": \"email@gmail.com\",\r\n                    \"xLgr_dest\": \"LOGRADOURO\",\r\n                    \"xPais_dest\": \"Brasil\",\r\n                    \"cMun_dest\": 4309605,\r\n                    \"xMun_dest\": \"CONTAGEM\",\r\n                    \"UF_dest\": \"RS\",\r\n                    \"CEP_dest\": 32145900,\r\n                    \"cPais_dest\": 1058,\r\n                    \"fone_dest\": 5535485105\r\n                }\r\n            },\r\n            \"retirada\": {\r\n                \"CNPJ_ret\": \"\",\r\n                \"CPF_ret\": \"\",\r\n                \"xNome_ret\": \"\",\r\n                \"xLgr_ret\": \"\",\r\n                \"nro_ret\": \"\",\r\n                \"xCpl_ret\": \"\",\r\n                \"xBairro_ret\": \"\",\r\n                \"xMun_ret\": \"\",\r\n                \"cMun_ret\": 0,\r\n                \"UF_ret\": \"\",\r\n                \"CEP_ret\": 0,\r\n                \"cPais_ret\": 0,\r\n                \"xPais_ret\": \"\",\r\n                \"fone_ret\": 0,\r\n                \"email_ret\": \"\",\r\n                \"IE_ret\": \"\"\r\n            },\r\n            \"entrega\": {\r\n                \"CNPJ_entr\": \"\",\r\n                \"CPF_entr\": \"\",\r\n                \"xLgr_entr\": \"\",\r\n                \"nro_entr\": \"\",\r\n                \"xCpl_entr\": \"\",\r\n                \"xBairro_entr\": \"\",\r\n                \"cMun_entr\": 0,\r\n                \"xMun_entr\": \"\",\r\n                \"UF_entr\": \"\",\r\n                \"xNome_entr\": \"\",\r\n                \"CEP_entr\": 0,\r\n                \"cPais_entr\": 0,\r\n                \"xPais_entr\": \"\",\r\n                \"fone_entr\": 0,\r\n                \"email_entr\": \"\",\r\n                \"IE_entr\": \"\"\r\n            },\r\n            \"autXML\": [],\r\n            \"det\": [\r\n                {\r\n                    \"infADProd\": \"\",\r\n                    \"prod\": {\r\n                        \"cProd\": \"45093812504843\",\r\n                        \"cEAN\": \"SEM GTIN\",\r\n                        \"xProd\": \"MOTOCICLETA MARCA HONDA, TIPO CB 300 R ANO DE FABRICACAO 2009 ANO MODELO 2010\",\r\n                        \"NCM\": \"87112020\",\r\n                        \"EXTIPI\": \"87\",\r\n                        \"CFOP\": 5102,\r\n                        \"uCOM\": \"BOT\",\r\n                        \"qCOM\": \"1.0000\",\r\n                        \"vUnCom\": \"9975.2100000000\",\r\n                        \"vProd\": \"9975.21\",\r\n                        \"cEANTrib\": \"SEM GTIN\",\r\n                        \"uTrib\": \"BOT\",\r\n                        \"qTrib\": \"1.0000\",\r\n                        \"vUnTrib\": \"9975.2100000000\",\r\n                        \"vFrete\": \"0.00\",\r\n                        \"vSeg\": \"0.00\",\r\n                        \"vDesc\": \"0.00\",\r\n                        \"vOutro_item\": \"0.00\",\r\n                        \"indTot\": 1,\r\n                        \"nTipoItem\": 1,\r\n                        \"dProd\": 0,\r\n                        \"xPed_item\": \"\",\r\n                        \"nItemPed\": \"\",\r\n                        \"nFCI\": \"\",\r\n                        \"nRECOPI\": \"123456\",\r\n                        \"CEST\": \"\",\r\n                        \"cBenef\": \"RS051011\",\r\n                        \"indEscala\": \"\",\r\n                        \"CNPJFab\": \"\",\r\n                        \"NVEs\": [],\r\n                        \"detDI\": [],\r\n                        \"detExport\": [],\r\n                        \"veicProd\": {\r\n                            \"tpOp\": 1,\r\n                            \"chassi\": \"9C2NC4310AR033234\",\r\n                            \"cCor\": \"PRME\",\r\n                            \"xCor\": \"PRATA METALICO\",\r\n                            \"pot\": \"2600\",\r\n                            \"cilin\": \"0291\",\r\n                            \"PesoL\": \"000000143\",\r\n                            \"PesoB\": \"000000175\",\r\n                            \"nSerie\": \"0AR033234\",\r\n                            \"tpComb\": \"02\",\r\n                            \"nMotor\": \"NC43E1A033234\",\r\n                            \"CMT\": \"00179.000\",\r\n                            \"dist\": \"1402\",\r\n                            \"RENAVAM\": \"\",\r\n                            \"anoMod\": 2010,\r\n                            \"anoFab\": 2009,\r\n                            \"tpPint\": \"M\",\r\n                            \"tpVeic\": \"04\",\r\n                            \"espVeic\": 1,\r\n                            \"VIN\": \"N\",\r\n                            \"condVeic\": 1,\r\n                            \"cMod\": 2710,\r\n                            \"cCorDENATRAN\": \"04\",\r\n                            \"lota\": 2,\r\n                            \"tpRest\": 0\r\n                        },\r\n                        \"med\": [],\r\n                        \"arma\": [],\r\n                        \"comb\": {\r\n                            \"cProdANP\": 0,\r\n                            \"pMixGN\": 0,\r\n                            \"CODIF\": \"\",\r\n                            \"qTemp\": \"0.0000\",\r\n                            \"UFcons\": \"\",\r\n                            \"descANP\": \"\",\r\n                            \"pGLP\": 0,\r\n                            \"pGNn\": 0,\r\n                            \"pGNi\": 0,\r\n                            \"vPart\": \"0.00\",\r\n                            \"CIDE\": {\r\n                                \"qBCprod\": \"0.0000\",\r\n                                \"vAliqProd\": \"0.0000\",\r\n                                \"vCIDE\": \"0.00\"\r\n                            },\r\n                            \"Encerrante\": {\r\n                                \"nBico\": 0,\r\n                                \"nBomba\": 0,\r\n                                \"nTanque\": 0,\r\n                                \"vEncerranteIni\": 0,\r\n                                \"vEncerranteFim\": 0\r\n                            }\r\n                        },\r\n                        \"Rastro\": []\r\n                    },\r\n                    \"imposto\": {\r\n                        \"ICMS\": {\r\n                            \"orig\": 0,\r\n                            \"CST\": \"40\",\r\n                            \"modBC\": \"0\",\r\n                            \"vBC\": \"0.00\",\r\n                            \"pICMS\": 0,\r\n                            \"vICMS_icms\": \"0.00\",\r\n                            \"modBCST\": 0,\r\n                            \"pMVAST\": 0,\r\n                            \"pRedBCST\": 0,\r\n                            \"vBCST\": \"0.00\",\r\n                            \"vBCSTRet\": \"0.00\",\r\n                            \"pICMSST\": 0,\r\n                            \"vICMSST_icms\": \"0.00\",\r\n                            \"vICMSSTRet\": \"0.00\",\r\n                            \"pRedBC\": 0,\r\n                            \"motDesICMS\": 0,\r\n                            \"vICMSDeson\": \"0.00\",\r\n                            \"vICMSOp\": \"0.00\",\r\n                            \"pDif\": 0,\r\n                            \"vICMSDif\": \"0.00\",\r\n                            \"pBCOp\": 0,\r\n                            \"UFST\": \"\",\r\n                            \"vBCSTDest\": \"0.00\",\r\n                            \"vICMSSTDest_icms\": \"0.00\",\r\n                            \"pCredSN\": 0,\r\n                            \"vCredICMSSN\": \"0.00\",\r\n                            \"GerarTagStRet\": \"\",\r\n                            \"pFCP\": 0,\r\n                            \"vFCP\": \"0.00\",\r\n                            \"vBCFCP\": \"0.00\",\r\n                            \"vBCFCPST\": \"0.00\",\r\n                            \"pFCPST\": 0,\r\n                            \"vFCPST\": \"0.00\",\r\n                            \"pST\": 0,\r\n                            \"vICMSSubstituto\": \"0.00\",\r\n                            \"vBCFCPSTRet\": \"0.00\",\r\n                            \"pFCPSTRet\": 0,\r\n                            \"vFCPSTRet\": \"0.00\",\r\n                            \"GerarICMSST\": \"\",\r\n                            \"pRedBCEfet\": 0,\r\n                            \"vBCEfet\": \"0.00\",\r\n                            \"pICMSEfet\": 0,\r\n                            \"vICMSEfet\": \"0.00\"\r\n                        },\r\n                        \"IPI\": {\r\n                            \"clEnq\": \"\",\r\n                            \"CNPJProd\": \"\",\r\n                            \"cSelo\": \"\",\r\n                            \"qSelo\": 0,\r\n                            \"cEnq\": \"\",\r\n                            \"CSTIPI\": {\r\n                                \"CST_IPI\": \"\",\r\n                                \"vBC_IPI\": \"0.00\",\r\n                                \"qUnid_IPI\": \"0.0000\",\r\n                                \"vUnid_IPI\": \"0.0000\",\r\n                                \"pIPI\": 0,\r\n                                \"vIPI\": \"0.00\"\r\n                            }\r\n                        },\r\n                        \"II\": {\r\n                            \"vBC_imp\": \"0.00\",\r\n                            \"vDespAdu\": \"0.00\",\r\n                            \"vII\": \"0.00\",\r\n                            \"vIOF\": \"0.00\"\r\n                        },\r\n                        \"PIS\": {\r\n                            \"CST_pis\": \"07\",\r\n                            \"vBC_pis\": \"0.00\",\r\n                            \"pPIS\": 0,\r\n                            \"vPIS\": \"0.00\",\r\n                            \"qBCprod_pis\": \"0.0000\",\r\n                            \"vAliqProd_pis\": \"0.0000\"\r\n                        },\r\n                        \"PISST\": {\r\n                            \"vBC_pis_ST\": \"0.00\",\r\n                            \"pPIS_ST\": 0,\r\n                            \"qBCprod_pis_ST\": \"0.0000\",\r\n                            \"vAliqProd_pis_ST\": \"0.0000\",\r\n                            \"vPIS_ST\": \"0.00\"\r\n                        },\r\n                        \"COFINS\": {\r\n                            \"CST_cofins\": \"07\",\r\n                            \"vBC_cofins\": \"0.00\",\r\n                            \"pCOFINS\": 0,\r\n                            \"vCOFINS\": \"0.00\",\r\n                            \"qBCProd_cofins\": \"0.0000\",\r\n                            \"vAliqProd_cofins\": \"0.0000\"\r\n                        },\r\n                        \"COFINSST\": {\r\n                            \"vBC_cofins_ST\": \"0.00\",\r\n                            \"pCOFINS_cofins_ST\": 0,\r\n                            \"qBCProd_cofins_ST\": \"0.0000\",\r\n                            \"vAliqProd_cofins_ST\": \"0.0000\",\r\n                            \"vCOFINS_cofins_ST\": \"0.00\"\r\n                        },\r\n                        \"ISSQN\": {\r\n                            \"vBC_issqn\": \"0.00\",\r\n                            \"vAliq\": 0,\r\n                            \"vISSQN\": \"0.00\",\r\n                            \"cMunFg_issqn\": 0,\r\n                            \"cListServ\": \"\",\r\n                            \"cSitTrib\": \"\",\r\n                            \"vDeducao\": \"0.00\",\r\n                            \"vOutro_issqn\": \"0.00\",\r\n                            \"vDescIncond\": \"0.00\",\r\n                            \"vDescCond\": \"0.00\",\r\n                            \"indISSRet\": 0,\r\n                            \"vISSRet\": \"0.00\",\r\n                            \"indISS\": 0,\r\n                            \"cServico\": \"\",\r\n                            \"cMun_issqn\": 0,\r\n                            \"cPais_issqn\": 0,\r\n                            \"nProcesso\": \"\",\r\n                            \"indIncentivo\": 0\r\n                        },\r\n                        \"ICMSUFDest\": {\r\n                            \"vBCUFDest\": \"0.00\",\r\n                            \"pFCPUFDest\": 0,\r\n                            \"pICMSUFDest\": 0,\r\n                            \"pICMSInter\": 0,\r\n                            \"pICMSInterPart\": 0,\r\n                            \"vFCPUFDest\": \"0.00\",\r\n                            \"vICMSUFDest\": \"0.00\",\r\n                            \"vICMSUFRemet\": \"0.00\",\r\n                            \"vBCFCPUFDest\": \"0.00\"\r\n                        }\r\n                    },\r\n                    \"impostoDevol\": {\r\n                        \"pDevol\": 0,\r\n                        \"IPIDevol\": {\r\n                            \"vIPIDevol\": \"0.00\"\r\n                        }\r\n                    }\r\n                }\r\n            ],\r\n            \"total\": {\r\n                \"ICMStot\": {\r\n                    \"vBC_ttlnfe\": \"0.00\",\r\n                    \"vICMS_ttlnfe\": \"0.00\",\r\n                    \"vICMSDeson_ttlnfe\": \"0.00\",\r\n                    \"vBCST_ttlnfe\": \"0.00\",\r\n                    \"vST_ttlnfe\": \"0.00\",\r\n                    \"vProd_ttlnfe\": \"9975.21\",\r\n                    \"vFrete_ttlnfe\": \"0.00\",\r\n                    \"vSeg_ttlnfe\": \"0.00\",\r\n                    \"vDesc_ttlnfe\": \"0.00\",\r\n                    \"vII_ttlnfe\": \"0.00\",\r\n                    \"vIPI_ttlnfe\": \"0.00\",\r\n                    \"vPIS_ttlnfe\": \"0.00\",\r\n                    \"vCOFINS_ttlnfe\": \"0.00\",\r\n                    \"vOutro\": \"0.00\",\r\n                    \"vNF\": \"9975.21\",\r\n                    \"vTotTrib_ttlnfe\": \"\",\r\n                    \"vFCPUFDest_ttlnfe\": \"0.00\",\r\n                    \"vICMSUFDest_ttlnfe\": \"0.00\",\r\n                    \"vICMSUFRemet_ttlnfe\": \"0.00\",\r\n                    \"vFCP_ttlnfe\": \"0.00\",\r\n                    \"vFCPST_ttlnfe\": \"0.00\",\r\n                    \"vFCPSTRet_ttlnfe\": \"0.00\",\r\n                    \"vIPIDevol_ttlnfe\": \"0.00\",\r\n                    \"vAFRMM_ttlnfe\": \"0.00\"\r\n                },\r\n                \"ISSQNtot\": {\r\n                    \"vServ\": \"0.00\",\r\n                    \"vBC_ttlnfe_iss\": \"0.00\",\r\n                    \"vISS\": \"0.00\",\r\n                    \"vPIS_servttlnfe\": \"0.00\",\r\n                    \"vCOFINS_servttlnfe\": \"0.00\",\r\n                    \"dCompet\": \"0000-00-00\",\r\n                    \"vDeducao_servttlnfe\": \"0.00\",\r\n                    \"vOutro_servttlnfe\": \"0.00\",\r\n                    \"vDescIncond_servttlnfe\": \"0.00\",\r\n                    \"vDescCond_servttlnfe\": \"0.00\",\r\n                    \"vISSRet_servttlnfe\": \"0.00\",\r\n                    \"cRegTrib\": 0\r\n                },\r\n                \"retTrib\": {\r\n                    \"vRetPIS\": \"0.00\",\r\n                    \"vRetCOFINS_servttlnfe\": \"0.00\",\r\n                    \"vRetCSLL\": \"0.00\",\r\n                    \"vBCIRRF\": \"0.00\",\r\n                    \"vIRRF\": \"0.00\",\r\n                    \"vBCRetPrev\": \"0.00\",\r\n                    \"vRetPrev\": \"0.00\"\r\n                }\r\n            },\r\n            \"transp\": {\r\n                \"modFrete\": 9,\r\n                \"balsa\": \"\",\r\n                \"vagao\": \"\",\r\n                \"transporta\": {\r\n                    \"CNPJ_transp\": \"\",\r\n                    \"CPF_transp\": \"\",\r\n                    \"xNome_transp\": \"\",\r\n                    \"IE_transp\": \"\",\r\n                    \"xEnder\": \"\",\r\n                    \"xMun_transp\": \"\",\r\n                    \"UF_transp\": \"\"\r\n                },\r\n                \"retTransp\": {\r\n                    \"vServ_transp\": \"0.00\",\r\n                    \"vBCRet\": \"0.00\",\r\n                    \"pICMSRet\": 0,\r\n                    \"vICMSRet\": \"0.00\",\r\n                    \"CFOP_transp\": 0,\r\n                    \"cMunFG_transp\": 0\r\n                },\r\n                \"veicTransp\": {\r\n                    \"placa\": \"IDN1565\",\r\n                    \"UF_veictransp\": \"RS\",\r\n                    \"RNTC\": \"00811004\"\r\n                },\r\n                \"reboque\": [\r\n                    {\r\n                        \"placa_rebtransp\": \"CHI488\",\r\n                        \"UF_rebtransp\": \"RS\",\r\n                        \"RNTC_rebtransp\": \"00811004\"\r\n                    },\r\n                    {\r\n                        \"placa_rebtransp\": \"CH4985\",\r\n                        \"UF_rebtransp\": \"RS\",\r\n                        \"RNTC_rebtransp\": \"00811004\"\r\n                    }\r\n                ],\r\n                \"vol\": []\r\n            },\r\n            \"cobr\": {\r\n                \"fat\": {\r\n                    \"nFat\": \"\",\r\n                    \"vOrig\": \"0.00\",\r\n                    \"vDesc_cob\": \"0.00\",\r\n                    \"vLiq\": \"0.00\"\r\n                },\r\n                \"dup\": []\r\n            },\r\n            \"pag\": [\r\n                {\r\n                    \"indPag_pag\": \"\",\r\n                    \"tPag\": \"01\",\r\n                    \"vPag\": \"9975.21\",\r\n                    \"card\": {\r\n                        \"tipoIntegracao\": 0,\r\n                        \"CNPJ_card\": \"\",\r\n                        \"tBand\": \"\",\r\n                        \"cAut\": \"\"\r\n                    }\r\n                }\r\n            ],\r\n            \"infAdic\": {\r\n                \"infAdFisco\": \"INFORMACOES DE INTERESSE DO FISCO\",\r\n                \"infCpl\": \"INFORMACOES COMPLEMENTARES\",\r\n                \"obsCont\": [],\r\n                \"procRef\": []\r\n            },\r\n            \"exporta\": {\r\n                \"UFEmbarq\": \"\",\r\n                \"xLocEmbarq\": \"\",\r\n                \"xLocDespacho\": \"\"\r\n            },\r\n            \"compra\": {\r\n                \"xNEmp\": \"\",\r\n                \"xPed\": \"\",\r\n                \"xCont\": \"\"\r\n            },\r\n            \"cana\": {\r\n                \"safra\": \"\",\r\n                \"ref\": \"\",\r\n                \"qTotMes\": \"0.0000000000\",\r\n                \"qTotAnt\": \"0.0000000000\",\r\n                \"qTotGer\": \"0.0000000000\",\r\n                \"vFor\": \"0.00\",\r\n                \"vTotDed\": \"0.00\",\r\n                \"vLiqFor\": \"0.00\",\r\n                \"canaDiario\": [],\r\n                \"canaDeducoes\": []\r\n            }\r\n        }\r\n    }\r\n]";
+        private string NFeSefaz = "{\n    \"Documento\":\"<NFe xmlns=\\\"http://www.portalfiscal.inf.br/nfe\\\"><infNFe versao=\\\"4.00\\\" Id=\\\"NFe43210206354976000149552450000000221000005000\\\"><ide><cUF>43</cUF><cNF>00000500</cNF><natOp>NFE CST 00</natOp><mod>55</mod><serie>245</serie><nNF>22</nNF><dhEmi>2021-02-18T13:50:27-03:00</dhEmi><tpNF>1</tpNF><idDest>2</idDest><cMunFG>4321808</cMunFG><tpImp>2</tpImp><tpEmis>1</tpEmis><cDV>2</cDV><tpAmb>2</tpAmb><finNFe>1</finNFe><indFinal>0</indFinal><indPres>9</indPres><indIntermed>0</indIntermed><procEmi>0</procEmi><verProc>MigrateInvoiCy</verProc></ide><emit><CNPJ>06354976000149</CNPJ><xNome>MIGRATE</xNome><xFant>MIGRATE</xFant><enderEmit><xLgr>RUA PE CACIQUE</xLgr><nro>985</nro><xCpl>2 ANDAR</xCpl><xBairro>CENTRO</xBairro><cMun>4321808</cMun><xMun>TRÊS DE MAIO</xMun><UF>RS</UF><CEP>98910000</CEP><fone>5535354830</fone></enderEmit><IE>1470049241</IE><IM>111</IM><CRT>3</CRT></emit><dest><CNPJ>48662175000352</CNPJ><xNome>NF-E EMITIDA EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL</xNome><enderDest><xLgr>AV JOSE MAZZI</xLgr><nro>1328</nro><xBairro>CENTRO</xBairro><cMun>3518602</cMun><xMun>GUARIBA</xMun><UF>SP</UF><CEP>14840000</CEP><cPais>1058</cPais><xPais>Brasil</xPais></enderDest><indIEDest>1</indIEDest><IE>334005675110</IE></dest><det nItem=\\\"1\\\"><prod><cProd>06011B60D4000</cProd><cEAN>7891009866031</cEAN><xProd>Furad Impac GSB 550 RE 550W 127V14 acc</xProd><NCM>84672100</NCM><CEST>2110000</CEST><cBenef>RS051011</cBenef><CFOP>6401</CFOP><uCom>PC</uCom><qCom>10.0000</qCom><vUnCom>188.4500000000</vUnCom><vProd>1884.50</vProd><cEANTrib>7891009866031</cEANTrib><uTrib>PC</uTrib><qTrib>10.0000</qTrib><vUnTrib>188.4500000000</vUnTrib><indTot>1</indTot><xPed>194101</xPed><nItemPed>000009</nItemPed><nFCI>47366CA3-981D-476E-B479-5C500787E175</nFCI></prod><imposto><ICMS><ICMS10><orig>5</orig><CST>10</CST><modBC>3</modBC><vBC>1884.50</vBC><pICMS>12.0000</pICMS><vICMS>226.14</vICMS><modBCST>4</modBCST><pMVAST>0.0000</pMVAST><vBCST>3101.53</vBCST><pICMSST>18.0000</pICMSST><vICMSST>332.14</vICMSST></ICMS10></ICMS><IPI><cEnq>999</cEnq><IPITrib><CST>50</CST><vBC>1884.50</vBC><pIPI>8.0000</pIPI><vIPI>150.76</vIPI></IPITrib></IPI><PIS><PISAliq><CST>01</CST><vBC>1884.50</vBC><pPIS>1.6500</pPIS><vPIS>31.09</vPIS></PISAliq></PIS><COFINS><COFINSAliq><CST>01</CST><vBC>1884.50</vBC><pCOFINS>7.6000</pCOFINS><vCOFINS>143.22</vCOFINS></COFINSAliq></COFINS></imposto><infAdProd>PC 194101</infAdProd></det><total><ICMSTot><vBC>1884.50</vBC><vICMS>226.14</vICMS><vICMSDeson>0.00</vICMSDeson><vFCP>0.00</vFCP><vBCST>3101.53</vBCST><vST>332.14</vST><vFCPST>0.00</vFCPST><vFCPSTRet>0.00</vFCPSTRet><vProd>1884.50</vProd><vFrete>0.00</vFrete><vSeg>0.00</vSeg><vDesc>0.00</vDesc><vII>0.00</vII><vIPI>150.76</vIPI><vIPIDevol>0.00</vIPIDevol><vPIS>31.09</vPIS><vCOFINS>143.22</vCOFINS><vOutro>0.00</vOutro><vNF>2367.40</vNF></ICMSTot></total><transp><modFrete>0</modFrete><transporta><CNPJ>48740351014204</CNPJ><xNome>BRASPRESS TRANSPORTES URGENTES LTDA</xNome><IE>9073946899</IE><xEnder>ROD REGIS BITTENCOURT - BR 116</xEnder><xMun>CAMPINA GRANDE DO SUL</xMun><UF>PR</UF></transporta><vol><qVol>1</qVol><esp>VOLUMES</esp><pesoL>18.225</pesoL><pesoB>19.100</pesoB></vol></transp><cobr><fat><nFat>0000477031</nFat><vOrig>2367.40</vOrig><vDesc>0.00</vDesc><vLiq>2367.40</vLiq></fat><dup><nDup>001</nDup><dVenc>2022-01-04</dVenc><vDup>2367.40</vDup></dup></cobr><pag><detPag><indPag>1</indPag><tPag>15</tPag><vPag>2367.40</vPag></detPag></pag><infAdic><infAdFisco>ICMS SUBST. TRIBUTARIA</infAdFisco><infCpl>R: 1715906838 O: 0203158316 Numero do Cliente: 0003589702 DVE: BR60 VENDA EFETUADA ATRAVES DE NDISTRIBUIDOR VDBN REPRESENTACOES COMERCIAIS FONE DISTRIB.: 41 2109 8284 PORTARIA CAT No 010 DE 31 DE JANEIRO DE 2020 Usuario: UC4CPIC DUVIDAS: VDBN REPRESENTACOES COMERCIAIS LTDA FONE: 41 2109 8284 EMAIL: posvendasbosch@negrao.com.br .</infCpl></infAdic><infRespTec><CNPJ>45990181003285</CNPJ><xContato>RESPONSAVEL TECNICO BOSCH</xContato><email>REMESSAS.DIVERSAS@BR.BOSCH.COM</email><fone>4738022693</fone><idCSRT>01</idCSRT><hashCSRT>2gfQWXWmGUeAKRzU2Gk/7IhOb+c=</hashCSRT></infRespTec></infNFe></NFe>\" \n}\n";
+        private string NFeDescarte = "[\r\n    {\r\n        \"ModeloDocumento\": \"NFe\",\r\n        \"tpAmb\": 2,\r\n        \"CNPJ_emit\": \"06354976000149\",\r\n        \"Numero\": 25,\r\n        \"Serie\": \"245\",\r\n        \"NumeroSubstituta\": 26,\r\n        \"SerieSubstituta\": \"245\"\r\n    }\r\n]";
+        private string NFeCancelamento = "[\r\n    {\r\n        \"ModeloDocumento\": \"NFe\",\r\n        \"Versao\": 4.00,\r\n        \"Evento\": {\r\n            \"NtfCnpjEmissor\": \"06354976000149\",\r\n            \"NtfNumero\": 26,\r\n            \"NtfSerie\": \"245\",\r\n            \"tpAmb\": 2,\r\n            \"EveInf\": {\r\n                \"EveDh\": \"2021-02-23T14:30:00\",\r\n                \"EveFusoHorario\": \"-03:00\",\r\n                \"EveTp\": 110111,\r\n                \"EvenSeq\": \"1\",\r\n                \"Evedet\": {\r\n                    \"EveDesc\": \"Cancelamento\",\r\n                    \"EvexJust\": \"estou cancelando este documento para testes\"\r\n                }\r\n            }\r\n        }\r\n    }\r\n]";
+        private string NFeCartaCorrecao = "[\r\n    {\r\n        \"ModeloDocumento\": \"NFe\",\r\n        \"Versao\": 4.00,\r\n        \"Evento\": {\r\n            \"NtfCnpjEmissor\": \"06354976000149\",\r\n            \"NtfNumero\": 26,\r\n            \"NtfSerie\": \"245\",\r\n            \"tpAmb\": 2,\r\n            \"EveInf\": {\r\n                \"EveDh\": \"2021-02-23T09:45:00\",\r\n                \"EveFusoHorario\": \"-03:00\",\r\n                \"EveTp\": 110110,\r\n                \"EvenSeq\": \"1\",\r\n                \"Evedet\": {\r\n                    \"EveDesc\": \"Carta de Correcao\",\r\n                    \"EveCorrecao\": \"No campo descrição do produto, Onde se lê: ANO MODELO 2010 Leia-se: ANO MODELO 2015.\",\r\n                    \"EveCondUso\": \"A Carta de Correcao e disciplinada pelo paragrafo 1o-A do art. 7o do Convenio S/N, de 15 de dezembro de 1970 e pode ser utilizada para regularizacao de erro ocorrido na emissao de documento fiscal, desde que o erro nao esteja relacionado com: I - as variaveis que determinam o valor do imposto tais como: base de calculo, aliquota, diferenca de preco, quantidade, valor da operacao ou da prestacao; II - a correcao de dados cadastrais que implique mudanca do remetente ou do destinatario; III - a data de emissao ou de saida.\"\r\n                }\r\n            }\r\n        }\r\n    }\r\n]";
+        private string NFeInutilizacao = "[\r\n    {\r\n        \"ModeloDocumento\": \"NFe\",\r\n        \"Versao\": 4.00,\r\n        \"CnpjEmissor\": \"06354976000149\",\r\n        \"tpAmb\": 2,\r\n        \"UfEmissor\": 43,\r\n        \"NumeroInicial\": 27,\r\n        \"NumeroFinal\": 27,\r\n        \"Serie\": \"245\",\r\n        \"Justificativa\": \"Pulos de numeração.\"\r\n    }\r\n]";
+        private string NFeExport = "";
+        private string NFeExportV2 = "";
+        private string NFeSubstituicao = "";
+        private string NFeEvento = "";
+        private string NFeImportacao = "";
 
-        public string NFCeEnvio = "[\r\n    {\r\n        \"Documento\": {\r\n            \"ModeloDocumento\": \"NFCe\",\r\n            \"Versao\": 4,\r\n            \"ide\": {\r\n                \"cUF\": 43,\r\n                \"natOp\": \"Venda no supermercado\",\r\n                \"mod\": \"65\",\r\n                \"serie\": \"260\",\r\n                \"nNF\": 2019,\r\n                \"dhEmi\": \"2021-02-23T09:20:55\",\r\n                \"fusoHorario\": \"-03:00\",\r\n                \"dhSaiEnt\": \"0000-00-00T00:00:00\",\r\n                \"tpNf\": 1,\r\n                \"idDest\": 1,\r\n                \"indFinal\": 1,\r\n                \"indPres\": 4,\r\n                \"cMunFg\": 4321808,\r\n                \"tpImp\": 4,\r\n                \"tpEmis\": 1,\r\n                \"tpAmb\": 2,\r\n                \"xJust\": \"\",\r\n                \"dhCont\": \"0000-00-00T00:00:00\",\r\n                \"finNFe\": 1,\r\n                \"EmailArquivos\": \"brunateste@migrate.com.br\",\r\n                \"indIntermed\": 1,\r\n                \"NFRef\": []\r\n            },\r\n            \"infIntermed\": {\r\n                \"CNPJ_int\": \"99325560000183\",\r\n                \"idCadIntTran\": \"Fredi\"\r\n            },\r\n            \"emit\": {\r\n                \"CNPJ_emit\": \"06354976000149\",\r\n                \"CPF_emit\": \"\",\r\n                \"xNome\": \"EMPRESA TESTE NFe e NFCe\",\r\n                \"xFant\": \"\",\r\n                \"IM\": \"\",\r\n                \"CNAE\": \"\",\r\n                \"IE\": \"1470049241\",\r\n                \"IEST\": \"\",\r\n                \"CRT\": 3,\r\n                \"enderEmit\": {\r\n                    \"xLgr\": \"Rua Sen. Salgado filho\",\r\n                    \"nro\": \"668\",\r\n                    \"xCpl\": \"\",\r\n                    \"xBairro\": \"centro\",\r\n                    \"cMun\": 4321808,\r\n                    \"xMun\": \"tres de maio\",\r\n                    \"UF\": \"RS\",\r\n                    \"CEP\": 98910000,\r\n                    \"cPais\": 0,\r\n                    \"xPais\": \"\",\r\n                    \"fone\": 0,\r\n                    \"fax\": 0,\r\n                    \"Email\": \"\"\r\n                }\r\n            },\r\n            \"dest\": {\r\n                \"CNPJ_dest\": \"\",\r\n                \"CPF_dest\": \"64356976045\",\r\n                \"idEstrangeiro\": \"\",\r\n                \"xNome_dest\": \"NF-E EMITIDA EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"IE_dest\": \"\",\r\n                \"ISUF\": \"\",\r\n                \"indIEDest\": 9,\r\n                \"IM_dest\": \"\",\r\n                \"enderDest\": {\r\n                    \"nro_dest\": \"12\",\r\n                    \"xCpl_dest\": \"CENTRO\",\r\n                    \"xBairro_dest\": \"Primavera\",\r\n                    \"xEmail_dest\": \"brunateste@migrate.com.br\",\r\n                    \"xLgr_dest\": \"Rua das Margaridas\",\r\n                    \"xPais_dest\": \"Brasil\",\r\n                    \"cMun_dest\": 4317103,\r\n                    \"xMun_dest\": \"CONTAGEM\",\r\n                    \"UF_dest\": \"RS\",\r\n                    \"CEP_dest\": 32145900,\r\n                    \"cPais_dest\": 1058,\r\n                    \"fone_dest\": 5543481105\r\n                }\r\n            },\r\n            \"retirada\": {\r\n                \"CNPJ_ret\": \"\",\r\n                \"CPF_ret\": \"\",\r\n                \"xNome_ret\": \"\",\r\n                \"xLgr_ret\": \"\",\r\n                \"nro_ret\": \"\",\r\n                \"xCpl_ret\": \"\",\r\n                \"xBairro_ret\": \"\",\r\n                \"xMun_ret\": \"\",\r\n                \"cMun_ret\": 0,\r\n                \"UF_ret\": \"\",\r\n                \"CEP_ret\": 0,\r\n                \"cPais_ret\": 0,\r\n                \"xPais_ret\": \"\",\r\n                \"fone_ret\": 0,\r\n                \"email_ret\": \"\",\r\n                \"IE_ret\": \"\"\r\n            },\r\n            \"entrega\": {\r\n                \"CNPJ_entr\": \"\",\r\n                \"CPF_entr\": \"\",\r\n                \"xLgr_entr\": \"\",\r\n                \"nro_entr\": \"\",\r\n                \"xCpl_entr\": \"\",\r\n                \"xBairro_entr\": \"\",\r\n                \"cMun_entr\": 0,\r\n                \"xMun_entr\": \"\",\r\n                \"UF_entr\": \"\",\r\n                \"xNome_entr\": \"\",\r\n                \"CEP_entr\": 0,\r\n                \"cPais_entr\": 0,\r\n                \"xPais_entr\": \"\",\r\n                \"fone_entr\": 0,\r\n                \"email_entr\": \"\",\r\n                \"IE_entr\": \"\"\r\n            },\r\n            \"autXML\": [\r\n                {\r\n                    \"CNPJ_aut\": \"\",\r\n                    \"CPF_aut\": \"16928779098\"\r\n                }\r\n            ],\r\n            \"det\": [\r\n                {\r\n                    \"infADProd\": \"Informações Adicionais\",\r\n                    \"prod\": {\r\n                        \"cProd\": \"1\",\r\n                        \"cEAN\": \"SEM GTIN\",\r\n                        \"xProd\": \"NOTA FISCAL EMITIDA EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                        \"NCM\": \"39231090\",\r\n                        \"EXTIPI\": \"\",\r\n                        \"CFOP\": 5102,\r\n                        \"uCOM\": \"UN\",\r\n                        \"qCOM\": \"1.0000\",\r\n                        \"vUnCom\": \"10.0000000000\",\r\n                        \"vProd\": \"10.00\",\r\n                        \"cEANTrib\": \"SEM GTIN\",\r\n                        \"uTrib\": \"UN\",\r\n                        \"qTrib\": \"1.0000\",\r\n                        \"vUnTrib\": \"10.0000000000\",\r\n                        \"vFrete\": \"0.00\",\r\n                        \"vSeg\": \"0.00\",\r\n                        \"vDesc\": \"0.00\",\r\n                        \"vOutro_item\": \"0.00\",\r\n                        \"indTot\": 1,\r\n                        \"nTipoItem\": 0,\r\n                        \"dProd\": 0,\r\n                        \"xPed_item\": \"\",\r\n                        \"nItemPed\": \"\",\r\n                        \"nFCI\": \"\",\r\n                        \"nRECOPI\": \"\",\r\n                        \"CEST\": \"1234567\",\r\n                        \"cBenef\": \"RS051011\"\r\n                    },\r\n                    \"imposto\": {\r\n                        \"vTotTrib\": \"\",\r\n                        \"ICMS\": {\r\n                            \"orig\": 0,\r\n                            \"CST\": \"40\",\r\n                            \"modBC\": \"\",\r\n                            \"vBC\": \"0.00\",\r\n                            \"pICMS\": 0,\r\n                            \"vICMS_icms\": \"0.00\",\r\n                            \"modBCST\": 0,\r\n                            \"pMVAST\": 0,\r\n                            \"pRedBCST\": 0,\r\n                            \"vBCST\": \"0.00\",\r\n                            \"vBCSTRet\": \"0.00\",\r\n                            \"pICMSST\": 0,\r\n                            \"vICMSST_icms\": \"0.00\",\r\n                            \"vICMSSTRet\": \"0.00\",\r\n                            \"pRedBC\": 0,\r\n                            \"motDesICMS\": 0,\r\n                            \"vICMSDeson\": \"0.00\",\r\n                            \"vICMSOp\": \"0.00\",\r\n                            \"pDif\": 0,\r\n                            \"vICMSDif\": \"0.00\",\r\n                            \"pBCOp\": 0,\r\n                            \"UFST\": \"\",\r\n                            \"vBCSTDest\": \"0.00\",\r\n                            \"vICMSSTDest_icms\": \"0.00\",\r\n                            \"pCredSN\": 0,\r\n                            \"vCredICMSSN\": \"0.00\",\r\n                            \"pFCP\": 0,\r\n                            \"vFCP\": \"0.00\",\r\n                            \"vBCFCP\": \"0.00\",\r\n                            \"vBCFCPST\": \"0.00\",\r\n                            \"pFCPST\": 0,\r\n                            \"vFCPST\": \"0.00\",\r\n                            \"pST\": 0,\r\n                            \"vICMSSubstituto\": \"0.00\",\r\n                            \"vBCFCPSTRet\": \"0.00\",\r\n                            \"pFCPSTRet\": 0,\r\n                            \"vFCPSTRet\": \"0.00\",\r\n                            \"GerarICMSST\": \"\",\r\n                            \"pRedBCEfet\": 0,\r\n                            \"vBCEfet\": \"0.00\",\r\n                            \"pICMSEfet\": 0,\r\n                            \"vICMSEfet\": \"0.00\"\r\n                        },\r\n                        \"IPI\": {\r\n                            \"CNPJProd\": \"\",\r\n                            \"cSelo\": \"\",\r\n                            \"qSelo\": 0,\r\n                            \"cEnq\": \"\",\r\n                            \"CSTIPI\": {\r\n                                \"CST_IPI\": \"\",\r\n                                \"vBC_IPI\": \"0.00\",\r\n                                \"qUnid_IPI\": \"0.0000\",\r\n                                \"vUnid_IPI\": \"0.0000\",\r\n                                \"pIPI\": 0,\r\n                                \"vIPI\": \"0.00\"\r\n                            }\r\n                        }\r\n                    }\r\n                }\r\n            ],\r\n            \"total\": {\r\n                \"ICMStot\": {\r\n                    \"vBC_ttlnfe\": \"0.00\",\r\n                    \"vICMS_ttlnfe\": \"0.00\",\r\n                    \"vICMSDeson_ttlnfe\": \"0.00\",\r\n                    \"vBCST_ttlnfe\": \"0.00\",\r\n                    \"vST_ttlnfe\": \"0.00\",\r\n                    \"vProd_ttlnfe\": \"10.00\",\r\n                    \"vFrete_ttlnfe\": \"0.00\",\r\n                    \"vSeg_ttlnfe\": \"0.00\",\r\n                    \"vDesc_ttlnfe\": \"0.00\",\r\n                    \"vII_ttlnfe\": \"0.00\",\r\n                    \"vIPI_ttlnfe\": \"0.00\",\r\n                    \"vPIS_ttlnfe\": \"0.00\",\r\n                    \"vCOFINS_ttlnfe\": \"0.00\",\r\n                    \"vOutro\": \"0.00\",\r\n                    \"vNF\": \"10.00\",\r\n                    \"vTotTrib_ttlnfe\": \"0\"\r\n                }\r\n            },\r\n            \"transp\": {\r\n                \"modFrete\": 2,\r\n                \"balsa\": \"\",\r\n                \"vagao\": \"\",\r\n                \"transporta\": {\r\n                    \"CNPJ_transp\": \"\",\r\n                    \"CPF_transp\": \"67492234007\",\r\n                    \"xNome_transp\": \"Transportador\",\r\n                    \"IE_transp\": \"\",\r\n                    \"xEnder\": \"\",\r\n                    \"xMun_transp\": \"\",\r\n                    \"UF_transp\": \"RS\"\r\n                },\r\n                \"retTransp\": {\r\n                    \"vServ_transp\": \"0.00\",\r\n                    \"vBCRet\": \"0.00\",\r\n                    \"pICMSRet\": 0,\r\n                    \"vICMSRet\": \"0.00\",\r\n                    \"CFOP_transp\": 0,\r\n                    \"cMunFG_transp\": 0\r\n                },\r\n                \"veicTransp\": {\r\n                    \"placa\": \"\",\r\n                    \"UF_veictransp\": \"\",\r\n                    \"RNTC\": \"\"\r\n                },\r\n                \"reboque\": [],\r\n                \"vol\": []\r\n            },\r\n            \"cobr\": {\r\n                \"fat\": {\r\n                    \"nFat\": \"\",\r\n                    \"vOrig\": \"0.00\",\r\n                    \"vDesc_cob\": \"0.00\",\r\n                    \"vLiq\": \"0.00\"\r\n                },\r\n                \"dup\": []\r\n            },\r\n            \"pag\": [\r\n                {\r\n                    \"indPag_pag\": \"\",\r\n                    \"tPag\": \"16\",\r\n                    \"vPag\": \"10.00\",\r\n                    \"card\": {\r\n                        \"tipoIntegracao\": 0,\r\n                        \"CNPJ_card\": \"\",\r\n                        \"tBand\": \"\",\r\n                        \"cAut\": \"\"\r\n                    }\r\n                }\r\n            ],\r\n            \"vTroco\": \"0.00\",\r\n            \"infSuplem\": {\r\n                \"qrCode\": \"\",\r\n                \"urlChave\": \"\"\r\n            },\r\n            \"RespTecnico\": {\r\n                \"CNPJ\": \"87610802000100\",\r\n                \"xContato\": \"Joao Carlos da Cunha Pereira\",\r\n                \"email\": \"testemigrate1@gmail.com\",\r\n                \"fone\": 35351285,\r\n                \"idCSRT\": \"12\",\r\n                \"hashCSRT\": \"v61uDiGY1KWigdgMfHGYi6qilQo=\"\r\n            }\r\n        }\r\n    }\r\n]";
-        public string NFCeSefaz = "";
-        public string NFCeDescarte = "[\r\n    {\r\n        \"ModeloDocumento\": \"NFCe\",\r\n        \"tpAmb\": 2,\r\n        \"CNPJ_emit\": \"06354976000149\",\r\n        \"Numero\": 25,\r\n        \"Serie\": \"245\",\r\n        \"NumeroSubstituta\": 26,\r\n        \"SerieSubstituta\": \"245\"\r\n    }\r\n]";
-        public string NFCeCancelamento = "[\n\t{\n\t\t\"ModeloDocumento\": \"NFCe\",\n\t\t\"Versao\": \"4.00\",\n\t\t\"Evento\": {\n\t      \"NtfCnpjEmissor\": \"06354976000149\",\n\t      \"NtfNumero\": \"52465495\",\n\t      \"NtfSerie\": \"251\",\n\t      \"tpAmb\": \"2\",\n\t      \"EveInf\": {\n\t         \"EveDh\": \"2020-05-20T11:00:00-03:00\",\n\t         \"EveFusoHorario\": \"-03:00\",\n\t         \"EveTp\": \"110111\",\n\t         \"EvenSeq\": \"1\",\n\t         \"Evedet\": {\n\t            \"EveDesc\": \"Cancelamento\",\n\t            \"EvenProt\": \"143200000317337\",\n\t            \"EvexJust\": \"Aqui vai a justificativa de cancelamento\"\n\t         }\n\t      }\n\t\t}\n\t}\n]";
-        public string NFCeCartaCorrecao = "";
-        public string NFCeInutilizacao = "";
-        public string NFCeExport = "[\\n\\t{\\n\\t\\t\\\"ModeloDocumento\\\": \\\"nfce\\\",\\n\\t\\t\\\"Versao\\\": \\\"4.0\\\",\\n\\t\\t\\\"tpAmb\\\": 2,\\n\\t\\t\\\"CnpjEmpresa\\\": \\\"06354976000149\\\",\\n\\t\\t\\\"StatutsDocumento\\\": 2,\\n\\t\\t\\\"NumeroInicial\\\": 1,\\n\\t\\t\\\"NumeroFinal\\\": 270,\\n\\t\\t\\\"Serie\\\": \\\"650\\\",\\n\\t\\t\\\"ResumoDocumento\\\": \\\"S\\\",\\n\\t\\t\\\"ParmSituacao\\\": \\\"S\\\",\\n\\t\\t\\\"ParmXMLCompleto\\\": \\\"S\\\",\\n\\t\\t\\\"ParmEventos\\";
-        public string NFCeExportV2 = "[\n\t{\n\t\t\"ModeloDocumento\": \"nfce\",\n\t\t\"Versao\": \"4.0\",\n\t\t\"tpAmb\": 2,\n\t\t\"CnpjEmpresa\": \"06354976000149\",\n\t\t\"StatutsDocumento\": 2,\n\t\t\"NumeroInicial\": 1,\n\t\t\"NumeroFinal\": 270,\n\t\t\"Serie\": \"650\",\n\t\t\"ResumoDocumento\": \"S\",\n\t\t\"ParmSituacao\": \"S\",\n\t\t\"ParmXMLCompleto\": \"S\",\n\t\t\"ParmEventos\": \"S\"\n\t}\n]";
-        public string NFCeSubstituicao = "";
-        public string NFCeEvento = "[\n\t{\n\t\t\"ModeloDocumento\": \"NFCe\",\n\t\t\"Versao\": \"4.00\",\n\t\t\"Evento\": {\n\t      \"NtfCnpjEmissor\": \"06354976000149\",\n\t      \"NtfNumero\": \"52465495\",\n\t      \"NtfSerie\": \"251\",\n\t      \"tpAmb\": \"2\",\n\t      \"EveInf\": {\n\t         \"EveDh\": \"2020-05-20T11:00:00-03:00\",\n\t         \"EveFusoHorario\": \"-03:00\",\n\t         \"EveTp\": \"110111\",\n\t         \"EvenSeq\": \"1\",\n\t         \"Evedet\": {\n\t            \"EveDesc\": \"Cancelamento\",\n\t            \"EvenProt\": \"143200000317337\",\n\t            \"EvexJust\": \"Aqui vai a justificativa de cancelamento\"\n\t         }\n\t      }\n\t\t}\n\t}\n]";
-        public string NFCeImportacao = "[\n\t{\n\t\t\"Documento\" : \"<nfeProc versao=\\\"4.00\\\" xmlns=\\\"http://www.portalfiscal.inf.br/nfe\\\"><NFe xmlns=\\\"http://www.portalfiscal.inf.br/nfe\\\"><infNFe versao=\\\"4.00\\\" Id=\\\"NFe43201206354976000149652510524655221000000025\\\"><ide><cUF>43</cUF><cNF>00000002</cNF><natOp>NFE CST 00</natOp><mod>65</mod><serie>1</serie><nNF>1</nNF><dhEmi>0000-00-00T00:00:00</dhEmi><tpNF>1</tpNF><idDest>1</idDest><cMunFG>4309605</cMunFG><tpImp>4</tpImp><tpEmis>1</tpEmis><cDV>5</cDV><tpAmb>2</tpAmb><finNFe>1</finNFe><indFinal>1</indFinal><indPres>1</indPres><procEmi>0</procEmi><verProc>MigrateInvoiCy</verProc></ide><emit><CNPJ>06354976000149</CNPJ><xNome>MIGRATE COMPANY SISTEMAS DE INFORMACAO</xNome><xFant>EMPRESA DE TESTE</xFant><enderEmit><xLgr>AAAAAAAAAAAA</xLgr><nro>1700</nro><xCpl>ANDAR 02</xCpl><xBairro>BARRA DO JACARE</xBairro><cMun>4321808</cMun><xMun>ENCANTADO</xMun><UF>RS</UF><CEP>95960000</CEP><cPais>1058</cPais><xPais>BRASIL</xPais><fone>5137513322</fone></enderEmit><IE>1470049241</IE><CRT>3</CRT></emit><dest><CNPJ>99999999000191</CNPJ><xNome>NF-E EMITIDA EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL</xNome><enderDest><xLgr>AAAAAAAAAAAAAAAAAA</xLgr><nro>67</nro><xCpl>CENTRO</xCpl><xBairro>KENNEDY</xBairro><cMun>4321808</cMun><xMun>Tres de Maio</xMun><UF>RS</UF><CEP>98910000</CEP><cPais>1058</cPais><xPais>Brasil</xPais><fone>999999999</fone></enderDest><indIEDest>9</indIEDest><email>aaaaa@aaaa.aaa</email></dest><det nItem=\\\"1\\\"><prod><cProd>00009</cProd><cEAN>SEM GTIN</cEAN><xProd>NOTA FISCAL EMITIDA EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL</xProd><NCM>87113000</NCM><EXTIPI>87</EXTIPI><CFOP>5101</CFOP><uCom>BOT</uCom><qCom>1.0000</qCom><vUnCom>100.0000000000</vUnCom><vProd>100.00</vProd><cEANTrib>SEM GTIN</cEANTrib><uTrib>BOT</uTrib><qTrib>1.0000</qTrib><vUnTrib>100.0000000000</vUnTrib><indTot>1</indTot></prod><imposto><vTotTrib>0.00</vTotTrib><ICMS><ICMS00><orig>0</orig><CST>00</CST><modBC>3</modBC><vBC>100.00</vBC><pICMS>17.0000</pICMS><vICMS>17.00</vICMS></ICMS00></ICMS><PIS><PISNT><CST>07</CST></PISNT></PIS><COFINS><COFINSNT><CST>07</CST></COFINSNT></COFINS></imposto><infAdProd>informacoes adicionais do produto</infAdProd></det><total><ICMSTot><vBC>100.00</vBC><vICMS>17.00</vICMS><vICMSDeson>0.00</vICMSDeson><vFCP>0.00</vFCP><vBCST>0.00</vBCST><vST>0.00</vST><vFCPST>0.00</vFCPST><vFCPSTRet>0.00</vFCPSTRet><vProd>100.00</vProd><vFrete>0.00</vFrete><vSeg>0.00</vSeg><vDesc>0.00</vDesc><vII>0.00</vII><vIPI>0.00</vIPI><vIPIDevol>0.00</vIPIDevol><vPIS>0.00</vPIS><vCOFINS>0</vCOFINS><vOutro>0.00</vOutro><vNF>100.00</vNF><vTotTrib>0.00</vTotTrib></ICMSTot></total><transp><modFrete>9</modFrete></transp><pag><detPag><tPag>01</tPag><vPag>100.00</vPag></detPag></pag><infAdic><infAdFisco>INFORMACOES DE INTERESSE DO FISCO</infAdFisco><infCpl>informacoes complementares da nota</infCpl></infAdic></infNFe><infNFeSupl><qrCode><![CDATA[https://www.sefaz.rs.gov.br/NFCE/NFCE-COM.aspx?p=43201206354976000149652510524655221000000025|2|2|1|BAA3EEC60CCD7C999116DF0805BAA98C6DA23BB2]]></qrCode><urlChave><![CDATA[www.sefaz.rs.gov.br/nfce/consulta]]></urlChave></infNFeSupl><Signature xmlns=\\\"http://www.w3.org/2000/09/xmldsig#\\\"><SignedInfo><CanonicalizationMethod Algorithm=\\\"http://www.w3.org/TR/2001/REC-xml-c14n-20010315\\\" /><SignatureMethod Algorithm=\\\"http://www.w3.org/2000/09/xmldsig#rsa-sha1\\\" /><Reference URI=\\\"#NFe43201206354976000149652510524655221000000025\\\"><Transforms><Transform Algorithm=\\\"http://www.w3.org/2000/09/xmldsig#enveloped-signature\\\" /><Transform Algorithm=\\\"http://www.w3.org/TR/2001/REC-xml-c14n-20010315\\\" /></Transforms><DigestMethod Algorithm=\\\"http://www.w3.org/2000/09/xmldsig#sha1\\\" /><DigestValue>pwSVTayc3xZcl0jkjfOvpZcm5CU=</DigestValue></Reference></SignedInfo><SignatureValue>UbuijwQbjaBgyIOUWjKS25q2H2FQXUCsAxcjhuTUsssIOV/Sl8ddQ4+VJP1/eFv5woOHIk7lgVAJofjRwzJQ1GlNJ0D2eatoTH5GKDKdS5w0kx2zR/Gl02waKzHMAmK3oXl3W03ZbXjy6OBHc0+PArD23k+TsHCl+1g4PUmlh9xDN6/ncYwjS1Tfq9fqQV50xiF2ttTuxHRUEQz4xzQ9ik6t+mMHsfWSq5VIiEcpo/Hau8aVosr8X+v8HhOcleVacyBdjzP0B9TfyQVOsL2AEpc7fQsqJKCQyUK8u46ZDlfo+fCiUVYrbKnuEWeq/k0/0mw34WLtTXl12YApV7gSxQ==</SignatureValue><KeyInfo><X509Data><X509Certificate>MIIH5DCCBcygAwIBAgIIPqwgASE+wgQwDQYJKoZIhvcNAQELBQAwgYkxCzAJBgNVBAYTAkJSMRMwEQYDVQQKEwpJQ1AtQnJhc2lsMTQwMgYDVQQLEytBdXRvcmlkYWRlIENlcnRpZmljYWRvcmEgUmFpeiBCcmFzaWxlaXJhIHYyMRIwEAYDVQQLEwlBQyBTT0xVVEkxGzAZBgNVBAMTEkFDIFNPTFVUSSBNdWx0aXBsYTAeFw0yMDAxMjExNDM2MTlaFw0yMTAxMjAxNDM2MTlaMIIBBDELMAkGA1UEBhMCQlIxEzARBgNVBAoTCklDUC1CcmFzaWwxNDAyBgNVBAsTK0F1dG9yaWRhZGUgQ2VydGlmaWNhZG9yYSBSYWl6IEJyYXNpbGVpcmEgdjIxEjAQBgNVBAsTCUFDIFNPTFVUSTEbMBkGA1UECxMSQUMgU09MVVRJIE11bHRpcGxhMRcwFQYDVQQLEw4wMzE1MTIwMDAwMDEzMzEaMBgGA1UECxMRQ2VydGlmaWNhZG8gUEogQTExRDBCBgNVBAMTO01JR1JBVEUgQ09NUEFOWSBTSVNURU1BUyBERSBJTkZPUk1BQ0FPIExUREEuOjA2MzU0OTc2MDAwMTQ5MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA0RO3miPcW5w6F0zjuqu87u9XqKUv4AAXHQSMb9XwsfrBOdeFzz3iibWaPAROKMQZIRiPVInH49QiIURb6cTPFUyc7rcoeBG/74z1saSb+KY4NxaKa0YoFpRnJoqyVkCID/WrKqAumktLeNzcQeALFVyycGq3LNl/mNzBZUNB1seu2pzqQwaE+PGzewVFYuKP5/KuOwXdoIlUI31J0lcrgVRsZbPdh/t+jrSsyBQyWEXOCQIdjnT9UIQQPr64TXtDHK5jXJBjmcIxtW/4VjOMCUYCJNhZ7T+DKaFN1u6m2K+EbZiZ3sfJ8mtej9BpLFhtuh6NGPY/Gfr3CXv/d5RyhwIDAQABo4IC0DCCAswwVAYIKwYBBQUHAQEESDBGMEQGCCsGAQUFBzAChjhodHRwOi8vY2NkLmFjc29sdXRpLmNvbS5ici9sY3IvYWMtc29sdXRpLW11bHRpcGxhLXYxLnA3YjAdBgNVHQ4EFgQUzjFHN/G/Syc2JqfQ1+iSecHOLiMwCQYDVR0TBAIwADAfBgNVHSMEGDAWgBQ1rjEU9l7Sek9Y/jSoGmeXCsSbBzBeBgNVHSAEVzBVMFMGBmBMAQIBJjBJMEcGCCsGAQUFBwIBFjtodHRwczovL2NjZC5hY3NvbHV0aS5jb20uYnIvZG9jcy9kcGMtYWMtc29sdXRpLW11bHRpcGxhLnBkZjCB3gYDVR0fBIHWMIHTMD6gPKA6hjhodHRwOi8vY2NkLmFjc29sdXRpLmNvbS5ici9sY3IvYWMtc29sdXRpLW11bHRpcGxhLXYxLmNybDA/oD2gO4Y5aHR0cDovL2NjZDIuYWNzb2x1dGkuY29tLmJyL2xjci9hYy1zb2x1dGktbXVsdGlwbGEtdjEuY3JsMFCgTqBMhkpodHRwOi8vcmVwb3NpdG9yaW8uaWNwYnJhc2lsLmdvdi5ici9sY3IvQUNTT0xVVEkvYWMtc29sdXRpLW11bHRpcGxhLXYxLmNybDAOBgNVHQ8BAf8EBAMCBeAwHQYDVR0lBBYwFAYIKwYBBQUHAwIGCCsGAQUFBwMEMIG4BgNVHREEgbAwga2BGWZpbmFuY2Vpcm9AbWlncmF0ZS5jb20uYnKgIgYFYEwBAwKgGRMXQURJTFNPTiBNT0FDSVIgV0VERElHRU6gGQYFYEwBAwOgEBMOMDYzNTQ5NzYwMDAxNDmgOAYFYEwBAwSgLxMtMjgwNTE5Nzg3NzM1MzA1NjA1MzAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwoBcGBWBMAQMHoA4TDDAwMDAwMDAwMDAwMDANBgkqhkiG9w0BAQsFAAOCAgEAYnQK+oKGphgfnmMTl1A95K2OHiMMkowaHXE0iWNi8pROOvxUnB8Duq/l6n3Zh58JfQolznx4O54olb8JPiKKJ8paFQ8i1rys4inIyJT2rF3N7MF2IC9jGPbMpcf+cxXYR1IYqir7nzrROB0Wlk7kl+nIEN6HUxraDVpeyhrJzH/cNLFYEobthXiYS4UIbH0pfZU2F64QZhnY1YKM6fD22LH22k9dI8VTZWQFQnzeSsutryJZfttn82DfTk6DDqRaYmKIZs9o9ihS3Hftzyq18MSR7k8aDrPGwTAGSBnr46jjKx7BpQ6gZ5YmAHFbv9dtbLUCKh3m0gdr5KerIftpipQ0hxJQPcbP8FUiiDaTwj4SqI+cups/45+aEddAEquXhUaR7ROQUeRMhXAS3+0aBdF+TcvxZIXewNSWh20A01H7f6QOnwoxgfwf/ulEgbRZFnLCW2wDsEr/PZw0rGOzWVQviCb23WbC33x72iFydHBtli969YaQSLhcvDrnQ2QNZphd/Zwvt0/Fd60aokZRGvZQELTuJq2pf73nxfgCZ8bobPUUNlANP6hLAWJPW1vUn8O87KnvdiWruHc6x6KK3RtRw0xAms/Kcai4h48P2KBHsJAmiaAoK/tznZ3ShFBDbQz7wFS1/ucW5IzjxAgXvxBKPoFjAFgDSO2vHwJBnYo=</X509Certificate></X509Data></KeyInfo></Signature></NFe><protNFe versao=\\\"4.00\\\"><infProt><tpAmb>2</tpAmb><verAplic>RSnfce202010200652</verAplic><chNFe>43201206354976000149652510524655221000000025</chNFe><dhRecbto>2020-12-11T16:23:09-03:00</dhRecbto><nProt>143200001216164</nProt><digVal>pwSVTayc3xZcl0jkjfOvpZcm5CU=</digVal><cStat>100</cStat><xMotivo>Autorizado o uso da NF-e</xMotivo></infProt></protNFe></nfeProc>\"\n\t}\n]";
+        private string NFCeEnvio = "[\r\n    {\r\n        \"Documento\": {\r\n            \"ModeloDocumento\": \"NFCe\",\r\n            \"Versao\": 4,\r\n            \"ide\": {\r\n                \"cUF\": 43,\r\n                \"natOp\": \"Venda no supermercado\",\r\n                \"mod\": \"65\",\r\n                \"serie\": \"260\",\r\n                \"nNF\": 2019,\r\n                \"dhEmi\": \"2021-02-23T09:20:55\",\r\n                \"fusoHorario\": \"-03:00\",\r\n                \"dhSaiEnt\": \"0000-00-00T00:00:00\",\r\n                \"tpNf\": 1,\r\n                \"idDest\": 1,\r\n                \"indFinal\": 1,\r\n                \"indPres\": 4,\r\n                \"cMunFg\": 4321808,\r\n                \"tpImp\": 4,\r\n                \"tpEmis\": 1,\r\n                \"tpAmb\": 2,\r\n                \"xJust\": \"\",\r\n                \"dhCont\": \"0000-00-00T00:00:00\",\r\n                \"finNFe\": 1,\r\n                \"EmailArquivos\": \"brunateste@migrate.com.br\",\r\n                \"indIntermed\": 1,\r\n                \"NFRef\": []\r\n            },\r\n            \"infIntermed\": {\r\n                \"CNPJ_int\": \"99325560000183\",\r\n                \"idCadIntTran\": \"Fredi\"\r\n            },\r\n            \"emit\": {\r\n                \"CNPJ_emit\": \"06354976000149\",\r\n                \"CPF_emit\": \"\",\r\n                \"xNome\": \"EMPRESA TESTE NFe e NFCe\",\r\n                \"xFant\": \"\",\r\n                \"IM\": \"\",\r\n                \"CNAE\": \"\",\r\n                \"IE\": \"1470049241\",\r\n                \"IEST\": \"\",\r\n                \"CRT\": 3,\r\n                \"enderEmit\": {\r\n                    \"xLgr\": \"Rua Sen. Salgado filho\",\r\n                    \"nro\": \"668\",\r\n                    \"xCpl\": \"\",\r\n                    \"xBairro\": \"centro\",\r\n                    \"cMun\": 4321808,\r\n                    \"xMun\": \"tres de maio\",\r\n                    \"UF\": \"RS\",\r\n                    \"CEP\": 98910000,\r\n                    \"cPais\": 0,\r\n                    \"xPais\": \"\",\r\n                    \"fone\": 0,\r\n                    \"fax\": 0,\r\n                    \"Email\": \"\"\r\n                }\r\n            },\r\n            \"dest\": {\r\n                \"CNPJ_dest\": \"\",\r\n                \"CPF_dest\": \"64356976045\",\r\n                \"idEstrangeiro\": \"\",\r\n                \"xNome_dest\": \"NF-E EMITIDA EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"IE_dest\": \"\",\r\n                \"ISUF\": \"\",\r\n                \"indIEDest\": 9,\r\n                \"IM_dest\": \"\",\r\n                \"enderDest\": {\r\n                    \"nro_dest\": \"12\",\r\n                    \"xCpl_dest\": \"CENTRO\",\r\n                    \"xBairro_dest\": \"Primavera\",\r\n                    \"xEmail_dest\": \"brunateste@migrate.com.br\",\r\n                    \"xLgr_dest\": \"Rua das Margaridas\",\r\n                    \"xPais_dest\": \"Brasil\",\r\n                    \"cMun_dest\": 4317103,\r\n                    \"xMun_dest\": \"CONTAGEM\",\r\n                    \"UF_dest\": \"RS\",\r\n                    \"CEP_dest\": 32145900,\r\n                    \"cPais_dest\": 1058,\r\n                    \"fone_dest\": 5543481105\r\n                }\r\n            },\r\n            \"retirada\": {\r\n                \"CNPJ_ret\": \"\",\r\n                \"CPF_ret\": \"\",\r\n                \"xNome_ret\": \"\",\r\n                \"xLgr_ret\": \"\",\r\n                \"nro_ret\": \"\",\r\n                \"xCpl_ret\": \"\",\r\n                \"xBairro_ret\": \"\",\r\n                \"xMun_ret\": \"\",\r\n                \"cMun_ret\": 0,\r\n                \"UF_ret\": \"\",\r\n                \"CEP_ret\": 0,\r\n                \"cPais_ret\": 0,\r\n                \"xPais_ret\": \"\",\r\n                \"fone_ret\": 0,\r\n                \"email_ret\": \"\",\r\n                \"IE_ret\": \"\"\r\n            },\r\n            \"entrega\": {\r\n                \"CNPJ_entr\": \"\",\r\n                \"CPF_entr\": \"\",\r\n                \"xLgr_entr\": \"\",\r\n                \"nro_entr\": \"\",\r\n                \"xCpl_entr\": \"\",\r\n                \"xBairro_entr\": \"\",\r\n                \"cMun_entr\": 0,\r\n                \"xMun_entr\": \"\",\r\n                \"UF_entr\": \"\",\r\n                \"xNome_entr\": \"\",\r\n                \"CEP_entr\": 0,\r\n                \"cPais_entr\": 0,\r\n                \"xPais_entr\": \"\",\r\n                \"fone_entr\": 0,\r\n                \"email_entr\": \"\",\r\n                \"IE_entr\": \"\"\r\n            },\r\n            \"autXML\": [\r\n                {\r\n                    \"CNPJ_aut\": \"\",\r\n                    \"CPF_aut\": \"16928779098\"\r\n                }\r\n            ],\r\n            \"det\": [\r\n                {\r\n                    \"infADProd\": \"Informações Adicionais\",\r\n                    \"prod\": {\r\n                        \"cProd\": \"1\",\r\n                        \"cEAN\": \"SEM GTIN\",\r\n                        \"xProd\": \"NOTA FISCAL EMITIDA EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                        \"NCM\": \"39231090\",\r\n                        \"EXTIPI\": \"\",\r\n                        \"CFOP\": 5102,\r\n                        \"uCOM\": \"UN\",\r\n                        \"qCOM\": \"1.0000\",\r\n                        \"vUnCom\": \"10.0000000000\",\r\n                        \"vProd\": \"10.00\",\r\n                        \"cEANTrib\": \"SEM GTIN\",\r\n                        \"uTrib\": \"UN\",\r\n                        \"qTrib\": \"1.0000\",\r\n                        \"vUnTrib\": \"10.0000000000\",\r\n                        \"vFrete\": \"0.00\",\r\n                        \"vSeg\": \"0.00\",\r\n                        \"vDesc\": \"0.00\",\r\n                        \"vOutro_item\": \"0.00\",\r\n                        \"indTot\": 1,\r\n                        \"nTipoItem\": 0,\r\n                        \"dProd\": 0,\r\n                        \"xPed_item\": \"\",\r\n                        \"nItemPed\": \"\",\r\n                        \"nFCI\": \"\",\r\n                        \"nRECOPI\": \"\",\r\n                        \"CEST\": \"1234567\",\r\n                        \"cBenef\": \"RS051011\"\r\n                    },\r\n                    \"imposto\": {\r\n                        \"vTotTrib\": \"\",\r\n                        \"ICMS\": {\r\n                            \"orig\": 0,\r\n                            \"CST\": \"40\",\r\n                            \"modBC\": \"\",\r\n                            \"vBC\": \"0.00\",\r\n                            \"pICMS\": 0,\r\n                            \"vICMS_icms\": \"0.00\",\r\n                            \"modBCST\": 0,\r\n                            \"pMVAST\": 0,\r\n                            \"pRedBCST\": 0,\r\n                            \"vBCST\": \"0.00\",\r\n                            \"vBCSTRet\": \"0.00\",\r\n                            \"pICMSST\": 0,\r\n                            \"vICMSST_icms\": \"0.00\",\r\n                            \"vICMSSTRet\": \"0.00\",\r\n                            \"pRedBC\": 0,\r\n                            \"motDesICMS\": 0,\r\n                            \"vICMSDeson\": \"0.00\",\r\n                            \"vICMSOp\": \"0.00\",\r\n                            \"pDif\": 0,\r\n                            \"vICMSDif\": \"0.00\",\r\n                            \"pBCOp\": 0,\r\n                            \"UFST\": \"\",\r\n                            \"vBCSTDest\": \"0.00\",\r\n                            \"vICMSSTDest_icms\": \"0.00\",\r\n                            \"pCredSN\": 0,\r\n                            \"vCredICMSSN\": \"0.00\",\r\n                            \"pFCP\": 0,\r\n                            \"vFCP\": \"0.00\",\r\n                            \"vBCFCP\": \"0.00\",\r\n                            \"vBCFCPST\": \"0.00\",\r\n                            \"pFCPST\": 0,\r\n                            \"vFCPST\": \"0.00\",\r\n                            \"pST\": 0,\r\n                            \"vICMSSubstituto\": \"0.00\",\r\n                            \"vBCFCPSTRet\": \"0.00\",\r\n                            \"pFCPSTRet\": 0,\r\n                            \"vFCPSTRet\": \"0.00\",\r\n                            \"GerarICMSST\": \"\",\r\n                            \"pRedBCEfet\": 0,\r\n                            \"vBCEfet\": \"0.00\",\r\n                            \"pICMSEfet\": 0,\r\n                            \"vICMSEfet\": \"0.00\"\r\n                        },\r\n                        \"IPI\": {\r\n                            \"CNPJProd\": \"\",\r\n                            \"cSelo\": \"\",\r\n                            \"qSelo\": 0,\r\n                            \"cEnq\": \"\",\r\n                            \"CSTIPI\": {\r\n                                \"CST_IPI\": \"\",\r\n                                \"vBC_IPI\": \"0.00\",\r\n                                \"qUnid_IPI\": \"0.0000\",\r\n                                \"vUnid_IPI\": \"0.0000\",\r\n                                \"pIPI\": 0,\r\n                                \"vIPI\": \"0.00\"\r\n                            }\r\n                        }\r\n                    }\r\n                }\r\n            ],\r\n            \"total\": {\r\n                \"ICMStot\": {\r\n                    \"vBC_ttlnfe\": \"0.00\",\r\n                    \"vICMS_ttlnfe\": \"0.00\",\r\n                    \"vICMSDeson_ttlnfe\": \"0.00\",\r\n                    \"vBCST_ttlnfe\": \"0.00\",\r\n                    \"vST_ttlnfe\": \"0.00\",\r\n                    \"vProd_ttlnfe\": \"10.00\",\r\n                    \"vFrete_ttlnfe\": \"0.00\",\r\n                    \"vSeg_ttlnfe\": \"0.00\",\r\n                    \"vDesc_ttlnfe\": \"0.00\",\r\n                    \"vII_ttlnfe\": \"0.00\",\r\n                    \"vIPI_ttlnfe\": \"0.00\",\r\n                    \"vPIS_ttlnfe\": \"0.00\",\r\n                    \"vCOFINS_ttlnfe\": \"0.00\",\r\n                    \"vOutro\": \"0.00\",\r\n                    \"vNF\": \"10.00\",\r\n                    \"vTotTrib_ttlnfe\": \"0\"\r\n                }\r\n            },\r\n            \"transp\": {\r\n                \"modFrete\": 2,\r\n                \"balsa\": \"\",\r\n                \"vagao\": \"\",\r\n                \"transporta\": {\r\n                    \"CNPJ_transp\": \"\",\r\n                    \"CPF_transp\": \"67492234007\",\r\n                    \"xNome_transp\": \"Transportador\",\r\n                    \"IE_transp\": \"\",\r\n                    \"xEnder\": \"\",\r\n                    \"xMun_transp\": \"\",\r\n                    \"UF_transp\": \"RS\"\r\n                },\r\n                \"retTransp\": {\r\n                    \"vServ_transp\": \"0.00\",\r\n                    \"vBCRet\": \"0.00\",\r\n                    \"pICMSRet\": 0,\r\n                    \"vICMSRet\": \"0.00\",\r\n                    \"CFOP_transp\": 0,\r\n                    \"cMunFG_transp\": 0\r\n                },\r\n                \"veicTransp\": {\r\n                    \"placa\": \"\",\r\n                    \"UF_veictransp\": \"\",\r\n                    \"RNTC\": \"\"\r\n                },\r\n                \"reboque\": [],\r\n                \"vol\": []\r\n            },\r\n            \"cobr\": {\r\n                \"fat\": {\r\n                    \"nFat\": \"\",\r\n                    \"vOrig\": \"0.00\",\r\n                    \"vDesc_cob\": \"0.00\",\r\n                    \"vLiq\": \"0.00\"\r\n                },\r\n                \"dup\": []\r\n            },\r\n            \"pag\": [\r\n                {\r\n                    \"indPag_pag\": \"\",\r\n                    \"tPag\": \"16\",\r\n                    \"vPag\": \"10.00\",\r\n                    \"card\": {\r\n                        \"tipoIntegracao\": 0,\r\n                        \"CNPJ_card\": \"\",\r\n                        \"tBand\": \"\",\r\n                        \"cAut\": \"\"\r\n                    }\r\n                }\r\n            ],\r\n            \"vTroco\": \"0.00\",\r\n            \"infSuplem\": {\r\n                \"qrCode\": \"\",\r\n                \"urlChave\": \"\"\r\n            },\r\n            \"RespTecnico\": {\r\n                \"CNPJ\": \"87610802000100\",\r\n                \"xContato\": \"Joao Carlos da Cunha Pereira\",\r\n                \"email\": \"testemigrate1@gmail.com\",\r\n                \"fone\": 35351285,\r\n                \"idCSRT\": \"12\",\r\n                \"hashCSRT\": \"v61uDiGY1KWigdgMfHGYi6qilQo=\"\r\n            }\r\n        }\r\n    }\r\n]";
+        private string NFCeSefaz = "";
+        private string NFCeDescarte = "[\r\n    {\r\n        \"ModeloDocumento\": \"NFCe\",\r\n        \"tpAmb\": 2,\r\n        \"CNPJ_emit\": \"06354976000149\",\r\n        \"Numero\": 25,\r\n        \"Serie\": \"245\",\r\n        \"NumeroSubstituta\": 26,\r\n        \"SerieSubstituta\": \"245\"\r\n    }\r\n]";
+        private string NFCeCancelamento = "[\n\t{\n\t\t\"ModeloDocumento\": \"NFCe\",\n\t\t\"Versao\": \"4.00\",\n\t\t\"Evento\": {\n\t      \"NtfCnpjEmissor\": \"06354976000149\",\n\t      \"NtfNumero\": \"52465495\",\n\t      \"NtfSerie\": \"251\",\n\t      \"tpAmb\": \"2\",\n\t      \"EveInf\": {\n\t         \"EveDh\": \"2020-05-20T11:00:00-03:00\",\n\t         \"EveFusoHorario\": \"-03:00\",\n\t         \"EveTp\": \"110111\",\n\t         \"EvenSeq\": \"1\",\n\t         \"Evedet\": {\n\t            \"EveDesc\": \"Cancelamento\",\n\t            \"EvenProt\": \"143200000317337\",\n\t            \"EvexJust\": \"Aqui vai a justificativa de cancelamento\"\n\t         }\n\t      }\n\t\t}\n\t}\n]";
+        private string NFCeCartaCorrecao = "";
+        private string NFCeInutilizacao = "";
+        private string NFCeExport = "[\\n\\t{\\n\\t\\t\\\"ModeloDocumento\\\": \\\"nfce\\\",\\n\\t\\t\\\"Versao\\\": \\\"4.0\\\",\\n\\t\\t\\\"tpAmb\\\": 2,\\n\\t\\t\\\"CnpjEmpresa\\\": \\\"06354976000149\\\",\\n\\t\\t\\\"StatutsDocumento\\\": 2,\\n\\t\\t\\\"NumeroInicial\\\": 1,\\n\\t\\t\\\"NumeroFinal\\\": 270,\\n\\t\\t\\\"Serie\\\": \\\"650\\\",\\n\\t\\t\\\"ResumoDocumento\\\": \\\"S\\\",\\n\\t\\t\\\"ParmSituacao\\\": \\\"S\\\",\\n\\t\\t\\\"ParmXMLCompleto\\\": \\\"S\\\",\\n\\t\\t\\\"ParmEventos\\";
+        private string NFCeExportV2 = "[\n\t{\n\t\t\"ModeloDocumento\": \"nfce\",\n\t\t\"Versao\": \"4.0\",\n\t\t\"tpAmb\": 2,\n\t\t\"CnpjEmpresa\": \"06354976000149\",\n\t\t\"StatutsDocumento\": 2,\n\t\t\"NumeroInicial\": 1,\n\t\t\"NumeroFinal\": 270,\n\t\t\"Serie\": \"650\",\n\t\t\"ResumoDocumento\": \"S\",\n\t\t\"ParmSituacao\": \"S\",\n\t\t\"ParmXMLCompleto\": \"S\",\n\t\t\"ParmEventos\": \"S\"\n\t}\n]";
+        private string NFCeSubstituicao = "";
+        private string NFCeEvento = "[\n\t{\n\t\t\"ModeloDocumento\": \"NFCe\",\n\t\t\"Versao\": \"4.00\",\n\t\t\"Evento\": {\n\t      \"NtfCnpjEmissor\": \"06354976000149\",\n\t      \"NtfNumero\": \"52465495\",\n\t      \"NtfSerie\": \"251\",\n\t      \"tpAmb\": \"2\",\n\t      \"EveInf\": {\n\t         \"EveDh\": \"2020-05-20T11:00:00-03:00\",\n\t         \"EveFusoHorario\": \"-03:00\",\n\t         \"EveTp\": \"110111\",\n\t         \"EvenSeq\": \"1\",\n\t         \"Evedet\": {\n\t            \"EveDesc\": \"Cancelamento\",\n\t            \"EvenProt\": \"143200000317337\",\n\t            \"EvexJust\": \"Aqui vai a justificativa de cancelamento\"\n\t         }\n\t      }\n\t\t}\n\t}\n]";
+        private string NFCeImportacao = "[\n\t{\n\t\t\"Documento\" : \"<nfeProc versao=\\\"4.00\\\" xmlns=\\\"http://www.portalfiscal.inf.br/nfe\\\"><NFe xmlns=\\\"http://www.portalfiscal.inf.br/nfe\\\"><infNFe versao=\\\"4.00\\\" Id=\\\"NFe43201206354976000149652510524655221000000025\\\"><ide><cUF>43</cUF><cNF>00000002</cNF><natOp>NFE CST 00</natOp><mod>65</mod><serie>1</serie><nNF>1</nNF><dhEmi>0000-00-00T00:00:00</dhEmi><tpNF>1</tpNF><idDest>1</idDest><cMunFG>4309605</cMunFG><tpImp>4</tpImp><tpEmis>1</tpEmis><cDV>5</cDV><tpAmb>2</tpAmb><finNFe>1</finNFe><indFinal>1</indFinal><indPres>1</indPres><procEmi>0</procEmi><verProc>MigrateInvoiCy</verProc></ide><emit><CNPJ>06354976000149</CNPJ><xNome>MIGRATE COMPANY SISTEMAS DE INFORMACAO</xNome><xFant>EMPRESA DE TESTE</xFant><enderEmit><xLgr>AAAAAAAAAAAA</xLgr><nro>1700</nro><xCpl>ANDAR 02</xCpl><xBairro>BARRA DO JACARE</xBairro><cMun>4321808</cMun><xMun>ENCANTADO</xMun><UF>RS</UF><CEP>95960000</CEP><cPais>1058</cPais><xPais>BRASIL</xPais><fone>5137513322</fone></enderEmit><IE>1470049241</IE><CRT>3</CRT></emit><dest><CNPJ>99999999000191</CNPJ><xNome>NF-E EMITIDA EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL</xNome><enderDest><xLgr>AAAAAAAAAAAAAAAAAA</xLgr><nro>67</nro><xCpl>CENTRO</xCpl><xBairro>KENNEDY</xBairro><cMun>4321808</cMun><xMun>Tres de Maio</xMun><UF>RS</UF><CEP>98910000</CEP><cPais>1058</cPais><xPais>Brasil</xPais><fone>999999999</fone></enderDest><indIEDest>9</indIEDest><email>aaaaa@aaaa.aaa</email></dest><det nItem=\\\"1\\\"><prod><cProd>00009</cProd><cEAN>SEM GTIN</cEAN><xProd>NOTA FISCAL EMITIDA EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL</xProd><NCM>87113000</NCM><EXTIPI>87</EXTIPI><CFOP>5101</CFOP><uCom>BOT</uCom><qCom>1.0000</qCom><vUnCom>100.0000000000</vUnCom><vProd>100.00</vProd><cEANTrib>SEM GTIN</cEANTrib><uTrib>BOT</uTrib><qTrib>1.0000</qTrib><vUnTrib>100.0000000000</vUnTrib><indTot>1</indTot></prod><imposto><vTotTrib>0.00</vTotTrib><ICMS><ICMS00><orig>0</orig><CST>00</CST><modBC>3</modBC><vBC>100.00</vBC><pICMS>17.0000</pICMS><vICMS>17.00</vICMS></ICMS00></ICMS><PIS><PISNT><CST>07</CST></PISNT></PIS><COFINS><COFINSNT><CST>07</CST></COFINSNT></COFINS></imposto><infAdProd>informacoes adicionais do produto</infAdProd></det><total><ICMSTot><vBC>100.00</vBC><vICMS>17.00</vICMS><vICMSDeson>0.00</vICMSDeson><vFCP>0.00</vFCP><vBCST>0.00</vBCST><vST>0.00</vST><vFCPST>0.00</vFCPST><vFCPSTRet>0.00</vFCPSTRet><vProd>100.00</vProd><vFrete>0.00</vFrete><vSeg>0.00</vSeg><vDesc>0.00</vDesc><vII>0.00</vII><vIPI>0.00</vIPI><vIPIDevol>0.00</vIPIDevol><vPIS>0.00</vPIS><vCOFINS>0</vCOFINS><vOutro>0.00</vOutro><vNF>100.00</vNF><vTotTrib>0.00</vTotTrib></ICMSTot></total><transp><modFrete>9</modFrete></transp><pag><detPag><tPag>01</tPag><vPag>100.00</vPag></detPag></pag><infAdic><infAdFisco>INFORMACOES DE INTERESSE DO FISCO</infAdFisco><infCpl>informacoes complementares da nota</infCpl></infAdic></infNFe><infNFeSupl><qrCode><![CDATA[https://www.sefaz.rs.gov.br/NFCE/NFCE-COM.aspx?p=43201206354976000149652510524655221000000025|2|2|1|BAA3EEC60CCD7C999116DF0805BAA98C6DA23BB2]]></qrCode><urlChave><![CDATA[www.sefaz.rs.gov.br/nfce/consulta]]></urlChave></infNFeSupl><Signature xmlns=\\\"http://www.w3.org/2000/09/xmldsig#\\\"><SignedInfo><CanonicalizationMethod Algorithm=\\\"http://www.w3.org/TR/2001/REC-xml-c14n-20010315\\\" /><SignatureMethod Algorithm=\\\"http://www.w3.org/2000/09/xmldsig#rsa-sha1\\\" /><Reference URI=\\\"#NFe43201206354976000149652510524655221000000025\\\"><Transforms><Transform Algorithm=\\\"http://www.w3.org/2000/09/xmldsig#enveloped-signature\\\" /><Transform Algorithm=\\\"http://www.w3.org/TR/2001/REC-xml-c14n-20010315\\\" /></Transforms><DigestMethod Algorithm=\\\"http://www.w3.org/2000/09/xmldsig#sha1\\\" /><DigestValue>pwSVTayc3xZcl0jkjfOvpZcm5CU=</DigestValue></Reference></SignedInfo><SignatureValue>UbuijwQbjaBgyIOUWjKS25q2H2FQXUCsAxcjhuTUsssIOV/Sl8ddQ4+VJP1/eFv5woOHIk7lgVAJofjRwzJQ1GlNJ0D2eatoTH5GKDKdS5w0kx2zR/Gl02waKzHMAmK3oXl3W03ZbXjy6OBHc0+PArD23k+TsHCl+1g4PUmlh9xDN6/ncYwjS1Tfq9fqQV50xiF2ttTuxHRUEQz4xzQ9ik6t+mMHsfWSq5VIiEcpo/Hau8aVosr8X+v8HhOcleVacyBdjzP0B9TfyQVOsL2AEpc7fQsqJKCQyUK8u46ZDlfo+fCiUVYrbKnuEWeq/k0/0mw34WLtTXl12YApV7gSxQ==</SignatureValue><KeyInfo><X509Data><X509Certificate>MIIH5DCCBcygAwIBAgIIPqwgASE+wgQwDQYJKoZIhvcNAQELBQAwgYkxCzAJBgNVBAYTAkJSMRMwEQYDVQQKEwpJQ1AtQnJhc2lsMTQwMgYDVQQLEytBdXRvcmlkYWRlIENlcnRpZmljYWRvcmEgUmFpeiBCcmFzaWxlaXJhIHYyMRIwEAYDVQQLEwlBQyBTT0xVVEkxGzAZBgNVBAMTEkFDIFNPTFVUSSBNdWx0aXBsYTAeFw0yMDAxMjExNDM2MTlaFw0yMTAxMjAxNDM2MTlaMIIBBDELMAkGA1UEBhMCQlIxEzARBgNVBAoTCklDUC1CcmFzaWwxNDAyBgNVBAsTK0F1dG9yaWRhZGUgQ2VydGlmaWNhZG9yYSBSYWl6IEJyYXNpbGVpcmEgdjIxEjAQBgNVBAsTCUFDIFNPTFVUSTEbMBkGA1UECxMSQUMgU09MVVRJIE11bHRpcGxhMRcwFQYDVQQLEw4wMzE1MTIwMDAwMDEzMzEaMBgGA1UECxMRQ2VydGlmaWNhZG8gUEogQTExRDBCBgNVBAMTO01JR1JBVEUgQ09NUEFOWSBTSVNURU1BUyBERSBJTkZPUk1BQ0FPIExUREEuOjA2MzU0OTc2MDAwMTQ5MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA0RO3miPcW5w6F0zjuqu87u9XqKUv4AAXHQSMb9XwsfrBOdeFzz3iibWaPAROKMQZIRiPVInH49QiIURb6cTPFUyc7rcoeBG/74z1saSb+KY4NxaKa0YoFpRnJoqyVkCID/WrKqAumktLeNzcQeALFVyycGq3LNl/mNzBZUNB1seu2pzqQwaE+PGzewVFYuKP5/KuOwXdoIlUI31J0lcrgVRsZbPdh/t+jrSsyBQyWEXOCQIdjnT9UIQQPr64TXtDHK5jXJBjmcIxtW/4VjOMCUYCJNhZ7T+DKaFN1u6m2K+EbZiZ3sfJ8mtej9BpLFhtuh6NGPY/Gfr3CXv/d5RyhwIDAQABo4IC0DCCAswwVAYIKwYBBQUHAQEESDBGMEQGCCsGAQUFBzAChjhodHRwOi8vY2NkLmFjc29sdXRpLmNvbS5ici9sY3IvYWMtc29sdXRpLW11bHRpcGxhLXYxLnA3YjAdBgNVHQ4EFgQUzjFHN/G/Syc2JqfQ1+iSecHOLiMwCQYDVR0TBAIwADAfBgNVHSMEGDAWgBQ1rjEU9l7Sek9Y/jSoGmeXCsSbBzBeBgNVHSAEVzBVMFMGBmBMAQIBJjBJMEcGCCsGAQUFBwIBFjtodHRwczovL2NjZC5hY3NvbHV0aS5jb20uYnIvZG9jcy9kcGMtYWMtc29sdXRpLW11bHRpcGxhLnBkZjCB3gYDVR0fBIHWMIHTMD6gPKA6hjhodHRwOi8vY2NkLmFjc29sdXRpLmNvbS5ici9sY3IvYWMtc29sdXRpLW11bHRpcGxhLXYxLmNybDA/oD2gO4Y5aHR0cDovL2NjZDIuYWNzb2x1dGkuY29tLmJyL2xjci9hYy1zb2x1dGktbXVsdGlwbGEtdjEuY3JsMFCgTqBMhkpodHRwOi8vcmVwb3NpdG9yaW8uaWNwYnJhc2lsLmdvdi5ici9sY3IvQUNTT0xVVEkvYWMtc29sdXRpLW11bHRpcGxhLXYxLmNybDAOBgNVHQ8BAf8EBAMCBeAwHQYDVR0lBBYwFAYIKwYBBQUHAwIGCCsGAQUFBwMEMIG4BgNVHREEgbAwga2BGWZpbmFuY2Vpcm9AbWlncmF0ZS5jb20uYnKgIgYFYEwBAwKgGRMXQURJTFNPTiBNT0FDSVIgV0VERElHRU6gGQYFYEwBAwOgEBMOMDYzNTQ5NzYwMDAxNDmgOAYFYEwBAwSgLxMtMjgwNTE5Nzg3NzM1MzA1NjA1MzAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwoBcGBWBMAQMHoA4TDDAwMDAwMDAwMDAwMDANBgkqhkiG9w0BAQsFAAOCAgEAYnQK+oKGphgfnmMTl1A95K2OHiMMkowaHXE0iWNi8pROOvxUnB8Duq/l6n3Zh58JfQolznx4O54olb8JPiKKJ8paFQ8i1rys4inIyJT2rF3N7MF2IC9jGPbMpcf+cxXYR1IYqir7nzrROB0Wlk7kl+nIEN6HUxraDVpeyhrJzH/cNLFYEobthXiYS4UIbH0pfZU2F64QZhnY1YKM6fD22LH22k9dI8VTZWQFQnzeSsutryJZfttn82DfTk6DDqRaYmKIZs9o9ihS3Hftzyq18MSR7k8aDrPGwTAGSBnr46jjKx7BpQ6gZ5YmAHFbv9dtbLUCKh3m0gdr5KerIftpipQ0hxJQPcbP8FUiiDaTwj4SqI+cups/45+aEddAEquXhUaR7ROQUeRMhXAS3+0aBdF+TcvxZIXewNSWh20A01H7f6QOnwoxgfwf/ulEgbRZFnLCW2wDsEr/PZw0rGOzWVQviCb23WbC33x72iFydHBtli969YaQSLhcvDrnQ2QNZphd/Zwvt0/Fd60aokZRGvZQELTuJq2pf73nxfgCZ8bobPUUNlANP6hLAWJPW1vUn8O87KnvdiWruHc6x6KK3RtRw0xAms/Kcai4h48P2KBHsJAmiaAoK/tznZ3ShFBDbQz7wFS1/ucW5IzjxAgXvxBKPoFjAFgDSO2vHwJBnYo=</X509Certificate></X509Data></KeyInfo></Signature></NFe><protNFe versao=\\\"4.00\\\"><infProt><tpAmb>2</tpAmb><verAplic>RSnfce202010200652</verAplic><chNFe>43201206354976000149652510524655221000000025</chNFe><dhRecbto>2020-12-11T16:23:09-03:00</dhRecbto><nProt>143200001216164</nProt><digVal>pwSVTayc3xZcl0jkjfOvpZcm5CU=</digVal><cStat>100</cStat><xMotivo>Autorizado o uso da NF-e</xMotivo></infProt></protNFe></nfeProc>\"\n\t}\n]";
 
-        public string MDFeEnvio = "[\r\n    {\r\n        \"Documento\": {\r\n            \"ModeloDocumento\": \"MDFe\",\r\n            \"Versao\": 3,\r\n            \"ide\": {\r\n                \"cMDF\": 9576,\r\n                \"cUF\": 43,\r\n                \"tpAmb\": 2,\r\n                \"tpEmit\": 1,\r\n                \"tpTransp\": 1,\r\n                \"serie\": \"261\",\r\n                \"nMDF\": 153,\r\n                \"modal\": 1,\r\n                \"dhEmi\": \"2021-02-22T16:29:00\",\r\n                \"fusoHorario\": \"-03:00\",\r\n                \"tpEmis\": 2,\r\n                \"UFIni\": \"RS\",\r\n                \"UFFim\": \"SC\",\r\n                \"dhIniViagem\": \"2021-02-22T16:29:00\",\r\n                \"indCanalVerde\": 0,\r\n                \"indCarregaPosterior\": 0,\r\n                \"MunicipiosCarreg\": [\r\n                    {\r\n                        \"cMunCarreg\": 4321808,\r\n                        \"xMunCarreg\": \"TRES DE MAIO\"\r\n                    },\r\n                    {\r\n                        \"cMunCarreg\": 4310207,\r\n                        \"xMunCarreg\": \"IJUI\"\r\n                    }\r\n                ]\r\n            },\r\n            \"emit\": {\r\n                \"CNPJ_emit\": \"06354976000149\",\r\n                \"CPF_emit\": \"\",\r\n                \"xNome\": \"Migrate\",\r\n                \"xFant\": \"\",\r\n                \"IE\": \"1470049241\",\r\n                \"enderEmit\": {\r\n                    \"xLgr\": \"Rua Padre Cacique\",\r\n                    \"nro\": \"900\",\r\n                    \"xCpl\": \"\",\r\n                    \"xBairro\": \"Centro\",\r\n                    \"cMun\": 4321808,\r\n                    \"xMun\": \"TRES DE MAIO\",\r\n                    \"CEP\": \"98910000\",\r\n                    \"UF\": \"RS\",\r\n                    \"fone\": \"\",\r\n                    \"email\": \"\"\r\n                }\r\n            },\r\n            \"InfModal\": {\r\n                \"versaoModal\": 3,\r\n                \"Rodoviario\": {\r\n                    \"RNTRC\": \"12345678\",\r\n                    \"CIOT\": \"\",\r\n                    \"codAgPorto\": \"UIASDHI332\",\r\n                    \"infCIOT\": [\r\n                        {\r\n                            \"CIOT\": \"123456789412\",\r\n                            \"CPF_CIOT\": \"03935034024\",\r\n                            \"CNPJ_CIOT\": \"\"\r\n                        }\r\n                    ],\r\n                    \"Contratante\": [\r\n                        {\r\n                            \"CPF_Contr\": \"65675003010\",\r\n                            \"CNPJ_Contr\": \"\",\r\n                            \"IdEstrangeiro_Contr\": \"\",\r\n                            \"Nome_Contr\": \"\"\r\n                        }\r\n                    ],\r\n                    \"VeicTracao\": {\r\n                        \"codigo_veiculo\": \"001\",\r\n                        \"placa\": \"KLB0152\",\r\n                        \"tara\": 13000,\r\n                        \"capacidadeKG\": 666,\r\n                        \"capacidadeM3\": 333,\r\n                        \"tipoRodado\": \"01\",\r\n                        \"tipoCarroceria\": \"02\",\r\n                        \"UF_veiculo\": \"RS\",\r\n                        \"RENAVAM\": \"\",\r\n                        \"Proprietario\": {\r\n                            \"CPF_prop\": \"65675003010\",\r\n                            \"CNPJ_prop\": \"\",\r\n                            \"RNTRC_prop\": \"12345679\",\r\n                            \"nome_prop\": \"Gaius Claudius Glaber\",\r\n                            \"IE_prop\": \"123456\",\r\n                            \"UF_prop\": \"RS\",\r\n                            \"tipo_prop\": 0\r\n                        },\r\n                        \"Condutores\": [\r\n                            {\r\n                                \"nome_cond\": \"Lentulus Batiatus\",\r\n                                \"CPF_cond\": \"58691676043\"\r\n                            },\r\n                            {\r\n                                \"nome_cond\": \"Edinaldo\",\r\n                                \"CPF_cond\": \"77657488239\"\r\n                            },\r\n                            {\r\n                                \"nome_cond\": \"Boras\",\r\n                                \"CPF_cond\": \"93244749011\"\r\n                            }\r\n                        ]\r\n                    },\r\n                    \"VeicReboque\": [\r\n                        {\r\n                            \"codigo_reb\": \"0027\",\r\n                            \"placa_reb\": \"IKS5686\",\r\n                            \"tara_reb\": 254,\r\n                            \"capacidadeKG_reb\": 652,\r\n                            \"capacidadeM3_reb\": 255,\r\n                            \"tipoCarroceria_reb\": \"01\",\r\n                            \"UF_reb\": \"RS\",\r\n                            \"RENAVAM\": \"\",\r\n                            \"Proprietario_reb\": {\r\n                                \"CPF_prop_reb\": \"\",\r\n                                \"CNPJ_prop_reb\": \"03702076000157\",\r\n                                \"RNTRC_reb\": \"12345678\",\r\n                                \"nome_prop_reb\": \"Ilithyia\",\r\n                                \"IE_prop_reb\": \"12245645\",\r\n                                \"UF_prop_reb\": \"RS\",\r\n                                \"tipo_prop_reb\": 1\r\n                            }\r\n                        }\r\n                    ],\r\n                    \"ValePedagio\": [\r\n                        {\r\n                            \"CNPJForn\": \"03702076000157\",\r\n                            \"CNPJPagamento\": \"14812749000148\",\r\n                            \"numeroCompra\": \"5544\",\r\n                            \"vValePed\": \"66.80\"\r\n                        }\r\n                    ],\r\n                    \"LacresRodo\": [\r\n                        {\r\n                            \"nLacreRodo\": \"1234567\"\r\n                        },\r\n                        {\r\n                            \"nLacreRodo\": \"343445\"\r\n                        }\r\n                    ]\r\n                }\r\n            },\r\n            \"Documentos\": {\r\n                \"MunicipiosDescarreg\": [\r\n                    {\r\n                        \"cMunDescarreg\": 4205407,\r\n                        \"xMunDescarreg\": \"FLORIANOPOLIS\",\r\n                        \"InfDoc\": [\r\n                            {\r\n                                \"tipoDocumento\": 1,\r\n                                \"chaveAcesso\": \"43130406354976000149572890000000211123456788\",\r\n                                \"segCodBarras\": \"\",\r\n                                \"numero_ct\": \"\",\r\n                                \"serie_ct\": \"\",\r\n                                \"subserie_ct\": 0,\r\n                                \"dEmi_ct\": \"0000-00-00\",\r\n                                \"valorCarga_ct\": \"0.00\",\r\n                                \"CNPJ_nf\": \"\",\r\n                                \"UF_nf\": \"\",\r\n                                \"numero_nf\": \"\",\r\n                                \"serie_nf\": \"\",\r\n                                \"dEmi_nf\": \"0000-00-00\",\r\n                                \"valor_nf\": \"0.00\",\r\n                                \"PINSuframa_nf\": \"\",\r\n                                \"indReentrega\": 1\r\n                            },\r\n                            {\r\n                                \"tipoDocumento\": 1,\r\n                                \"chaveAcesso\": \"43190806354976000149522600000012021000001001\",\r\n                                \"segCodBarras\": \"\",\r\n                                \"numero_ct\": \"\",\r\n                                \"serie_ct\": \"\",\r\n                                \"subserie_ct\": 0,\r\n                                \"dEmi_ct\": \"0000-00-00\",\r\n                                \"valorCarga_ct\": \"0.00\",\r\n                                \"CNPJ_nf\": \"\",\r\n                                \"UF_nf\": \"\",\r\n                                \"numero_nf\": \"\",\r\n                                \"serie_nf\": \"\",\r\n                                \"dEmi_nf\": \"0000-00-00\",\r\n                                \"valor_nf\": \"0.00\",\r\n                                \"PINSuframa_nf\": \"\",\r\n                                \"indReentrega\": 1\r\n                            }\r\n                        ]\r\n                    }\r\n                ]\r\n            },\r\n            \"Seguro\": [\r\n                {\r\n                    \"nApol\": \"4321\",\r\n                    \"Averbacoes\": [\r\n                        {\r\n                            \"nAver\": \"1\"\r\n                        },\r\n                        {\r\n                            \"nAver\": \"2\"\r\n                        }\r\n                    ],\r\n                    \"infResp\": {\r\n                        \"respSeg\": 1,\r\n                        \"CNPJ_respSeg\": \"12345678901344\",\r\n                        \"CPF_respSeg\": \"14743208645\"\r\n                    },\r\n                    \"infSeg\": {\r\n                        \"XSeg\": \"Seguro de Moto\",\r\n                        \"CNPJ_seg\": \"12345678901344\"\r\n                    }\r\n                }\r\n            ],\r\n            \"ProdPredominante\": {\r\n                \"tpCarga\": \"05\",\r\n                \"xProd\": \"Latas de atum\",\r\n                \"cEAN\": \"SEM GTIN\",\r\n                \"NCM\": \"98765432\",\r\n                \"Lotacao\": {\r\n                    \"LocalCarrega\": {\r\n                        \"CEP_LocalCarrega\": \"98910000\",\r\n                        \"Latitude_LocalCarrega\": \"\",\r\n                        \"Longitude_LocalCarrega\": \"\"\r\n                    },\r\n                    \"LocalDescarrega\": {\r\n                        \"CEP_LocalDescarrega\": \"98915000\",\r\n                        \"Latitude_LocalDescarrega\": \"\",\r\n                        \"Longitude_LocalDescarrega\": \"\"\r\n                    }\r\n                }\r\n            },\r\n            \"Totais\": {\r\n                \"qtdCTe\": 2,\r\n                \"qtdCT\": 0,\r\n                \"qtdNFe\": 0,\r\n                \"qtdNF\": 0,\r\n                \"qtdMDFe\": 0,\r\n                \"valorCarga\": \"5666.66\",\r\n                \"codUnidMedida\": \"01\",\r\n                \"qtdCarga\": \"666.0000\"\r\n            }\r\n        }\r\n    }\r\n]";
-        public string MDFeSefaz = "";
-        public string MDFeDescarte = "";
-        public string MDFeCancelamento = "";
-        public string MDFeCartaCorrecao = "";
-        public string MDFeInutilizacao = "";
-        public string MDFeExport = "";
-        public string MDFeExportV2 = "";
-        public string MDFeSubstituicao = "";
-        public string MDFeEvento = "";
-        public string MDFeImportacao = "";
+        private string MDFeEnvio = "[\r\n    {\r\n        \"Documento\": {\r\n            \"ModeloDocumento\": \"MDFe\",\r\n            \"Versao\": 3,\r\n            \"ide\": {\r\n                \"cMDF\": 9576,\r\n                \"cUF\": 43,\r\n                \"tpAmb\": 2,\r\n                \"tpEmit\": 1,\r\n                \"tpTransp\": 1,\r\n                \"serie\": \"261\",\r\n                \"nMDF\": 153,\r\n                \"modal\": 1,\r\n                \"dhEmi\": \"2021-02-22T16:29:00\",\r\n                \"fusoHorario\": \"-03:00\",\r\n                \"tpEmis\": 2,\r\n                \"UFIni\": \"RS\",\r\n                \"UFFim\": \"SC\",\r\n                \"dhIniViagem\": \"2021-02-22T16:29:00\",\r\n                \"indCanalVerde\": 0,\r\n                \"indCarregaPosterior\": 0,\r\n                \"MunicipiosCarreg\": [\r\n                    {\r\n                        \"cMunCarreg\": 4321808,\r\n                        \"xMunCarreg\": \"TRES DE MAIO\"\r\n                    },\r\n                    {\r\n                        \"cMunCarreg\": 4310207,\r\n                        \"xMunCarreg\": \"IJUI\"\r\n                    }\r\n                ]\r\n            },\r\n            \"emit\": {\r\n                \"CNPJ_emit\": \"06354976000149\",\r\n                \"CPF_emit\": \"\",\r\n                \"xNome\": \"Migrate\",\r\n                \"xFant\": \"\",\r\n                \"IE\": \"1470049241\",\r\n                \"enderEmit\": {\r\n                    \"xLgr\": \"Rua Padre Cacique\",\r\n                    \"nro\": \"900\",\r\n                    \"xCpl\": \"\",\r\n                    \"xBairro\": \"Centro\",\r\n                    \"cMun\": 4321808,\r\n                    \"xMun\": \"TRES DE MAIO\",\r\n                    \"CEP\": \"98910000\",\r\n                    \"UF\": \"RS\",\r\n                    \"fone\": \"\",\r\n                    \"email\": \"\"\r\n                }\r\n            },\r\n            \"InfModal\": {\r\n                \"versaoModal\": 3,\r\n                \"Rodoviario\": {\r\n                    \"RNTRC\": \"12345678\",\r\n                    \"CIOT\": \"\",\r\n                    \"codAgPorto\": \"UIASDHI332\",\r\n                    \"infCIOT\": [\r\n                        {\r\n                            \"CIOT\": \"123456789412\",\r\n                            \"CPF_CIOT\": \"03935034024\",\r\n                            \"CNPJ_CIOT\": \"\"\r\n                        }\r\n                    ],\r\n                    \"Contratante\": [\r\n                        {\r\n                            \"CPF_Contr\": \"65675003010\",\r\n                            \"CNPJ_Contr\": \"\",\r\n                            \"IdEstrangeiro_Contr\": \"\",\r\n                            \"Nome_Contr\": \"\"\r\n                        }\r\n                    ],\r\n                    \"VeicTracao\": {\r\n                        \"codigo_veiculo\": \"001\",\r\n                        \"placa\": \"KLB0152\",\r\n                        \"tara\": 13000,\r\n                        \"capacidadeKG\": 666,\r\n                        \"capacidadeM3\": 333,\r\n                        \"tipoRodado\": \"01\",\r\n                        \"tipoCarroceria\": \"02\",\r\n                        \"UF_veiculo\": \"RS\",\r\n                        \"RENAVAM\": \"\",\r\n                        \"Proprietario\": {\r\n                            \"CPF_prop\": \"65675003010\",\r\n                            \"CNPJ_prop\": \"\",\r\n                            \"RNTRC_prop\": \"12345679\",\r\n                            \"nome_prop\": \"Gaius Claudius Glaber\",\r\n                            \"IE_prop\": \"123456\",\r\n                            \"UF_prop\": \"RS\",\r\n                            \"tipo_prop\": 0\r\n                        },\r\n                        \"Condutores\": [\r\n                            {\r\n                                \"nome_cond\": \"Lentulus Batiatus\",\r\n                                \"CPF_cond\": \"58691676043\"\r\n                            },\r\n                            {\r\n                                \"nome_cond\": \"Edinaldo\",\r\n                                \"CPF_cond\": \"77657488239\"\r\n                            },\r\n                            {\r\n                                \"nome_cond\": \"Boras\",\r\n                                \"CPF_cond\": \"93244749011\"\r\n                            }\r\n                        ]\r\n                    },\r\n                    \"VeicReboque\": [\r\n                        {\r\n                            \"codigo_reb\": \"0027\",\r\n                            \"placa_reb\": \"IKS5686\",\r\n                            \"tara_reb\": 254,\r\n                            \"capacidadeKG_reb\": 652,\r\n                            \"capacidadeM3_reb\": 255,\r\n                            \"tipoCarroceria_reb\": \"01\",\r\n                            \"UF_reb\": \"RS\",\r\n                            \"RENAVAM\": \"\",\r\n                            \"Proprietario_reb\": {\r\n                                \"CPF_prop_reb\": \"\",\r\n                                \"CNPJ_prop_reb\": \"03702076000157\",\r\n                                \"RNTRC_reb\": \"12345678\",\r\n                                \"nome_prop_reb\": \"Ilithyia\",\r\n                                \"IE_prop_reb\": \"12245645\",\r\n                                \"UF_prop_reb\": \"RS\",\r\n                                \"tipo_prop_reb\": 1\r\n                            }\r\n                        }\r\n                    ],\r\n                    \"ValePedagio\": [\r\n                        {\r\n                            \"CNPJForn\": \"03702076000157\",\r\n                            \"CNPJPagamento\": \"14812749000148\",\r\n                            \"numeroCompra\": \"5544\",\r\n                            \"vValePed\": \"66.80\"\r\n                        }\r\n                    ],\r\n                    \"LacresRodo\": [\r\n                        {\r\n                            \"nLacreRodo\": \"1234567\"\r\n                        },\r\n                        {\r\n                            \"nLacreRodo\": \"343445\"\r\n                        }\r\n                    ]\r\n                }\r\n            },\r\n            \"Documentos\": {\r\n                \"MunicipiosDescarreg\": [\r\n                    {\r\n                        \"cMunDescarreg\": 4205407,\r\n                        \"xMunDescarreg\": \"FLORIANOPOLIS\",\r\n                        \"InfDoc\": [\r\n                            {\r\n                                \"tipoDocumento\": 1,\r\n                                \"chaveAcesso\": \"43130406354976000149572890000000211123456788\",\r\n                                \"segCodBarras\": \"\",\r\n                                \"numero_ct\": \"\",\r\n                                \"serie_ct\": \"\",\r\n                                \"subserie_ct\": 0,\r\n                                \"dEmi_ct\": \"0000-00-00\",\r\n                                \"valorCarga_ct\": \"0.00\",\r\n                                \"CNPJ_nf\": \"\",\r\n                                \"UF_nf\": \"\",\r\n                                \"numero_nf\": \"\",\r\n                                \"serie_nf\": \"\",\r\n                                \"dEmi_nf\": \"0000-00-00\",\r\n                                \"valor_nf\": \"0.00\",\r\n                                \"PINSuframa_nf\": \"\",\r\n                                \"indReentrega\": 1\r\n                            },\r\n                            {\r\n                                \"tipoDocumento\": 1,\r\n                                \"chaveAcesso\": \"43190806354976000149522600000012021000001001\",\r\n                                \"segCodBarras\": \"\",\r\n                                \"numero_ct\": \"\",\r\n                                \"serie_ct\": \"\",\r\n                                \"subserie_ct\": 0,\r\n                                \"dEmi_ct\": \"0000-00-00\",\r\n                                \"valorCarga_ct\": \"0.00\",\r\n                                \"CNPJ_nf\": \"\",\r\n                                \"UF_nf\": \"\",\r\n                                \"numero_nf\": \"\",\r\n                                \"serie_nf\": \"\",\r\n                                \"dEmi_nf\": \"0000-00-00\",\r\n                                \"valor_nf\": \"0.00\",\r\n                                \"PINSuframa_nf\": \"\",\r\n                                \"indReentrega\": 1\r\n                            }\r\n                        ]\r\n                    }\r\n                ]\r\n            },\r\n            \"Seguro\": [\r\n                {\r\n                    \"nApol\": \"4321\",\r\n                    \"Averbacoes\": [\r\n                        {\r\n                            \"nAver\": \"1\"\r\n                        },\r\n                        {\r\n                            \"nAver\": \"2\"\r\n                        }\r\n                    ],\r\n                    \"infResp\": {\r\n                        \"respSeg\": 1,\r\n                        \"CNPJ_respSeg\": \"12345678901344\",\r\n                        \"CPF_respSeg\": \"14743208645\"\r\n                    },\r\n                    \"infSeg\": {\r\n                        \"XSeg\": \"Seguro de Moto\",\r\n                        \"CNPJ_seg\": \"12345678901344\"\r\n                    }\r\n                }\r\n            ],\r\n            \"ProdPredominante\": {\r\n                \"tpCarga\": \"05\",\r\n                \"xProd\": \"Latas de atum\",\r\n                \"cEAN\": \"SEM GTIN\",\r\n                \"NCM\": \"98765432\",\r\n                \"Lotacao\": {\r\n                    \"LocalCarrega\": {\r\n                        \"CEP_LocalCarrega\": \"98910000\",\r\n                        \"Latitude_LocalCarrega\": \"\",\r\n                        \"Longitude_LocalCarrega\": \"\"\r\n                    },\r\n                    \"LocalDescarrega\": {\r\n                        \"CEP_LocalDescarrega\": \"98915000\",\r\n                        \"Latitude_LocalDescarrega\": \"\",\r\n                        \"Longitude_LocalDescarrega\": \"\"\r\n                    }\r\n                }\r\n            },\r\n            \"Totais\": {\r\n                \"qtdCTe\": 2,\r\n                \"qtdCT\": 0,\r\n                \"qtdNFe\": 0,\r\n                \"qtdNF\": 0,\r\n                \"qtdMDFe\": 0,\r\n                \"valorCarga\": \"5666.66\",\r\n                \"codUnidMedida\": \"01\",\r\n                \"qtdCarga\": \"666.0000\"\r\n            }\r\n        }\r\n    }\r\n]";
+        private string MDFeSefaz = "";
+        private string MDFeDescarte = "";
+        private string MDFeCancelamento = "";
+        private string MDFeCartaCorrecao = "";
+        private string MDFeInutilizacao = "";
+        private string MDFeExport = "";
+        private string MDFeExportV2 = "";
+        private string MDFeSubstituicao = "";
+        private string MDFeEvento = "";
+        private string MDFeImportacao = "";
 
-        public string NFSeEnvio = "[\r\n    {\r\n        \"Documento\": {\r\n            \"ModeloDocumento\": \"NFSE\",\r\n            \"Versao\": 1.00,\r\n            \"RPS\": [\r\n                {\r\n                    \"RPSNumero\": 5,\r\n                    \"RPSSerie\": \"251\",\r\n                    \"RPSTipo\": 1,\r\n                    \"dEmis\": \"2021-02-23T14:10:00\",\r\n                    \"dCompetencia\": \"2021-02-23T14:10:00\",\r\n                    \"LocalPrestServ\": 0,\r\n                    \"natOp\": 1,\r\n                    \"Operacao\": \"\",\r\n                    \"NumProcesso\": \"\",\r\n                    \"RegEspTrib\": 6,\r\n                    \"OptSN\": 2,\r\n                    \"IncCult\": 2,\r\n                    \"Status\": 1,\r\n                    \"cVerificaRPS\": \"\",\r\n                    \"EmpreitadaGlobal\": 0,\r\n                    \"tpAmb\": 2,\r\n                    \"RPSSubs\": {\r\n                        \"SubsNumero\": 0,\r\n                        \"SubsSerie\": \"\",\r\n                        \"SubsTipo\": 0,\r\n                        \"SubsNFSeNumero\": 0,\r\n                        \"SubsDEmisNFSe\": \"0000-00-00T00:00:00\"\r\n                    },\r\n                    \"Prestador\": {\r\n                        \"CNPJ_prest\": \"06354976000149\",\r\n                        \"xNome\": \"MIGRATE\",\r\n                        \"xFant\": \"MOBIZZ\",\r\n                        \"IM\": \"132236721\",\r\n                        \"IE\": \"1470049241\",\r\n                        \"CMC\": \"132236721\",\r\n                        \"enderPrest\": {\r\n                            \"TPEnd\": \"RUA\",\r\n                            \"xLgr\": \"R 14 DE JULHO\",\r\n                            \"nro\": \"72\",\r\n                            \"xCpl\": \"2 ANDAR\",\r\n                            \"xBairro\": \"CENTRO\",\r\n                            \"cMun\": 4321808,\r\n                            \"UF\": \"RS\",\r\n                            \"CEP\": 98700000,\r\n                            \"fone\": \"5535354830\",\r\n                            \"Email\": \"\"\r\n                        }\r\n                    },\r\n                    \"ListaItens\": [\r\n                        {\r\n                            \"ItemSeq\": 1,\r\n                            \"ItemCod\": \"1\",\r\n                            \"ItemDesc\": \"\",\r\n                            \"ItemQtde\": 1,\r\n                            \"ItemvUnit\": 100.00,\r\n                            \"ItemuMed\": \"UN\",\r\n                            \"ItemvlDed\": 0,\r\n                            \"ItemTributavel\": \"\",\r\n                            \"ItemcCnae\": \"6201500\",\r\n                            \"ItemTributMunicipio\": \"\",\r\n                            \"ItemnAlvara\": \"\",\r\n                            \"ItemvIss\": 3.00,\r\n                            \"ItemvDesconto\": 0.00,\r\n                            \"ItemAliquota\": 0.030000,\r\n                            \"ItemVlrTotal\": 0,\r\n                            \"ItemBaseCalculo\": 100.00,\r\n                            \"ItemvlrISSRetido\": 0,\r\n                            \"ItemIssRetido\": 2,\r\n                            \"ItemRespRetencao\": 0,\r\n                            \"ItemIteListServico\": \"01.05\",\r\n                            \"ItemExigibilidadeISS\": 1,\r\n                            \"ItemcMunIncidencia\": 0,\r\n                            \"ItemNumProcesso\": \"\",\r\n                            \"ItemDedTipo\": \"\",\r\n                            \"ItemDedCPFRef\": \"\",\r\n                            \"ItemDedCNPJRef\": \"\",\r\n                            \"ItemDedNFRef\": 0,\r\n                            \"ItemDedvlTotRef\": 0,\r\n                            \"ItemDedPer\": 0,\r\n                            \"ItemVlrLiquido\": 100.00,\r\n                            \"ItemValAliqINSS\": 0.0000,\r\n                            \"ItemValINSS\": 0,\r\n                            \"ItemValAliqIR\": 0.0000,\r\n                            \"ItemValIR\": 0,\r\n                            \"ItemValAliqCOFINS\": 0.0000,\r\n                            \"ItemValCOFINS\": 0,\r\n                            \"ItemValAliqCSLL\": 0.0000,\r\n                            \"ItemValCSLL\": 0,\r\n                            \"ItemValAliqPIS\": 0.0000,\r\n                            \"ItemValPIS\": 0,\r\n                            \"ItemRedBC\": 0,\r\n                            \"ItemRedBCRetido\": 0,\r\n                            \"ItemBCRetido\": 0,\r\n                            \"ItemValAliqISSRetido\": 0,\r\n                            \"ItemPaisImpDevido\": \"\",\r\n                            \"ItemJustDed\": \"\",\r\n                            \"ItemvOutrasRetencoes\": 0,\r\n                            \"ItemDescIncondicionado\": 0.00,\r\n                            \"ItemDescCondicionado\": 0.00,\r\n                            \"ItemTotalAproxTribServ\": 0,\r\n                        }\r\n                    ],\r\n                    \"ListaParcelas\": [],\r\n                    \"Servico\": {\r\n                        \"Valores\": {\r\n                            \"ValServicos\": 100.00,\r\n                            \"ValDeducoes\": 0,\r\n                            \"ValPIS\": 0,\r\n                            \"ValBCPIS\": 0,\r\n                            \"ValCOFINS\": 0,\r\n                            \"ValBCCOFINS\": 0,\r\n                            \"ValINSS\": 0,\r\n                            \"ValBCINSS\": 0,\r\n                            \"ValIR\": 0,\r\n                            \"ValBCIRRF\": 0,\r\n                            \"ValCSLL\": 0,\r\n                            \"ValBCCSLL\": 0,\r\n                            \"RespRetencao\": 0,\r\n                            \"Tributavel\": \"\",\r\n                            \"ValISS\": 3.00,\r\n                            \"ISSRetido\": 2,\r\n                            \"ValISSRetido\": 0,\r\n                            \"ValTotal\": 0,\r\n                            \"ValTotalRecebido\": 0,\r\n                            \"ValBaseCalculo\": 100.00,\r\n                            \"ValOutrasRetencoes\": 0,\r\n                            \"ValAliqISS\": 0.030000,\r\n                            \"ValAliqPIS\": 0.0000,\r\n                            \"PISRetido\": 0,\r\n                            \"ValAliqCOFINS\": 0.0000,\r\n                            \"COFINSRetido\": 0,\r\n                            \"ValAliqIR\": 0.0000,\r\n                            \"IRRetido\": 0,\r\n                            \"ValAliqCSLL\": 0.0000,\r\n                            \"CSLLRetido\": 0,\r\n                            \"ValAliqINSS\": 0.0000,\r\n                            \"INSSRetido\": 0,\r\n                            \"ValAliqCpp\": 0,\r\n                            \"CppRetido\": 0,\r\n                            \"ValCpp\": 0,\r\n                            \"OutrasRetencoesRetido\": 0,\r\n                            \"ValBCOutrasRetencoes\": 0,\r\n                            \"ValAliqOutrasRetencoes\": 0,\r\n                            \"ValAliqTotTributos\": 0,\r\n                            \"ValLiquido\": 100.00,\r\n                            \"ValDescIncond\": 0.00,\r\n                            \"ValDescCond\": 0.00,\r\n                            \"ValAcrescimos\": 0,\r\n                            \"ValAliqISSoMunic\": 0.0000,\r\n                            \"InfValPIS\": \"\",\r\n                            \"InfValCOFINS\": \"\",\r\n                            \"ValLiqFatura\": 0,\r\n                            \"ValBCISSRetido\": 0,\r\n                            \"NroFatura\": 0,\r\n                            \"CargaTribValor\": 0,\r\n                            \"CargaTribPercentual\": 0,\r\n                            \"CargaTribFonte\": \"\",\r\n                            \"JustDed\": \"\",\r\n                            \"ValCredito\": 0,\r\n                            \"OutrosImp\": 0,\r\n                            \"ValRedBC\": 0,\r\n                            \"ValRetFederais\": 0,\r\n                            \"ValAproxTrib\": 0,\r\n                            \"NroDeducao\": \"0\"\r\n                        },\r\n                        \"LocalPrestacao\": {\r\n                            \"SerEndTpLgr\": \"\",\r\n                            \"SerEndLgr\": \"\",\r\n                            \"SerEndNumero\": \"\",\r\n                            \"SerEndComplemento\": \"\",\r\n                            \"SerEndBairro\": \"\",\r\n                            \"SerEndxMun\": \"\",\r\n                            \"SerEndcMun\": 0,\r\n                            \"SerEndCep\": 0,\r\n                            \"SerEndSiglaUF\": \"\"\r\n                        },\r\n                        \"IteListServico\": \"01.05\",\r\n                        \"Cnae\": 6201500,\r\n                        \"fPagamento\": \"\",\r\n                        \"tpag\": 0,\r\n                        \"TributMunicipio\": \"\",\r\n                        \"TributMunicDesc\": \"\",\r\n                        \"Discriminacao\": \"MAO DE OBRA\",\r\n                        \"cMun\": 4310207,\r\n                        \"SerQuantidade\": 0,\r\n                        \"SerUnidade\": \"\",\r\n                        \"SerNumAlvara\": \"\",\r\n                        \"PaiPreServico\": \"\",\r\n                        \"cMunIncidencia\": 0,\r\n                        \"dVencimento\": \"0000-00-00T00:00:00\",\r\n                        \"ObsInsPagamento\": \"\",\r\n                        \"ObrigoMunic\": 0,\r\n                        \"TributacaoISS\": 0,\r\n                        \"CodigoAtividadeEconomica\": \"\",\r\n                        \"ServicoViasPublicas\": 0,\r\n                        \"NumeroParcelas\": 0,\r\n                        \"NroOrcamento\": 0,\r\n                        \"CodigoNBS\": \"\"\r\n                    },\r\n                    \"Tomador\": {\r\n                        \"TomaCNPJ\": \"11635726000117\",\r\n                        \"TomaCPF\": \"\",\r\n                        \"TomaIE\": \"\",\r\n                        \"TomaIM\": \"\",\r\n                        \"TomaRazaoSocial\": \"AGROX\",\r\n                        \"TomatpLgr\": \"AV\",\r\n                        \"TomaEndereco\": \"Pde Dehon\",\r\n                        \"TomaNumero\": \"153\",\r\n                        \"TomaComplemento\": \"\",\r\n                        \"TomaBairro\": \"CENTRO\",\r\n                        \"TomacMun\": 4302204,\r\n                        \"TomaxMun\": \"Três de Maio\",\r\n                        \"TomaUF\": \"RS\",\r\n                        \"TomaPais\": \"BR\",\r\n                        \"TomaCEP\": 98910000,\r\n                        \"TomaTelefone\": \"5537481089\",\r\n                        \"TomaTipoTelefone\": \"\",\r\n                        \"TomaEmail\": \"\",\r\n                        \"TomaSite\": \"\",\r\n                        \"TomaIME\": \"\",\r\n                        \"TomaSituacaoEspecial\": \"\",\r\n                        \"DocTomadorEstrangeiro\": \"\",\r\n                        \"TomaRegEspTrib\": 0,\r\n                        \"TomaCadastroMunicipio\": 0,\r\n                        \"TomaOrgaoPublico\": 0\r\n                    },\r\n                    \"IntermServico\": {\r\n                        \"IntermRazaoSocial\": \"\",\r\n                        \"IntermCNPJ\": \"\",\r\n                        \"IntermCPF\": \"\",\r\n                        \"IntermIM\": \"\",\r\n                        \"IntermEmail\": \"\",\r\n                        \"IntermEndereco\": \"\",\r\n                        \"IntermNumero\": \"\",\r\n                        \"IntermComplemento\": \"\",\r\n                        \"IntermBairro\": \"\",\r\n                        \"IntermCep\": 0,\r\n                        \"IntermCmun\": 0,\r\n                        \"IntermXmun\": \"\",\r\n                        \"IntermFone\": \"\",\r\n                        \"IntermIE\": \"\"\r\n                    },\r\n                    \"ConstCivil\": {\r\n                        \"CodObra\": \"\",\r\n                        \"Art\": \"\",\r\n                        \"ObraLog\": \"\",\r\n                        \"ObraCompl\": \"\",\r\n                        \"ObraNumero\": \"\",\r\n                        \"ObraBairro\": \"\",\r\n                        \"ObraCEP\": 0,\r\n                        \"ObraMun\": 0,\r\n                        \"ObraUF\": \"\",\r\n                        \"ObraPais\": \"\",\r\n                        \"ObraCEI\": \"\",\r\n                        \"ObraMatricula\": \"\",\r\n                        \"ObraValRedBC\": 0,\r\n                        \"ObraTipo\": 0,\r\n                        \"ObraNomeFornecedor\": \"\",\r\n                        \"ObraNumeroNF\": 0,\r\n                        \"ObraDataNF\": \"0000-00-00T00:00:00\",\r\n                        \"ObraNumEncapsulamento\": \"\",\r\n                        \"AbatimentoMateriais\": 0,\r\n                        \"ListaMaterial\": []\r\n                    },\r\n                    \"ListaDed\": [],\r\n                    \"Transportadora\": {\r\n                        \"TraNome\": \"\",\r\n                        \"TraCPFCNPJ\": \"\",\r\n                        \"TraIE\": \"\",\r\n                        \"TraPlaca\": \"\",\r\n                        \"TraEnd\": \"\",\r\n                        \"TraMun\": 0,\r\n                        \"TraUF\": \"\",\r\n                        \"TraPais\": \"\",\r\n                        \"TraTipoFrete\": 0\r\n                    },\r\n                    \"NFSOutrasinformacoes\": \"\",\r\n                    \"RPSCanhoto\": 0,\r\n                    \"Arquivo\": \"\",\r\n                    \"ExtensaoArquivo\": \"\"\r\n                }\r\n            ]\r\n        }\r\n    }\r\n]";
-        public string NFSeSefaz = "";
-        public string NFSeDescarte = "";
-        public string NFSeCancelamento = "";
-        public string NFSeCartaCorrecao = "";
-        public string NFSeInutilizacao = "";
-        public string NFSeExport = "";
-        public string NFSeExportV2 = "";
-        public string NFSeSubstituicao = "[\t{ \"Documento\" :{\"ModeloDocumento\":\"NFSE\",\"Versao\":1.00,\"RPS\":[{\"RPSNumero\":1,\"RPSSerie\":\"251\",\"RPSTipo\":1,\"dEmis\":\"2020-05-19T10:47:00\",\"dCompetencia\":\"2020-05-19T10:03:00\",\"LocalPrestServ\":0,\"natOp\":1,\"Operacao\":\"\",\"NumProcesso\":\"\",\"RegEspTrib\":6,\"OptSN\":1,\"IncCult\":2,\"Status\":1,\"cVerificaRPS\":\"\",\"EmpreitadaGlobal\":0,\"tpAmb\":2,\"RPSSubs\":{\"SubsNumero\":0,\"SubsSerie\":\"\",\"SubsTipo\":0,\"SubsNFSeNumero\":0,\"SubsDEmisNFSe\":\"0000-00-00T00:00:00\"},\"Prestador\":{\"CNPJ_prest\":\"24388580000285\",\"xNome\":\"MOBIZZ\",\"xFant\":\"MOBIZZ\",\"IM\":\"726431\",\"IE\":\"0650154223\",\"CMC\":\"\",\"enderPrest\":{\"TPEnd\":\"RUA\",\"xLgr\":\"R 14 DE JULHO\",\"nro\":\"72\",\"xCpl\":\"\",\"xBairro\":\"CENTRO\",\"cMun\":4310207,\"UF\":\"RS\",\"CEP\":98700000,\"fone\":\"\",\"Email\":\"\"}},\"ListaItens\":[{\"ItemSeq\":1,\"ItemCod\":\"1\",\"ItemDesc\":\"\",\"ItemQtde\":1,\"ItemvUnit\":399.00,\"ItemuMed\":\"UN\",\"ItemvlDed\":0,\"ItemTributavel\":\"\",\"ItemcCnae\":\"\",\"ItemTributMunicipio\":\"9512600\",\"ItemnAlvara\":\"\",\"ItemvIss\":11.97,\"ItemvDesconto\":0.00,\"ItemAliquota\":0.0300,\"ItemVlrTotal\":399.00,\"ItemBaseCalculo\":399.00,\"ItemvlrISSRetido\":0,\"ItemIssRetido\":2,\"ItemRespRetencao\":0,\"ItemIteListServico\":\"14.02\",\"ItemExigibilidadeISS\":1,\"ItemcMunIncidencia\":0,\"ItemNumProcesso\":\"\",\"ItemDedTipo\":\"\",\"ItemDedCPFRef\":\"\",\"ItemDedCNPJRef\":\"\",\"ItemDedNFRef\":0,\"ItemDedvlTotRef\":0,\"ItemDedPer\":0,\"ItemVlrLiquido\":399.00,\"ItemValAliqINSS\":0.0000,\"ItemValINSS\":0.00,\"ItemValAliqIR\":0.0000,\"ItemValIR\":0.00,\"ItemValAliqCOFINS\":0.0000,\"ItemValCOFINS\":0.00,\"ItemValAliqCSLL\":0.0000,\"ItemValCSLL\":0.00,\"ItemValAliqPIS\":0.0000,\"ItemValPIS\":0.00,\"ItemRedBC\":0,\"ItemRedBCRetido\":0,\"ItemBCRetido\":0,\"ItemValAliqISSRetido\":0.0000,\"ItemPaisImpDevido\":\"\",\"ItemJustDed\":\"\",\"ItemvOutrasRetencoes\":0,\"ItemDescIncondicionado\":0.00,\"ItemDescCondicionado\":0.00,\"ItemTotalAproxTribServ\":0}],\"ListaParcelas\":[],\"Servico\":{\"Valores\":{\"ValServicos\":399.00,\"ValDeducoes\":0,\"ValPIS\":0.00,\"ValBCPIS\":0,\"ValCOFINS\":0.00,\"ValBCCOFINS\":0,\"ValINSS\":0.00,\"ValBCINSS\":0,\"ValIR\":0.00,\"ValBCIRRF\":0,\"ValCSLL\":0.00,\"ValBCCSLL\":0,\"RespRetencao\":0,\"Tributavel\":\"\",\"ValISS\":11.97,\"ISSRetido\":2,\"ValISSRetido\":0,\"ValTotal\":0,\"ValTotalRecebido\":0,\"ValBaseCalculo\":399.00,\"ValOutrasRetencoes\":0,\"ValAliqISS\":0.030000,\"ValAliqPIS\":0.0000,\"PISRetido\":0,\"ValAliqCOFINS\":0.0000,\"COFINSRetido\":0,\"ValAliqIR\":0.0000,\"IRRetido\":0,\"ValAliqCSLL\":0.0000,\"CSLLRetido\":0,\"ValAliqINSS\":0.0000,\"INSSRetido\":0,\"ValAliqCpp\":0,\"CppRetido\":0,\"ValCpp\":0,\"OutrasRetencoesRetido\":0,\"ValBCOutrasRetencoes\":0,\"ValAliqOutrasRetencoes\":0,\"ValAliqTotTributos\":0,\"ValLiquido\":399.00,\"ValDescIncond\":0.00,\"ValDescCond\":0.00,\"ValAcrescimos\":0,\"ValAliqISSoMunic\":0.0000,\"InfValPIS\":\"\",\"InfValCOFINS\":\"\",\"ValLiqFatura\":0,\"ValBCISSRetido\":0,\"NroFatura\":0,\"CargaTribValor\":0,\"CargaTribPercentual\":0,\"CargaTribFonte\":\"\",\"JustDed\":\"\",\"ValCredito\":0,\"OutrosImp\":0,\"ValRedBC\":0,\"ValRetFederais\":0,\"ValAproxTrib\":0},\"LocalPrestacao\":{\"SerEndTpLgr\":\"\",\"SerEndLgr\":\"\",\"SerEndNumero\":\"\",\"SerEndComplemento\":\"\",\"SerEndBairro\":\"\",\"SerEndxMun\":\"\",\"SerEndcMun\":0,\"SerEndCep\":0,\"SerEndSiglaUF\":\"\"},\"IteListServico\":\"14.02\",\"Cnae\":0,\"fPagamento\":\"\",\"tpag\":0,\"TributMunicipio\":\"9512600\",\"TributMunicDesc\":\"\",\"Discriminacao\":\"MAO DE OBRA\",\"cMun\":4310207,\"SerQuantidade\":0,\"SerUnidade\":\"\",\"SerNumAlvara\":\"\",\"PaiPreServico\":\"BR\",\"cMunIncidencia\":0,\"dVencimento\":\"0000-00-00T00:00:00\",\"ObsInsPagamento\":\"\",\"ObrigoMunic\":0,\"TributacaoISS\":0,\"CodigoAtividadeEconomica\":\"\",\"ServicoViasPublicas\":0,\"NumeroParcelas\":0,\"NroOrcamento\":0,\"CodigoNBS\":\"\"},\"Tomador\":{\"TomaCNPJ\":\"02663790000110\",\"TomaCPF\":\"\",\"TomaIE\":\"\",\"TomaIM\":\"\",\"TomaRazaoSocial\":\"AGROINOX\",\"TomatpLgr\":\"AV\",\"TomaEndereco\":\"Pde Dehon\",\"TomaNumero\":\"153\",\"TomaComplemento\":\"\",\"TomaBairro\":\"CENTRO\",\"TomacMun\":4302204,\"TomaxMun\":\"Boa vista do buricá\",\"TomaUF\":\"RS\",\"TomaPais\":\"BR\",\"TomaCEP\":98918000,\"TomaTelefone\":\"5537481081\",\"TomaTipoTelefone\":\"\",\"TomaEmail\":\"\",\"TomaSite\":\"\",\"TomaIME\":\"\",\"TomaSituacaoEspecial\":\"\",\"DocTomadorEstrangeiro\":\"\",\"TomaRegEspTrib\":0,\"TomaCadastroMunicipio\":0,\"TomaOrgaoPublico\":0},\"IntermServico\":{\"IntermRazaoSocial\":\"\",\"IntermCNPJ\":\"\",\"IntermCPF\":\"\",\"IntermIM\":\"\",\"IntermEmail\":\"\",\"IntermEndereco\":\"\",\"IntermNumero\":\"\",\"IntermComplemento\":\"\",\"IntermBairro\":\"\",\"IntermCep\":0,\"IntermCmun\":0,\"IntermXmun\":\"\",\"IntermFone\":\"\",\"IntermIE\":\"\"},\"ConstCivil\":{\"CodObra\":\"\",\"Art\":\"\",\"ObraLog\":\"\",\"ObraCompl\":\"\",\"ObraNumero\":\"\",\"ObraBairro\":\"\",\"ObraCEP\":0,\"ObraMun\":0,\"ObraUF\":\"\",\"ObraPais\":\"\",\"ObraCEI\":\"\",\"ObraMatricula\":\"\",\"ObraValRedBC\":0,\"ObraTipo\":0,\"ObraNomeFornecedor\":\"\",\"ObraNumeroNF\":0,\"ObraDataNF\":\"0000-00-00T00:00:00\",\"ObraNumEncapsulamento\":\"\",\"AbatimentoMateriais\":0,\"ListaMaterial\":[]},\"ListaDed\":[],\"Transportadora\":{\"TraNome\":\"\",\"TraCPFCNPJ\":\"\",\"TraIE\":\"\",\"TraPlaca\":\"\",\"TraEnd\":\"\",\"TraMun\":0,\"TraUF\":\"\",\"TraPais\":\"\",\"TraTipoFrete\":0},\"NFSOutrasinformacoes\":\"\",\"RPSCanhoto\":0,\"Arquivo\":\"\",\"ExtensaoArquivo\":\"\"}]}}\t]";
-        public string NFSeEvento = "";
-        public string NFSeImportacao = "";
+        private string NFSeEnvio = "[\r\n    {\r\n        \"Documento\": {\r\n            \"ModeloDocumento\": \"NFSE\",\r\n            \"Versao\": 1.00,\r\n            \"RPS\": [\r\n                {\r\n                    \"RPSNumero\": 5,\r\n                    \"RPSSerie\": \"251\",\r\n                    \"RPSTipo\": 1,\r\n                    \"dEmis\": \"2021-02-23T14:10:00\",\r\n                    \"dCompetencia\": \"2021-02-23T14:10:00\",\r\n                    \"LocalPrestServ\": 0,\r\n                    \"natOp\": 1,\r\n                    \"Operacao\": \"\",\r\n                    \"NumProcesso\": \"\",\r\n                    \"RegEspTrib\": 6,\r\n                    \"OptSN\": 2,\r\n                    \"IncCult\": 2,\r\n                    \"Status\": 1,\r\n                    \"cVerificaRPS\": \"\",\r\n                    \"EmpreitadaGlobal\": 0,\r\n                    \"tpAmb\": 2,\r\n                    \"RPSSubs\": {\r\n                        \"SubsNumero\": 0,\r\n                        \"SubsSerie\": \"\",\r\n                        \"SubsTipo\": 0,\r\n                        \"SubsNFSeNumero\": 0,\r\n                        \"SubsDEmisNFSe\": \"0000-00-00T00:00:00\"\r\n                    },\r\n                    \"Prestador\": {\r\n                        \"CNPJ_prest\": \"06354976000149\",\r\n                        \"xNome\": \"MIGRATE\",\r\n                        \"xFant\": \"MOBIZZ\",\r\n                        \"IM\": \"132236721\",\r\n                        \"IE\": \"1470049241\",\r\n                        \"CMC\": \"132236721\",\r\n                        \"enderPrest\": {\r\n                            \"TPEnd\": \"RUA\",\r\n                            \"xLgr\": \"R 14 DE JULHO\",\r\n                            \"nro\": \"72\",\r\n                            \"xCpl\": \"2 ANDAR\",\r\n                            \"xBairro\": \"CENTRO\",\r\n                            \"cMun\": 4321808,\r\n                            \"UF\": \"RS\",\r\n                            \"CEP\": 98700000,\r\n                            \"fone\": \"5535354830\",\r\n                            \"Email\": \"\"\r\n                        }\r\n                    },\r\n                    \"ListaItens\": [\r\n                        {\r\n                            \"ItemSeq\": 1,\r\n                            \"ItemCod\": \"1\",\r\n                            \"ItemDesc\": \"\",\r\n                            \"ItemQtde\": 1,\r\n                            \"ItemvUnit\": 100.00,\r\n                            \"ItemuMed\": \"UN\",\r\n                            \"ItemvlDed\": 0,\r\n                            \"ItemTributavel\": \"\",\r\n                            \"ItemcCnae\": \"6201500\",\r\n                            \"ItemTributMunicipio\": \"\",\r\n                            \"ItemnAlvara\": \"\",\r\n                            \"ItemvIss\": 3.00,\r\n                            \"ItemvDesconto\": 0.00,\r\n                            \"ItemAliquota\": 0.030000,\r\n                            \"ItemVlrTotal\": 0,\r\n                            \"ItemBaseCalculo\": 100.00,\r\n                            \"ItemvlrISSRetido\": 0,\r\n                            \"ItemIssRetido\": 2,\r\n                            \"ItemRespRetencao\": 0,\r\n                            \"ItemIteListServico\": \"01.05\",\r\n                            \"ItemExigibilidadeISS\": 1,\r\n                            \"ItemcMunIncidencia\": 0,\r\n                            \"ItemNumProcesso\": \"\",\r\n                            \"ItemDedTipo\": \"\",\r\n                            \"ItemDedCPFRef\": \"\",\r\n                            \"ItemDedCNPJRef\": \"\",\r\n                            \"ItemDedNFRef\": 0,\r\n                            \"ItemDedvlTotRef\": 0,\r\n                            \"ItemDedPer\": 0,\r\n                            \"ItemVlrLiquido\": 100.00,\r\n                            \"ItemValAliqINSS\": 0.0000,\r\n                            \"ItemValINSS\": 0,\r\n                            \"ItemValAliqIR\": 0.0000,\r\n                            \"ItemValIR\": 0,\r\n                            \"ItemValAliqCOFINS\": 0.0000,\r\n                            \"ItemValCOFINS\": 0,\r\n                            \"ItemValAliqCSLL\": 0.0000,\r\n                            \"ItemValCSLL\": 0,\r\n                            \"ItemValAliqPIS\": 0.0000,\r\n                            \"ItemValPIS\": 0,\r\n                            \"ItemRedBC\": 0,\r\n                            \"ItemRedBCRetido\": 0,\r\n                            \"ItemBCRetido\": 0,\r\n                            \"ItemValAliqISSRetido\": 0,\r\n                            \"ItemPaisImpDevido\": \"\",\r\n                            \"ItemJustDed\": \"\",\r\n                            \"ItemvOutrasRetencoes\": 0,\r\n                            \"ItemDescIncondicionado\": 0.00,\r\n                            \"ItemDescCondicionado\": 0.00,\r\n                            \"ItemTotalAproxTribServ\": 0,\r\n                        }\r\n                    ],\r\n                    \"ListaParcelas\": [],\r\n                    \"Servico\": {\r\n                        \"Valores\": {\r\n                            \"ValServicos\": 100.00,\r\n                            \"ValDeducoes\": 0,\r\n                            \"ValPIS\": 0,\r\n                            \"ValBCPIS\": 0,\r\n                            \"ValCOFINS\": 0,\r\n                            \"ValBCCOFINS\": 0,\r\n                            \"ValINSS\": 0,\r\n                            \"ValBCINSS\": 0,\r\n                            \"ValIR\": 0,\r\n                            \"ValBCIRRF\": 0,\r\n                            \"ValCSLL\": 0,\r\n                            \"ValBCCSLL\": 0,\r\n                            \"RespRetencao\": 0,\r\n                            \"Tributavel\": \"\",\r\n                            \"ValISS\": 3.00,\r\n                            \"ISSRetido\": 2,\r\n                            \"ValISSRetido\": 0,\r\n                            \"ValTotal\": 0,\r\n                            \"ValTotalRecebido\": 0,\r\n                            \"ValBaseCalculo\": 100.00,\r\n                            \"ValOutrasRetencoes\": 0,\r\n                            \"ValAliqISS\": 0.030000,\r\n                            \"ValAliqPIS\": 0.0000,\r\n                            \"PISRetido\": 0,\r\n                            \"ValAliqCOFINS\": 0.0000,\r\n                            \"COFINSRetido\": 0,\r\n                            \"ValAliqIR\": 0.0000,\r\n                            \"IRRetido\": 0,\r\n                            \"ValAliqCSLL\": 0.0000,\r\n                            \"CSLLRetido\": 0,\r\n                            \"ValAliqINSS\": 0.0000,\r\n                            \"INSSRetido\": 0,\r\n                            \"ValAliqCpp\": 0,\r\n                            \"CppRetido\": 0,\r\n                            \"ValCpp\": 0,\r\n                            \"OutrasRetencoesRetido\": 0,\r\n                            \"ValBCOutrasRetencoes\": 0,\r\n                            \"ValAliqOutrasRetencoes\": 0,\r\n                            \"ValAliqTotTributos\": 0,\r\n                            \"ValLiquido\": 100.00,\r\n                            \"ValDescIncond\": 0.00,\r\n                            \"ValDescCond\": 0.00,\r\n                            \"ValAcrescimos\": 0,\r\n                            \"ValAliqISSoMunic\": 0.0000,\r\n                            \"InfValPIS\": \"\",\r\n                            \"InfValCOFINS\": \"\",\r\n                            \"ValLiqFatura\": 0,\r\n                            \"ValBCISSRetido\": 0,\r\n                            \"NroFatura\": 0,\r\n                            \"CargaTribValor\": 0,\r\n                            \"CargaTribPercentual\": 0,\r\n                            \"CargaTribFonte\": \"\",\r\n                            \"JustDed\": \"\",\r\n                            \"ValCredito\": 0,\r\n                            \"OutrosImp\": 0,\r\n                            \"ValRedBC\": 0,\r\n                            \"ValRetFederais\": 0,\r\n                            \"ValAproxTrib\": 0,\r\n                            \"NroDeducao\": \"0\"\r\n                        },\r\n                        \"LocalPrestacao\": {\r\n                            \"SerEndTpLgr\": \"\",\r\n                            \"SerEndLgr\": \"\",\r\n                            \"SerEndNumero\": \"\",\r\n                            \"SerEndComplemento\": \"\",\r\n                            \"SerEndBairro\": \"\",\r\n                            \"SerEndxMun\": \"\",\r\n                            \"SerEndcMun\": 0,\r\n                            \"SerEndCep\": 0,\r\n                            \"SerEndSiglaUF\": \"\"\r\n                        },\r\n                        \"IteListServico\": \"01.05\",\r\n                        \"Cnae\": 6201500,\r\n                        \"fPagamento\": \"\",\r\n                        \"tpag\": 0,\r\n                        \"TributMunicipio\": \"\",\r\n                        \"TributMunicDesc\": \"\",\r\n                        \"Discriminacao\": \"MAO DE OBRA\",\r\n                        \"cMun\": 4310207,\r\n                        \"SerQuantidade\": 0,\r\n                        \"SerUnidade\": \"\",\r\n                        \"SerNumAlvara\": \"\",\r\n                        \"PaiPreServico\": \"\",\r\n                        \"cMunIncidencia\": 0,\r\n                        \"dVencimento\": \"0000-00-00T00:00:00\",\r\n                        \"ObsInsPagamento\": \"\",\r\n                        \"ObrigoMunic\": 0,\r\n                        \"TributacaoISS\": 0,\r\n                        \"CodigoAtividadeEconomica\": \"\",\r\n                        \"ServicoViasPublicas\": 0,\r\n                        \"NumeroParcelas\": 0,\r\n                        \"NroOrcamento\": 0,\r\n                        \"CodigoNBS\": \"\"\r\n                    },\r\n                    \"Tomador\": {\r\n                        \"TomaCNPJ\": \"11635726000117\",\r\n                        \"TomaCPF\": \"\",\r\n                        \"TomaIE\": \"\",\r\n                        \"TomaIM\": \"\",\r\n                        \"TomaRazaoSocial\": \"AGROX\",\r\n                        \"TomatpLgr\": \"AV\",\r\n                        \"TomaEndereco\": \"Pde Dehon\",\r\n                        \"TomaNumero\": \"153\",\r\n                        \"TomaComplemento\": \"\",\r\n                        \"TomaBairro\": \"CENTRO\",\r\n                        \"TomacMun\": 4302204,\r\n                        \"TomaxMun\": \"Três de Maio\",\r\n                        \"TomaUF\": \"RS\",\r\n                        \"TomaPais\": \"BR\",\r\n                        \"TomaCEP\": 98910000,\r\n                        \"TomaTelefone\": \"5537481089\",\r\n                        \"TomaTipoTelefone\": \"\",\r\n                        \"TomaEmail\": \"\",\r\n                        \"TomaSite\": \"\",\r\n                        \"TomaIME\": \"\",\r\n                        \"TomaSituacaoEspecial\": \"\",\r\n                        \"DocTomadorEstrangeiro\": \"\",\r\n                        \"TomaRegEspTrib\": 0,\r\n                        \"TomaCadastroMunicipio\": 0,\r\n                        \"TomaOrgaoPublico\": 0\r\n                    },\r\n                    \"IntermServico\": {\r\n                        \"IntermRazaoSocial\": \"\",\r\n                        \"IntermCNPJ\": \"\",\r\n                        \"IntermCPF\": \"\",\r\n                        \"IntermIM\": \"\",\r\n                        \"IntermEmail\": \"\",\r\n                        \"IntermEndereco\": \"\",\r\n                        \"IntermNumero\": \"\",\r\n                        \"IntermComplemento\": \"\",\r\n                        \"IntermBairro\": \"\",\r\n                        \"IntermCep\": 0,\r\n                        \"IntermCmun\": 0,\r\n                        \"IntermXmun\": \"\",\r\n                        \"IntermFone\": \"\",\r\n                        \"IntermIE\": \"\"\r\n                    },\r\n                    \"ConstCivil\": {\r\n                        \"CodObra\": \"\",\r\n                        \"Art\": \"\",\r\n                        \"ObraLog\": \"\",\r\n                        \"ObraCompl\": \"\",\r\n                        \"ObraNumero\": \"\",\r\n                        \"ObraBairro\": \"\",\r\n                        \"ObraCEP\": 0,\r\n                        \"ObraMun\": 0,\r\n                        \"ObraUF\": \"\",\r\n                        \"ObraPais\": \"\",\r\n                        \"ObraCEI\": \"\",\r\n                        \"ObraMatricula\": \"\",\r\n                        \"ObraValRedBC\": 0,\r\n                        \"ObraTipo\": 0,\r\n                        \"ObraNomeFornecedor\": \"\",\r\n                        \"ObraNumeroNF\": 0,\r\n                        \"ObraDataNF\": \"0000-00-00T00:00:00\",\r\n                        \"ObraNumEncapsulamento\": \"\",\r\n                        \"AbatimentoMateriais\": 0,\r\n                        \"ListaMaterial\": []\r\n                    },\r\n                    \"ListaDed\": [],\r\n                    \"Transportadora\": {\r\n                        \"TraNome\": \"\",\r\n                        \"TraCPFCNPJ\": \"\",\r\n                        \"TraIE\": \"\",\r\n                        \"TraPlaca\": \"\",\r\n                        \"TraEnd\": \"\",\r\n                        \"TraMun\": 0,\r\n                        \"TraUF\": \"\",\r\n                        \"TraPais\": \"\",\r\n                        \"TraTipoFrete\": 0\r\n                    },\r\n                    \"NFSOutrasinformacoes\": \"\",\r\n                    \"RPSCanhoto\": 0,\r\n                    \"Arquivo\": \"\",\r\n                    \"ExtensaoArquivo\": \"\"\r\n                }\r\n            ]\r\n        }\r\n    }\r\n]";
+        private string NFSeSefaz = "";
+        private string NFSeDescarte = "";
+        private string NFSeCancelamento = "";
+        private string NFSeCartaCorrecao = "";
+        private string NFSeInutilizacao = "";
+        private string NFSeExport = "";
+        private string NFSeExportV2 = "";
+        private string NFSeSubstituicao = "[\t{ \"Documento\" :{\"ModeloDocumento\":\"NFSE\",\"Versao\":1.00,\"RPS\":[{\"RPSNumero\":1,\"RPSSerie\":\"251\",\"RPSTipo\":1,\"dEmis\":\"2020-05-19T10:47:00\",\"dCompetencia\":\"2020-05-19T10:03:00\",\"LocalPrestServ\":0,\"natOp\":1,\"Operacao\":\"\",\"NumProcesso\":\"\",\"RegEspTrib\":6,\"OptSN\":1,\"IncCult\":2,\"Status\":1,\"cVerificaRPS\":\"\",\"EmpreitadaGlobal\":0,\"tpAmb\":2,\"RPSSubs\":{\"SubsNumero\":0,\"SubsSerie\":\"\",\"SubsTipo\":0,\"SubsNFSeNumero\":0,\"SubsDEmisNFSe\":\"0000-00-00T00:00:00\"},\"Prestador\":{\"CNPJ_prest\":\"24388580000285\",\"xNome\":\"MOBIZZ\",\"xFant\":\"MOBIZZ\",\"IM\":\"726431\",\"IE\":\"0650154223\",\"CMC\":\"\",\"enderPrest\":{\"TPEnd\":\"RUA\",\"xLgr\":\"R 14 DE JULHO\",\"nro\":\"72\",\"xCpl\":\"\",\"xBairro\":\"CENTRO\",\"cMun\":4310207,\"UF\":\"RS\",\"CEP\":98700000,\"fone\":\"\",\"Email\":\"\"}},\"ListaItens\":[{\"ItemSeq\":1,\"ItemCod\":\"1\",\"ItemDesc\":\"\",\"ItemQtde\":1,\"ItemvUnit\":399.00,\"ItemuMed\":\"UN\",\"ItemvlDed\":0,\"ItemTributavel\":\"\",\"ItemcCnae\":\"\",\"ItemTributMunicipio\":\"9512600\",\"ItemnAlvara\":\"\",\"ItemvIss\":11.97,\"ItemvDesconto\":0.00,\"ItemAliquota\":0.0300,\"ItemVlrTotal\":399.00,\"ItemBaseCalculo\":399.00,\"ItemvlrISSRetido\":0,\"ItemIssRetido\":2,\"ItemRespRetencao\":0,\"ItemIteListServico\":\"14.02\",\"ItemExigibilidadeISS\":1,\"ItemcMunIncidencia\":0,\"ItemNumProcesso\":\"\",\"ItemDedTipo\":\"\",\"ItemDedCPFRef\":\"\",\"ItemDedCNPJRef\":\"\",\"ItemDedNFRef\":0,\"ItemDedvlTotRef\":0,\"ItemDedPer\":0,\"ItemVlrLiquido\":399.00,\"ItemValAliqINSS\":0.0000,\"ItemValINSS\":0.00,\"ItemValAliqIR\":0.0000,\"ItemValIR\":0.00,\"ItemValAliqCOFINS\":0.0000,\"ItemValCOFINS\":0.00,\"ItemValAliqCSLL\":0.0000,\"ItemValCSLL\":0.00,\"ItemValAliqPIS\":0.0000,\"ItemValPIS\":0.00,\"ItemRedBC\":0,\"ItemRedBCRetido\":0,\"ItemBCRetido\":0,\"ItemValAliqISSRetido\":0.0000,\"ItemPaisImpDevido\":\"\",\"ItemJustDed\":\"\",\"ItemvOutrasRetencoes\":0,\"ItemDescIncondicionado\":0.00,\"ItemDescCondicionado\":0.00,\"ItemTotalAproxTribServ\":0}],\"ListaParcelas\":[],\"Servico\":{\"Valores\":{\"ValServicos\":399.00,\"ValDeducoes\":0,\"ValPIS\":0.00,\"ValBCPIS\":0,\"ValCOFINS\":0.00,\"ValBCCOFINS\":0,\"ValINSS\":0.00,\"ValBCINSS\":0,\"ValIR\":0.00,\"ValBCIRRF\":0,\"ValCSLL\":0.00,\"ValBCCSLL\":0,\"RespRetencao\":0,\"Tributavel\":\"\",\"ValISS\":11.97,\"ISSRetido\":2,\"ValISSRetido\":0,\"ValTotal\":0,\"ValTotalRecebido\":0,\"ValBaseCalculo\":399.00,\"ValOutrasRetencoes\":0,\"ValAliqISS\":0.030000,\"ValAliqPIS\":0.0000,\"PISRetido\":0,\"ValAliqCOFINS\":0.0000,\"COFINSRetido\":0,\"ValAliqIR\":0.0000,\"IRRetido\":0,\"ValAliqCSLL\":0.0000,\"CSLLRetido\":0,\"ValAliqINSS\":0.0000,\"INSSRetido\":0,\"ValAliqCpp\":0,\"CppRetido\":0,\"ValCpp\":0,\"OutrasRetencoesRetido\":0,\"ValBCOutrasRetencoes\":0,\"ValAliqOutrasRetencoes\":0,\"ValAliqTotTributos\":0,\"ValLiquido\":399.00,\"ValDescIncond\":0.00,\"ValDescCond\":0.00,\"ValAcrescimos\":0,\"ValAliqISSoMunic\":0.0000,\"InfValPIS\":\"\",\"InfValCOFINS\":\"\",\"ValLiqFatura\":0,\"ValBCISSRetido\":0,\"NroFatura\":0,\"CargaTribValor\":0,\"CargaTribPercentual\":0,\"CargaTribFonte\":\"\",\"JustDed\":\"\",\"ValCredito\":0,\"OutrosImp\":0,\"ValRedBC\":0,\"ValRetFederais\":0,\"ValAproxTrib\":0},\"LocalPrestacao\":{\"SerEndTpLgr\":\"\",\"SerEndLgr\":\"\",\"SerEndNumero\":\"\",\"SerEndComplemento\":\"\",\"SerEndBairro\":\"\",\"SerEndxMun\":\"\",\"SerEndcMun\":0,\"SerEndCep\":0,\"SerEndSiglaUF\":\"\"},\"IteListServico\":\"14.02\",\"Cnae\":0,\"fPagamento\":\"\",\"tpag\":0,\"TributMunicipio\":\"9512600\",\"TributMunicDesc\":\"\",\"Discriminacao\":\"MAO DE OBRA\",\"cMun\":4310207,\"SerQuantidade\":0,\"SerUnidade\":\"\",\"SerNumAlvara\":\"\",\"PaiPreServico\":\"BR\",\"cMunIncidencia\":0,\"dVencimento\":\"0000-00-00T00:00:00\",\"ObsInsPagamento\":\"\",\"ObrigoMunic\":0,\"TributacaoISS\":0,\"CodigoAtividadeEconomica\":\"\",\"ServicoViasPublicas\":0,\"NumeroParcelas\":0,\"NroOrcamento\":0,\"CodigoNBS\":\"\"},\"Tomador\":{\"TomaCNPJ\":\"02663790000110\",\"TomaCPF\":\"\",\"TomaIE\":\"\",\"TomaIM\":\"\",\"TomaRazaoSocial\":\"AGROINOX\",\"TomatpLgr\":\"AV\",\"TomaEndereco\":\"Pde Dehon\",\"TomaNumero\":\"153\",\"TomaComplemento\":\"\",\"TomaBairro\":\"CENTRO\",\"TomacMun\":4302204,\"TomaxMun\":\"Boa vista do buricá\",\"TomaUF\":\"RS\",\"TomaPais\":\"BR\",\"TomaCEP\":98918000,\"TomaTelefone\":\"5537481081\",\"TomaTipoTelefone\":\"\",\"TomaEmail\":\"\",\"TomaSite\":\"\",\"TomaIME\":\"\",\"TomaSituacaoEspecial\":\"\",\"DocTomadorEstrangeiro\":\"\",\"TomaRegEspTrib\":0,\"TomaCadastroMunicipio\":0,\"TomaOrgaoPublico\":0},\"IntermServico\":{\"IntermRazaoSocial\":\"\",\"IntermCNPJ\":\"\",\"IntermCPF\":\"\",\"IntermIM\":\"\",\"IntermEmail\":\"\",\"IntermEndereco\":\"\",\"IntermNumero\":\"\",\"IntermComplemento\":\"\",\"IntermBairro\":\"\",\"IntermCep\":0,\"IntermCmun\":0,\"IntermXmun\":\"\",\"IntermFone\":\"\",\"IntermIE\":\"\"},\"ConstCivil\":{\"CodObra\":\"\",\"Art\":\"\",\"ObraLog\":\"\",\"ObraCompl\":\"\",\"ObraNumero\":\"\",\"ObraBairro\":\"\",\"ObraCEP\":0,\"ObraMun\":0,\"ObraUF\":\"\",\"ObraPais\":\"\",\"ObraCEI\":\"\",\"ObraMatricula\":\"\",\"ObraValRedBC\":0,\"ObraTipo\":0,\"ObraNomeFornecedor\":\"\",\"ObraNumeroNF\":0,\"ObraDataNF\":\"0000-00-00T00:00:00\",\"ObraNumEncapsulamento\":\"\",\"AbatimentoMateriais\":0,\"ListaMaterial\":[]},\"ListaDed\":[],\"Transportadora\":{\"TraNome\":\"\",\"TraCPFCNPJ\":\"\",\"TraIE\":\"\",\"TraPlaca\":\"\",\"TraEnd\":\"\",\"TraMun\":0,\"TraUF\":\"\",\"TraPais\":\"\",\"TraTipoFrete\":0},\"NFSOutrasinformacoes\":\"\",\"RPSCanhoto\":0,\"Arquivo\":\"\",\"ExtensaoArquivo\":\"\"}]}}\t]";
+        private string NFSeEvento = "";
+        private string NFSeImportacao = "";
 
-        public string CTeEnvio = "[\r\n    {\r\n        \"Documento\": {\r\n            \"ModeloDocumento\": \"CTe\",\r\n            \"Versao\": 3.00,\r\n            \"ide\": {\r\n                \"cCT\": 11,\r\n                \"cUF\": 43,\r\n                \"natOp\": \"Aqui vai descrita a Natureza de operação do seu conhecimento\",\r\n                \"CFOP\": 6352,\r\n                \"mod\": \"57\",\r\n                \"serie\": \"305\",\r\n                \"nCT\": 55,\r\n                \"dhEmi\": \"2021-02-22T16:02:14\",\r\n                \"fusoHorario\": \"-03:00\",\r\n                \"tpImp\": 1,\r\n                \"tpEmis\": 1,\r\n                \"tpAmb\": 2,\r\n                \"tpCTe\": 0,\r\n                \"procEmi\": 0,\r\n                \"indGlobalizado\": 0,\r\n                \"refCTE\": \"\",\r\n                \"cMunEnv\": 4310405,\r\n                \"xMunEnv\": \"Independencia\",\r\n                \"UFEnv\": \"RS\",\r\n                \"modal\": \"01\",\r\n                \"tpServ\": 0,\r\n                \"cMunIni\": 4310405,\r\n                \"xMunIni\": \"Independencia\",\r\n                \"UFIni\": \"RS\",\r\n                \"cMunFim\": 3524006,\r\n                \"xMunFim\": \"Boa Vista do Burica\",\r\n                \"UFFim\": \"SP\",\r\n                \"retira\": 0,\r\n                \"xDetRetira\": \"Retirar o pacote no local de entrega indicado\",\r\n                \"indIEToma\": 9,\r\n                \"dhCont\": \"0000-00-00T00:00:00\",\r\n                \"xJust\": \"\",\r\n                \"EmailArquivos\": \"\",\r\n                \"tomador\": {\r\n                    \"toma\": \"4\",\r\n                    \"CNPJ_toma\": \"\",\r\n                    \"CPF_toma\": \"02702315011\",\r\n                    \"IE_toma\": \"\",\r\n                    \"xNome_toma\": \"Ricardo Pieper\",\r\n                    \"xFant_toma\": \"\",\r\n                    \"fone_toma\": \"5597207496\",\r\n                    \"email_toma\": \"\",\r\n                    \"enderTomador\": {\r\n                        \"xLgr_toma\": \"Rua Senador Salgado Filho\",\r\n                        \"nro_toma\": \"463\",\r\n                        \"xCpl_toma\": \"Complemento nao informado\",\r\n                        \"xBairro_toma\": \"Centro\",\r\n                        \"cMun_toma\": 4310405,\r\n                        \"xMun_toma\": \"Independencia\",\r\n                        \"CEP_toma\": \"98915000\",\r\n                        \"UF_toma\": \"RS\",\r\n                        \"cPais_toma\": 1046,\r\n                        \"xPais_toma\": \"Brasil\"\r\n                    }\r\n                }\r\n            },\r\n            \"compl\": {\r\n                \"xCaracAd\": \"Encomenda\",\r\n                \"xCaracSer\": \"Transporte\",\r\n                \"xEmi\": \"Jose Campos\",\r\n                \"origCalc\": \"\",\r\n                \"destCalc\": \"\",\r\n                \"xObs\": \"Observações\",\r\n                \"fluxo\": {\r\n                    \"xOrig\": \"Angra dos Reis\",\r\n                    \"xDest\": \"Maria da Silva de Andrade da Conceição Ferreira Pinto Júnior\",\r\n                    \"xRota\": \"129386\",\r\n                    \"passagem\": [\r\n                        {\r\n                            \"xPass\": \"LHR\"\r\n                        },\r\n                        {\r\n                            \"xPass\": \"CGH\"\r\n                        },\r\n                        {\r\n                            \"xPass\": \"POA\"\r\n                        }\r\n                    ]\r\n                },\r\n                \"Entrega\": {\r\n                    \"tpPer\": \"1\",\r\n                    \"dProg\": \"2021-03-03\",\r\n                    \"dIni\": \"2021-03-03\",\r\n                    \"dFim\": \"2021-03-03\",\r\n                    \"tpHor\": \"1\",\r\n                    \"hProg\": \"10:30:22\",\r\n                    \"hIni\": \"10:30:22\",\r\n                    \"hFim\": \"10:30:22\"\r\n                },\r\n                \"ObsCont\": [],\r\n                \"ObsFisco\": []\r\n            },\r\n            \"emit\": {\r\n                \"CNPJ_emit\": \"06354976000149\",\r\n                \"IE\": \"1470049241\",\r\n                \"IEST\": \"123456789\",\r\n                \"xNome\": \"CT-E EMITIDO EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"xFant\": \"EMPRESA 1\",\r\n\t\t\t\t\"CRT\":\"1\",\r\n                \"enderEmit\": {\r\n                    \"xLgr\": \"End\",\r\n                    \"nro\": \"1234\",\r\n                    \"xCpl\": \"teste\",\r\n                    \"xBairro\": \"Centro\",\r\n                    \"cMun\": 4310405,\r\n                    \"xMun\": \"Independencia\",\r\n                    \"CEP\": \"98910000\",\r\n                    \"UF\": \"RS\",\r\n                    \"fone\": \"35354560\"\r\n                }\r\n            },\r\n            \"rem\": {\r\n                \"CNPJ_rem\": \"06354976000149\",\r\n                \"CPF_rem\": \"\",\r\n                \"IE_rem\": \"1470049241\",\r\n                \"xNome_rem\": \"CT-E EMITIDO EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"xFant_rem\": \"Migrate Company\",\r\n                \"fone_rem\": \"5535355540\",\r\n                \"email_rem\": \"email@email.com.br\",\r\n                \"enderRem\": {\r\n                    \"xLgr_rem\": \"Rua Padre Cacique\",\r\n                    \"nro_rem\": \"556\",\r\n                    \"xCpl_rem\": \"Próximo as Lojas Denise\",\r\n                    \"xBairro_rem\": \"Centro\",\r\n                    \"cMun_rem\": 4310405,\r\n                    \"xMun_rem\": \"Independencia\",\r\n                    \"CEP_rem\": \"98910000\",\r\n                    \"UF_rem\": \"RS\",\r\n                    \"cPais_rem\": 1046,\r\n                    \"xPais_rem\": \"Brasil\"\r\n                }\r\n            },\r\n            \"exped\": {\r\n                \"CNPJ_exp\": \"23640261001810\",\r\n                \"CPF_exp\": \"\",\r\n                \"IE_exp\": \"421025353115\",\r\n                \"xNome_exp\": \"CT-E EMITIDO EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"fone_exp\": \"5535358668\",\r\n                \"email_exp\": \"email@email.com.br\",\r\n                \"enderExped\": {\r\n                    \"xLgr_exp\": \"Rua das Palmeiras\",\r\n                    \"nro_exp\": \"115\",\r\n                    \"xCpl_exp\": \"Próximo ao bar do Zé\",\r\n                    \"xBairro_exp\": \"Centro\",\r\n                    \"cMun_exp\": 3524006,\r\n                    \"xMun_exp\": \"Independencia\",\r\n                    \"CEP_exp\": \"98915000\",\r\n                    \"UF_exp\": \"SP\",\r\n                    \"cPais_exp\": 1046,\r\n                    \"xPais_exp\": \"Brasil\"\r\n                }\r\n            },\r\n            \"receb\": {\r\n                \"CNPJ_rec\": \"\",\r\n                \"CPF_rec\": \"01968897070\",\r\n                \"IE_rec\": \"\",\r\n                \"xNome_rec\": \"CT-E EMITIDO EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"fone_rec\": \"5597207556\",\r\n                \"email_rec\": \"email@email.com.br\",\r\n                \"enderReceb\": {\r\n                    \"xLgr_rec\": \"Rua Senador Salgado Filho\",\r\n                    \"nro_rec\": \"463\",\r\n                    \"xCpl_rec\": \"\",\r\n                    \"xBairro_rec\": \"Centro\",\r\n                    \"cMun_rec\": 3524006,\r\n                    \"xMun_rec\": \"Independência\",\r\n                    \"CEP_rec\": \"98915000\",\r\n                    \"UF_rec\": \"SP\",\r\n                    \"cPais_rec\": 1046,\r\n                    \"xPais_rec\": \"Brasil\"\r\n                }\r\n            },\r\n            \"dest\": {\r\n                \"CNPJ_dest\": \"92554179000147\",\r\n                \"CPF_dest\": \"\",\r\n                \"IE_dest\": \"0080093515\",\r\n                \"xNome_dest\": \"CT-E EMITIDO EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"fone_dest\": \"5599733466\",\r\n                \"ISUF_dest\": \"\",\r\n                \"email_dest\": \"\",\r\n                \"enderDest\": {\r\n                    \"xLgr_dest\": \"Rua Travessa\",\r\n                    \"nro_dest\": \"280\",\r\n                    \"xCpl_dest\": \"Próximo ao hospital\",\r\n                    \"xBairro_dest\": \"Rita\",\r\n                    \"cMun_dest\": 4301602,\r\n                    \"xMun_dest\": \"BAGE\",\r\n                    \"CEP_dest\": \"98910000\",\r\n                    \"UF_dest\": \"RS\",\r\n                    \"cPais_dest\": 1046,\r\n                    \"xPais_dest\": \"Brasil\"\r\n                }\r\n            },\r\n            \"vPrest\": {\r\n                \"vTPrest\": \"32321.23\",\r\n                \"vRec\": \"12312.23\",\r\n                \"Comp\": [\r\n                    {\r\n                        \"xNome_comp\": \"Componente 13\",\r\n                        \"vComp\": \"812.44\"\r\n                    },\r\n                    {\r\n                        \"xNome_comp\": \"Componente 20\",\r\n                        \"vComp\": \"1612.44\"\r\n                    }\r\n                ]\r\n            },\r\n            \"imp\": {\r\n                \"vTotImp\": \"59.90\",\r\n                \"vTotTrib\": \"0.00\",\r\n                \"infAdFisco\": \"Informacoes de interesse do fisco estarão escritas aqui\",\r\n                \"ICMS\": {\r\n                    \"CST\": \"00\",\r\n                    \"vBC\": \"12.00\",\r\n                    \"pICMS\": 13,\r\n                    \"vICMS\": \"1.56\",\r\n                    \"pRedBC\": 0,\r\n                    \"vBCSTRet\": \"0.00\",\r\n                    \"vICMSSTRet\": \"0.00\",\r\n                    \"pICMSSTRet\": 0,\r\n                    \"vCred\": \"0.00\",\r\n                    \"pRedBCOutraUF\": 0,\r\n                    \"vBCOutraUF\": \"0.00\",\r\n                    \"pICMSOutraUF\": 0,\r\n                    \"vICMSOutraUF\": \"0.00\",\r\n                    \"indSN\": 1\r\n                },\r\n                \"ICMSUFFim\": {\r\n                    \"vBCUFFim\": \"0.00\",\r\n                    \"pFCPUFFim\": 0,\r\n                    \"pICMSUFFim\": 0,\r\n                    \"pICMSInter\": 0,\r\n                    \"pICMSInterPart\": 0,\r\n                    \"vFCPUFFim\": \"0.00\",\r\n                    \"vICMSUFIni\": \"0.00\",\r\n                    \"vICMSUFFim\": \"0.00\"\r\n                }\r\n            },\r\n            \"infCTeNorm\": {\r\n                \"infCarga\": {\r\n                    \"vMerc\": \"599.65\",\r\n                    \"proPred\": \"Pepitas de ouro\",\r\n                    \"xOutCat\": \"Que valem mais do que dinheiro\",\r\n                    \"vCargaAverb\": \"200.00\",\r\n                    \"infQ\": [\r\n                        {\r\n                            \"cUnid\": \"01\",\r\n                            \"tpMed\": \"PESO BASE DE CÁLCULO\",\r\n                            \"qCarga\": \"899.0000\"\r\n                        }\r\n                    ]\r\n                },\r\n                \"infDoc\": [\r\n                    {\r\n                        \"tipoDocumento\": 2,\r\n                        \"nRoma_nf\": \"\",\r\n                        \"nPed_nf\": \"\",\r\n                        \"mod_nf\": \"\",\r\n                        \"serie_nf\": \"\",\r\n                        \"nDoc_nf\": \"\",\r\n                        \"dEmi_nf\": \"0000-00-00\",\r\n                        \"vBC_nf\": \"0.00\",\r\n                        \"vICMS_nf\": \"0.00\",\r\n                        \"vBCST_nf\": \"0.00\",\r\n                        \"vST_nf\": \"0.00\",\r\n                        \"vProd_nf\": \"0.00\",\r\n                        \"vNF_nf\": \"0.00\",\r\n                        \"CFOP_nf\": 0,\r\n                        \"peso_nf\": \"0.000\",\r\n                        \"PINSuframa_nf\": \"\",\r\n                        \"dPrev_nf\": \"0000-00-00\",\r\n                        \"chave_nfe\": \"43210106354976000149553050000003831000241294\",\r\n                        \"PINSuframa_nfe\": \"991233459\",\r\n                        \"dPrev_nfe\": \"2017-05-22\",\r\n                        \"tpDoc_outros\": \"\",\r\n                        \"desc_outros\": \"\",\r\n                        \"nDoc_outros\": \"\",\r\n                        \"dEmi_outros\": \"0000-00-00\",\r\n                        \"vDocFisc_outros\": \"0.00\",\r\n                        \"dPrev_outros\": \"0000-00-00\",\r\n                        \"locRet\": {\r\n                            \"CNPJ_locRet\": \"\",\r\n                            \"CPF_locRet\": \"\",\r\n                            \"xNome_locRet\": \"\",\r\n                            \"xLgr_locRet\": \"\",\r\n                            \"nro_locRet\": \"\",\r\n                            \"xCpl_locRet\": \"\",\r\n                            \"xBairro_locRet\": \"\",\r\n                            \"cMun_locRet\": \"\",\r\n                            \"xMun_locRet\": \"\",\r\n                            \"UF_locRet\": \"\"\r\n                        },\r\n                        \"UnidadesTransp\": [\r\n                            {\r\n                                \"tipoUnidTransporte\": 5,\r\n                                \"idUnidTransporte\": \"BCIASUDX988\",\r\n                                \"qtdRateada\": 999,\r\n                                \"LacresUnidTransp\": [\r\n                                    {\r\n                                        \"nLacreUnidTransp\": \"999\"\r\n                                    }\r\n                                ],\r\n                                \"UnidadesCargaTran\": [\r\n                                    {\r\n                                        \"tipoUnidCargaTran\": 2,\r\n                                        \"idUnidCargaTran\": \"22\",\r\n                                        \"qtdRateadaCargaTran\": 10,\r\n                                        \"LacresUnidCargaTran\": [\r\n                                            {\r\n                                                \"nLacreUnidCargaTran\": \"100\"\r\n                                            }\r\n                                        ]\r\n                                    }\r\n                                ]\r\n                            }\r\n                        ],\r\n                        \"UnidadesCarga\": []\r\n                    }\r\n                ],\r\n                \"docAnt\": [],\r\n                \"infModal\": {\r\n                    \"versaoModal\": 3.00,\r\n                    \"rodo\": {\r\n                        \"RNTRC\": \"89173646\",\r\n                        \"occ\": []\r\n                    },\r\n                    \"aereo\": {\r\n                        \"nMinu\": 0,\r\n                        \"nOCA\": 0,\r\n                        \"dPrev_aer\": \"0000-00-00\",\r\n                        \"tarifa\": {\r\n                            \"CL\": \"\",\r\n                            \"cTar\": \"\",\r\n                            \"vTar\": \"0.00\"\r\n                        },\r\n                        \"NatCarga\": {\r\n                            \"xDime\": \"\",\r\n                            \"InfManu\": []\r\n                        },\r\n                        \"periAereo\": []\r\n                    },\r\n                    \"aquav\": {\r\n                        \"vPrestAFRMM\": \"0.00\",\r\n                        \"vAFRMM\": \"0.00\",\r\n                        \"nBooking\": \"\",\r\n                        \"nCtrl\": \"\",\r\n                        \"xNavio\": \"\",\r\n                        \"nViag\": \"\",\r\n                        \"direc\": \"\",\r\n                        \"prtEmb\": \"\",\r\n                        \"prtTrans\": \"\",\r\n                        \"prtDest\": \"\",\r\n                        \"tpNav\": \"\",\r\n                        \"irin\": \"\",\r\n                        \"detCont\": [],\r\n                        \"balsa\": []\r\n                    },\r\n                    \"ferrov\": {\r\n                        \"tpTraf\": 0,\r\n                        \"fluxo_fer\": \"\",\r\n                        \"idTrem\": \"\",\r\n                        \"vFrete\": \"0.00\",\r\n                        \"trafMut\": {\r\n                            \"respFat_fer\": 0,\r\n                            \"ferrEmi_fer\": 0,\r\n                            \"vFrete_fer\": \"0.00\",\r\n                            \"chCTeFerroOrigem\": \"\"\r\n                        },\r\n                        \"FerroEnv\": [],\r\n                        \"ferroSub\": {\r\n                            \"CNPJ_fer\": \"\",\r\n                            \"cInt_fer\": \"\",\r\n                            \"IE_fer\": \"\",\r\n                            \"xNome_fer\": \"\",\r\n                            \"xLgr_fer\": \"\",\r\n                            \"nro_fer\": \"\",\r\n                            \"xCpl_fer\": \"\",\r\n                            \"xBairro_fer\": \"\",\r\n                            \"cMun_fer\": 0,\r\n                            \"xMun_fer\": \"\",\r\n                            \"CEP_fer\": 0,\r\n                            \"UF_fer\": \"\"\r\n                        }\r\n                    },\r\n                    \"duto\": {\r\n                        \"vTar_duto\": \"0.000000\",\r\n                        \"dIni_duto\": \"0000-00-00\",\r\n                        \"dFim_duto\": \"0000-00-00\"\r\n                    },\r\n                    \"multimodal\": {\r\n                        \"COTM\": \"\",\r\n                        \"indNegociavel\": 0,\r\n                        \"segMulti\": {\r\n                            \"xSegMulti\": \"\",\r\n                            \"CNPJSegMulti\": \"\",\r\n                            \"nApolSegMulti\": \"\",\r\n                            \"nAverSegMulti\": \"\"\r\n                        }\r\n                    }\r\n                },\r\n                \"veicNovos\": [\r\n                    {\r\n                        \"chassi\": \"11111111111111111\",\r\n                        \"cCor\": \"1123\",\r\n                        \"xCor\": \"VERDE MUSGO\",\r\n                        \"cMod\": \"7878\",\r\n                        \"vUnit\": \"15000.00\",\r\n                        \"vFrete_vNov\": \"500.99\"\r\n                    }\r\n                ],\r\n                \"cobr\": {\r\n                    \"fat\": {\r\n                        \"nFat\": \"9789988123\",\r\n                        \"vOrig\": \"12200.23\",\r\n                        \"vDesc\": \"5542.12\",\r\n                        \"vLiq\": \"7981.22\"\r\n                    },\r\n                    \"dup\": [\r\n                        {\r\n                            \"nDup\": \"56789658\",\r\n                            \"dVenc\": \"2014-07-03\",\r\n                            \"vDup\": \"12321.33\"\r\n                        }\r\n                    ]\r\n                },\r\n                \"infCteSub\": {\r\n                    \"chCTe_sub\": \"\",\r\n                    \"refCteAnu\": \"\",\r\n                    \"indAlteraToma\": \"\",\r\n                    \"tomaICMS\": {\r\n                        \"refNFe_sub\": \"\",\r\n                        \"refCte_sub\": \"\",\r\n                        \"refNF_CT\": {\r\n                            \"CNPJ_sub\": \"\",\r\n                            \"CPF_sub\": \"\",\r\n                            \"mod_sub\": \"\",\r\n                            \"serie_sub\": \"\",\r\n                            \"subserie_sub\": \"\",\r\n                            \"nro_sub\": 0,\r\n                            \"valor_sub\": \"0.00\",\r\n                            \"dEmi_sub\": \"0000-00-00\"\r\n                        }\r\n                    }\r\n                },\r\n                \"infGlobalizado\": {\r\n                    \"xObsGlob\": \"\"\r\n                },\r\n                \"infServVinc\": []\r\n            },\r\n            \"infCteComp\": {\r\n                \"chave_comp\": \"\"\r\n            },\r\n            \"infCteAnu\": {\r\n                \"chCte_anu\": \"\",\r\n                \"dEmi_anu\": \"0000-00-00\"\r\n            },\r\n            \"AutXml\": [],\r\n            \"infRespTec\": {\r\n                \"CNPJ_respTec\": \"\",\r\n                \"xContato\": \"\",\r\n                \"email_respTec\": \"\",\r\n                \"fone_respTec\": \"\",\r\n                \"idCSRT\": \"\",\r\n                \"hashCSRT\": \"\"\r\n            }\r\n        }\r\n    }\r\n]";
-        public string CTeSefaz = "";
-        public string CTeDescarte = "";
-        public string CTeCancelamento = "";
-        public string CTeCartaCorrecao = "";
-        public string CTeInutilizacao = "";
-        public string CTeExport = "";
-        public string CTeExportV2 = "";
-        public string CTeSubstituicao = "";
-        public string CTeEvento = "";
-        public string CTeImportacao = "";
+        private string CTeEnvio = "[\r\n    {\r\n        \"Documento\": {\r\n            \"ModeloDocumento\": \"CTe\",\r\n            \"Versao\": 3.00,\r\n            \"ide\": {\r\n                \"cCT\": 11,\r\n                \"cUF\": 43,\r\n                \"natOp\": \"Aqui vai descrita a Natureza de operação do seu conhecimento\",\r\n                \"CFOP\": 6352,\r\n                \"mod\": \"57\",\r\n                \"serie\": \"305\",\r\n                \"nCT\": 55,\r\n                \"dhEmi\": \"2021-02-22T16:02:14\",\r\n                \"fusoHorario\": \"-03:00\",\r\n                \"tpImp\": 1,\r\n                \"tpEmis\": 1,\r\n                \"tpAmb\": 2,\r\n                \"tpCTe\": 0,\r\n                \"procEmi\": 0,\r\n                \"indGlobalizado\": 0,\r\n                \"refCTE\": \"\",\r\n                \"cMunEnv\": 4310405,\r\n                \"xMunEnv\": \"Independencia\",\r\n                \"UFEnv\": \"RS\",\r\n                \"modal\": \"01\",\r\n                \"tpServ\": 0,\r\n                \"cMunIni\": 4310405,\r\n                \"xMunIni\": \"Independencia\",\r\n                \"UFIni\": \"RS\",\r\n                \"cMunFim\": 3524006,\r\n                \"xMunFim\": \"Boa Vista do Burica\",\r\n                \"UFFim\": \"SP\",\r\n                \"retira\": 0,\r\n                \"xDetRetira\": \"Retirar o pacote no local de entrega indicado\",\r\n                \"indIEToma\": 9,\r\n                \"dhCont\": \"0000-00-00T00:00:00\",\r\n                \"xJust\": \"\",\r\n                \"EmailArquivos\": \"\",\r\n                \"tomador\": {\r\n                    \"toma\": \"4\",\r\n                    \"CNPJ_toma\": \"\",\r\n                    \"CPF_toma\": \"02702315011\",\r\n                    \"IE_toma\": \"\",\r\n                    \"xNome_toma\": \"Ricardo Pieper\",\r\n                    \"xFant_toma\": \"\",\r\n                    \"fone_toma\": \"5597207496\",\r\n                    \"email_toma\": \"\",\r\n                    \"enderTomador\": {\r\n                        \"xLgr_toma\": \"Rua Senador Salgado Filho\",\r\n                        \"nro_toma\": \"463\",\r\n                        \"xCpl_toma\": \"Complemento nao informado\",\r\n                        \"xBairro_toma\": \"Centro\",\r\n                        \"cMun_toma\": 4310405,\r\n                        \"xMun_toma\": \"Independencia\",\r\n                        \"CEP_toma\": \"98915000\",\r\n                        \"UF_toma\": \"RS\",\r\n                        \"cPais_toma\": 1046,\r\n                        \"xPais_toma\": \"Brasil\"\r\n                    }\r\n                }\r\n            },\r\n            \"compl\": {\r\n                \"xCaracAd\": \"Encomenda\",\r\n                \"xCaracSer\": \"Transporte\",\r\n                \"xEmi\": \"Jose Campos\",\r\n                \"origCalc\": \"\",\r\n                \"destCalc\": \"\",\r\n                \"xObs\": \"Observações\",\r\n                \"fluxo\": {\r\n                    \"xOrig\": \"Angra dos Reis\",\r\n                    \"xDest\": \"Maria da Silva de Andrade da Conceição Ferreira Pinto Júnior\",\r\n                    \"xRota\": \"129386\",\r\n                    \"passagem\": [\r\n                        {\r\n                            \"xPass\": \"LHR\"\r\n                        },\r\n                        {\r\n                            \"xPass\": \"CGH\"\r\n                        },\r\n                        {\r\n                            \"xPass\": \"POA\"\r\n                        }\r\n                    ]\r\n                },\r\n                \"Entrega\": {\r\n                    \"tpPer\": \"1\",\r\n                    \"dProg\": \"2021-03-03\",\r\n                    \"dIni\": \"2021-03-03\",\r\n                    \"dFim\": \"2021-03-03\",\r\n                    \"tpHor\": \"1\",\r\n                    \"hProg\": \"10:30:22\",\r\n                    \"hIni\": \"10:30:22\",\r\n                    \"hFim\": \"10:30:22\"\r\n                },\r\n                \"ObsCont\": [],\r\n                \"ObsFisco\": []\r\n            },\r\n            \"emit\": {\r\n                \"CNPJ_emit\": \"06354976000149\",\r\n                \"IE\": \"1470049241\",\r\n                \"IEST\": \"123456789\",\r\n                \"xNome\": \"CT-E EMITIDO EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"xFant\": \"EMPRESA 1\",\r\n\t\t\t\t\"CRT\":\"1\",\r\n                \"enderEmit\": {\r\n                    \"xLgr\": \"End\",\r\n                    \"nro\": \"1234\",\r\n                    \"xCpl\": \"teste\",\r\n                    \"xBairro\": \"Centro\",\r\n                    \"cMun\": 4310405,\r\n                    \"xMun\": \"Independencia\",\r\n                    \"CEP\": \"98910000\",\r\n                    \"UF\": \"RS\",\r\n                    \"fone\": \"35354560\"\r\n                }\r\n            },\r\n            \"rem\": {\r\n                \"CNPJ_rem\": \"06354976000149\",\r\n                \"CPF_rem\": \"\",\r\n                \"IE_rem\": \"1470049241\",\r\n                \"xNome_rem\": \"CT-E EMITIDO EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"xFant_rem\": \"Migrate Company\",\r\n                \"fone_rem\": \"5535355540\",\r\n                \"email_rem\": \"email@email.com.br\",\r\n                \"enderRem\": {\r\n                    \"xLgr_rem\": \"Rua Padre Cacique\",\r\n                    \"nro_rem\": \"556\",\r\n                    \"xCpl_rem\": \"Próximo as Lojas Denise\",\r\n                    \"xBairro_rem\": \"Centro\",\r\n                    \"cMun_rem\": 4310405,\r\n                    \"xMun_rem\": \"Independencia\",\r\n                    \"CEP_rem\": \"98910000\",\r\n                    \"UF_rem\": \"RS\",\r\n                    \"cPais_rem\": 1046,\r\n                    \"xPais_rem\": \"Brasil\"\r\n                }\r\n            },\r\n            \"exped\": {\r\n                \"CNPJ_exp\": \"23640261001810\",\r\n                \"CPF_exp\": \"\",\r\n                \"IE_exp\": \"421025353115\",\r\n                \"xNome_exp\": \"CT-E EMITIDO EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"fone_exp\": \"5535358668\",\r\n                \"email_exp\": \"email@email.com.br\",\r\n                \"enderExped\": {\r\n                    \"xLgr_exp\": \"Rua das Palmeiras\",\r\n                    \"nro_exp\": \"115\",\r\n                    \"xCpl_exp\": \"Próximo ao bar do Zé\",\r\n                    \"xBairro_exp\": \"Centro\",\r\n                    \"cMun_exp\": 3524006,\r\n                    \"xMun_exp\": \"Independencia\",\r\n                    \"CEP_exp\": \"98915000\",\r\n                    \"UF_exp\": \"SP\",\r\n                    \"cPais_exp\": 1046,\r\n                    \"xPais_exp\": \"Brasil\"\r\n                }\r\n            },\r\n            \"receb\": {\r\n                \"CNPJ_rec\": \"\",\r\n                \"CPF_rec\": \"01968897070\",\r\n                \"IE_rec\": \"\",\r\n                \"xNome_rec\": \"CT-E EMITIDO EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"fone_rec\": \"5597207556\",\r\n                \"email_rec\": \"email@email.com.br\",\r\n                \"enderReceb\": {\r\n                    \"xLgr_rec\": \"Rua Senador Salgado Filho\",\r\n                    \"nro_rec\": \"463\",\r\n                    \"xCpl_rec\": \"\",\r\n                    \"xBairro_rec\": \"Centro\",\r\n                    \"cMun_rec\": 3524006,\r\n                    \"xMun_rec\": \"Independência\",\r\n                    \"CEP_rec\": \"98915000\",\r\n                    \"UF_rec\": \"SP\",\r\n                    \"cPais_rec\": 1046,\r\n                    \"xPais_rec\": \"Brasil\"\r\n                }\r\n            },\r\n            \"dest\": {\r\n                \"CNPJ_dest\": \"92554179000147\",\r\n                \"CPF_dest\": \"\",\r\n                \"IE_dest\": \"0080093515\",\r\n                \"xNome_dest\": \"CT-E EMITIDO EM AMBIENTE DE HOMOLOGACAO - SEM VALOR FISCAL\",\r\n                \"fone_dest\": \"5599733466\",\r\n                \"ISUF_dest\": \"\",\r\n                \"email_dest\": \"\",\r\n                \"enderDest\": {\r\n                    \"xLgr_dest\": \"Rua Travessa\",\r\n                    \"nro_dest\": \"280\",\r\n                    \"xCpl_dest\": \"Próximo ao hospital\",\r\n                    \"xBairro_dest\": \"Rita\",\r\n                    \"cMun_dest\": 4301602,\r\n                    \"xMun_dest\": \"BAGE\",\r\n                    \"CEP_dest\": \"98910000\",\r\n                    \"UF_dest\": \"RS\",\r\n                    \"cPais_dest\": 1046,\r\n                    \"xPais_dest\": \"Brasil\"\r\n                }\r\n            },\r\n            \"vPrest\": {\r\n                \"vTPrest\": \"32321.23\",\r\n                \"vRec\": \"12312.23\",\r\n                \"Comp\": [\r\n                    {\r\n                        \"xNome_comp\": \"Componente 13\",\r\n                        \"vComp\": \"812.44\"\r\n                    },\r\n                    {\r\n                        \"xNome_comp\": \"Componente 20\",\r\n                        \"vComp\": \"1612.44\"\r\n                    }\r\n                ]\r\n            },\r\n            \"imp\": {\r\n                \"vTotImp\": \"59.90\",\r\n                \"vTotTrib\": \"0.00\",\r\n                \"infAdFisco\": \"Informacoes de interesse do fisco estarão escritas aqui\",\r\n                \"ICMS\": {\r\n                    \"CST\": \"00\",\r\n                    \"vBC\": \"12.00\",\r\n                    \"pICMS\": 13,\r\n                    \"vICMS\": \"1.56\",\r\n                    \"pRedBC\": 0,\r\n                    \"vBCSTRet\": \"0.00\",\r\n                    \"vICMSSTRet\": \"0.00\",\r\n                    \"pICMSSTRet\": 0,\r\n                    \"vCred\": \"0.00\",\r\n                    \"pRedBCOutraUF\": 0,\r\n                    \"vBCOutraUF\": \"0.00\",\r\n                    \"pICMSOutraUF\": 0,\r\n                    \"vICMSOutraUF\": \"0.00\",\r\n                    \"indSN\": 1\r\n                },\r\n                \"ICMSUFFim\": {\r\n                    \"vBCUFFim\": \"0.00\",\r\n                    \"pFCPUFFim\": 0,\r\n                    \"pICMSUFFim\": 0,\r\n                    \"pICMSInter\": 0,\r\n                    \"pICMSInterPart\": 0,\r\n                    \"vFCPUFFim\": \"0.00\",\r\n                    \"vICMSUFIni\": \"0.00\",\r\n                    \"vICMSUFFim\": \"0.00\"\r\n                }\r\n            },\r\n            \"infCTeNorm\": {\r\n                \"infCarga\": {\r\n                    \"vMerc\": \"599.65\",\r\n                    \"proPred\": \"Pepitas de ouro\",\r\n                    \"xOutCat\": \"Que valem mais do que dinheiro\",\r\n                    \"vCargaAverb\": \"200.00\",\r\n                    \"infQ\": [\r\n                        {\r\n                            \"cUnid\": \"01\",\r\n                            \"tpMed\": \"PESO BASE DE CÁLCULO\",\r\n                            \"qCarga\": \"899.0000\"\r\n                        }\r\n                    ]\r\n                },\r\n                \"infDoc\": [\r\n                    {\r\n                        \"tipoDocumento\": 2,\r\n                        \"nRoma_nf\": \"\",\r\n                        \"nPed_nf\": \"\",\r\n                        \"mod_nf\": \"\",\r\n                        \"serie_nf\": \"\",\r\n                        \"nDoc_nf\": \"\",\r\n                        \"dEmi_nf\": \"0000-00-00\",\r\n                        \"vBC_nf\": \"0.00\",\r\n                        \"vICMS_nf\": \"0.00\",\r\n                        \"vBCST_nf\": \"0.00\",\r\n                        \"vST_nf\": \"0.00\",\r\n                        \"vProd_nf\": \"0.00\",\r\n                        \"vNF_nf\": \"0.00\",\r\n                        \"CFOP_nf\": 0,\r\n                        \"peso_nf\": \"0.000\",\r\n                        \"PINSuframa_nf\": \"\",\r\n                        \"dPrev_nf\": \"0000-00-00\",\r\n                        \"chave_nfe\": \"43210106354976000149553050000003831000241294\",\r\n                        \"PINSuframa_nfe\": \"991233459\",\r\n                        \"dPrev_nfe\": \"2017-05-22\",\r\n                        \"tpDoc_outros\": \"\",\r\n                        \"desc_outros\": \"\",\r\n                        \"nDoc_outros\": \"\",\r\n                        \"dEmi_outros\": \"0000-00-00\",\r\n                        \"vDocFisc_outros\": \"0.00\",\r\n                        \"dPrev_outros\": \"0000-00-00\",\r\n                        \"locRet\": {\r\n                            \"CNPJ_locRet\": \"\",\r\n                            \"CPF_locRet\": \"\",\r\n                            \"xNome_locRet\": \"\",\r\n                            \"xLgr_locRet\": \"\",\r\n                            \"nro_locRet\": \"\",\r\n                            \"xCpl_locRet\": \"\",\r\n                            \"xBairro_locRet\": \"\",\r\n                            \"cMun_locRet\": \"\",\r\n                            \"xMun_locRet\": \"\",\r\n                            \"UF_locRet\": \"\"\r\n                        },\r\n                        \"UnidadesTransp\": [\r\n                            {\r\n                                \"tipoUnidTransporte\": 5,\r\n                                \"idUnidTransporte\": \"BCIASUDX988\",\r\n                                \"qtdRateada\": 999,\r\n                                \"LacresUnidTransp\": [\r\n                                    {\r\n                                        \"nLacreUnidTransp\": \"999\"\r\n                                    }\r\n                                ],\r\n                                \"UnidadesCargaTran\": [\r\n                                    {\r\n                                        \"tipoUnidCargaTran\": 2,\r\n                                        \"idUnidCargaTran\": \"22\",\r\n                                        \"qtdRateadaCargaTran\": 10,\r\n                                        \"LacresUnidCargaTran\": [\r\n                                            {\r\n                                                \"nLacreUnidCargaTran\": \"100\"\r\n                                            }\r\n                                        ]\r\n                                    }\r\n                                ]\r\n                            }\r\n                        ],\r\n                        \"UnidadesCarga\": []\r\n                    }\r\n                ],\r\n                \"docAnt\": [],\r\n                \"infModal\": {\r\n                    \"versaoModal\": 3.00,\r\n                    \"rodo\": {\r\n                        \"RNTRC\": \"89173646\",\r\n                        \"occ\": []\r\n                    },\r\n                    \"aereo\": {\r\n                        \"nMinu\": 0,\r\n                        \"nOCA\": 0,\r\n                        \"dPrev_aer\": \"0000-00-00\",\r\n                        \"tarifa\": {\r\n                            \"CL\": \"\",\r\n                            \"cTar\": \"\",\r\n                            \"vTar\": \"0.00\"\r\n                        },\r\n                        \"NatCarga\": {\r\n                            \"xDime\": \"\",\r\n                            \"InfManu\": []\r\n                        },\r\n                        \"periAereo\": []\r\n                    },\r\n                    \"aquav\": {\r\n                        \"vPrestAFRMM\": \"0.00\",\r\n                        \"vAFRMM\": \"0.00\",\r\n                        \"nBooking\": \"\",\r\n                        \"nCtrl\": \"\",\r\n                        \"xNavio\": \"\",\r\n                        \"nViag\": \"\",\r\n                        \"direc\": \"\",\r\n                        \"prtEmb\": \"\",\r\n                        \"prtTrans\": \"\",\r\n                        \"prtDest\": \"\",\r\n                        \"tpNav\": \"\",\r\n                        \"irin\": \"\",\r\n                        \"detCont\": [],\r\n                        \"balsa\": []\r\n                    },\r\n                    \"ferrov\": {\r\n                        \"tpTraf\": 0,\r\n                        \"fluxo_fer\": \"\",\r\n                        \"idTrem\": \"\",\r\n                        \"vFrete\": \"0.00\",\r\n                        \"trafMut\": {\r\n                            \"respFat_fer\": 0,\r\n                            \"ferrEmi_fer\": 0,\r\n                            \"vFrete_fer\": \"0.00\",\r\n                            \"chCTeFerroOrigem\": \"\"\r\n                        },\r\n                        \"FerroEnv\": [],\r\n                        \"ferroSub\": {\r\n                            \"CNPJ_fer\": \"\",\r\n                            \"cInt_fer\": \"\",\r\n                            \"IE_fer\": \"\",\r\n                            \"xNome_fer\": \"\",\r\n                            \"xLgr_fer\": \"\",\r\n                            \"nro_fer\": \"\",\r\n                            \"xCpl_fer\": \"\",\r\n                            \"xBairro_fer\": \"\",\r\n                            \"cMun_fer\": 0,\r\n                            \"xMun_fer\": \"\",\r\n                            \"CEP_fer\": 0,\r\n                            \"UF_fer\": \"\"\r\n                        }\r\n                    },\r\n                    \"duto\": {\r\n                        \"vTar_duto\": \"0.000000\",\r\n                        \"dIni_duto\": \"0000-00-00\",\r\n                        \"dFim_duto\": \"0000-00-00\"\r\n                    },\r\n                    \"multimodal\": {\r\n                        \"COTM\": \"\",\r\n                        \"indNegociavel\": 0,\r\n                        \"segMulti\": {\r\n                            \"xSegMulti\": \"\",\r\n                            \"CNPJSegMulti\": \"\",\r\n                            \"nApolSegMulti\": \"\",\r\n                            \"nAverSegMulti\": \"\"\r\n                        }\r\n                    }\r\n                },\r\n                \"veicNovos\": [\r\n                    {\r\n                        \"chassi\": \"11111111111111111\",\r\n                        \"cCor\": \"1123\",\r\n                        \"xCor\": \"VERDE MUSGO\",\r\n                        \"cMod\": \"7878\",\r\n                        \"vUnit\": \"15000.00\",\r\n                        \"vFrete_vNov\": \"500.99\"\r\n                    }\r\n                ],\r\n                \"cobr\": {\r\n                    \"fat\": {\r\n                        \"nFat\": \"9789988123\",\r\n                        \"vOrig\": \"12200.23\",\r\n                        \"vDesc\": \"5542.12\",\r\n                        \"vLiq\": \"7981.22\"\r\n                    },\r\n                    \"dup\": [\r\n                        {\r\n                            \"nDup\": \"56789658\",\r\n                            \"dVenc\": \"2014-07-03\",\r\n                            \"vDup\": \"12321.33\"\r\n                        }\r\n                    ]\r\n                },\r\n                \"infCteSub\": {\r\n                    \"chCTe_sub\": \"\",\r\n                    \"refCteAnu\": \"\",\r\n                    \"indAlteraToma\": \"\",\r\n                    \"tomaICMS\": {\r\n                        \"refNFe_sub\": \"\",\r\n                        \"refCte_sub\": \"\",\r\n                        \"refNF_CT\": {\r\n                            \"CNPJ_sub\": \"\",\r\n                            \"CPF_sub\": \"\",\r\n                            \"mod_sub\": \"\",\r\n                            \"serie_sub\": \"\",\r\n                            \"subserie_sub\": \"\",\r\n                            \"nro_sub\": 0,\r\n                            \"valor_sub\": \"0.00\",\r\n                            \"dEmi_sub\": \"0000-00-00\"\r\n                        }\r\n                    }\r\n                },\r\n                \"infGlobalizado\": {\r\n                    \"xObsGlob\": \"\"\r\n                },\r\n                \"infServVinc\": []\r\n            },\r\n            \"infCteComp\": {\r\n                \"chave_comp\": \"\"\r\n            },\r\n            \"infCteAnu\": {\r\n                \"chCte_anu\": \"\",\r\n                \"dEmi_anu\": \"0000-00-00\"\r\n            },\r\n            \"AutXml\": [],\r\n            \"infRespTec\": {\r\n                \"CNPJ_respTec\": \"\",\r\n                \"xContato\": \"\",\r\n                \"email_respTec\": \"\",\r\n                \"fone_respTec\": \"\",\r\n                \"idCSRT\": \"\",\r\n                \"hashCSRT\": \"\"\r\n            }\r\n        }\r\n    }\r\n]";
+        private string CTeSefaz = "";
+        private string CTeDescarte = "";
+        private string CTeCancelamento = "";
+        private string CTeCartaCorrecao = "";
+        private string CTeInutilizacao = "";
+        private string CTeExport = "";
+        private string CTeExportV2 = "";
+        private string CTeSubstituicao = "";
+        private string CTeEvento = "";
+        private string CTeImportacao = "";
 
-        public string EmpresaCadastro = "[\r\n    {\r\n        \"EmpNomFantasia\": \"Cadastro Empresa REST\",\r\n        \"EmpApelido\": \"Cadastro Empresa REST\",\r\n        \"EmpRazSocial\": \"Cadastro Empresa REST\",\r\n        \"EmpCNPJ\": \"36762131000153\",\r\n        \"EmpCPF\": \"\",\r\n        \"EmpIE\": \"9353071823\",\r\n        \"EmpTelefone\": \"5535354830\",\r\n        \"EmpEndereco\": \"RUA PE CACIQUE\",\r\n        \"EmpNumero\": \"985\",\r\n        \"EmpBairro\": \"CENTRO\",\r\n        \"EmpCEP\": \"98910000\",\r\n        \"EmpComplemento\": \"2 ANDAR\",\r\n        \"EmpIM\": \"136235781\",\r\n        \"MunCodigo\": 4321808,\r\n        \"EmpCNAE\": \"\",\r\n        \"EmpCRT\": 1,\r\n        \"EmpIEST\": \"\",\r\n        \"EmpMtrCNPJ\": \"\",\r\n        \"EmpMarca\": \"\",\r\n        \"EmpMarcaExtensao\": \"\",\r\n\t\t\"EmpArqApelido\":\"\",\r\n        \"EmpEmail\": \"\",\r\n        \"EmpObservacao\": \"\",\r\n        \"EmpTpoEndereco\": \"RUA\",\r\n        \"Certificado\": {\r\n            \"ArquivoPFX\": \"\",\r\n            \"Senha\": \"\",\r\n            \"TipoCertificado\": 0,\r\n            \"MetodoCriptografiaSenha\": \"\",\r\n            \"UsarCertPreCadastrado\": \"\"\r\n        },\r\n        \"Usuarios\": [],\r\n        \"CaixaEnvioEmail\": {\r\n            \"EmlValCertificado\": \"\",\r\n            \"EmlTmpEspera\": 0,\r\n            \"EmlAutenticacao\": \"\",\r\n            \"EmlSSL\": \"\",\r\n            \"EmlPorta\": 0,\r\n            \"EmlHost\": \"\",\r\n            \"EmlSenha\": \"\",\r\n            \"EmlUsuario\": \"\",\r\n            \"EmlNomeRemetente\": \"\",\r\n            \"EmlEmailRemetente\": \"\",\r\n            \"EmlStatus\": \"\"\r\n        },\r\n        \"CaixasLeituraEmail\": [],\r\n        \"Parametros\": {\r\n            \"NFe\": {\r\n                \"FormaRetornoPDFIntegracao\": 3,\r\n                \"FormaRetornoXMLIntegracao\": 3,\r\n                \"UltimoNSU\": \"000000000029764\",\r\n                \"JustInut\": \"\",\r\n                \"JustCanc\": \"\",\r\n                \"JustCont\": \"\",\r\n                \"TextoEmailAutorizado\": \"\",\r\n                \"TextoEmailCancelado\": \"\",\r\n                \"TextoEmailCCe\": \"\",\r\n                \"EmailTitulo\": \"\",\r\n                \"EnviarPDFEmail\": \"\",\r\n                \"EnviarXMLEmail\": \"\",\r\n                \"OrdemContingencia\": [\r\n                    {\r\n                        \"OrdemContingenciaNFe\": 4\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFe\": 1\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFe\": 2\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFe\": 0\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFe\": 3\r\n                    }\r\n                ]\r\n            },\r\n            \"NFSe\": {\r\n                \"CodCMC\": \"444\",\r\n                \"AmbienteEmissao\": 2,\r\n                \"MaxRPSLote\": 1,\r\n                \"TipoProcessamento\": \"S\",\r\n                \"UsuarioAutent\": \"usuario\",\r\n                \"SenhaAutent\": \"senha\",\r\n                \"ImprimeNFSeEfetivada\": \"S\",\r\n                \"EnviarNFSeTomador\": \"N\",\r\n                \"EmpRetImpPrefeitura\": \"N\",\r\n                \"EmpConcatOutrasInfo\": \"S\",\r\n                \"EmpImprimeCanhoto\": 2,\r\n                \"EmpChavePrimaria\": \"\",\r\n                \"FormaRetornoPDFIntegracao\": 1,\r\n                \"FormaRetornoXmlIntegracao\": 1,\r\n                \"EmpDiscriminacaoBetha\": \"N\",\r\n                \"EmpAutorizacaoRegimeRPS\": \"1234\",\r\n                \"EmpEspelhoNovo\": \"S\",\r\n                \"EmpEnviaCPFPrestador\": \"N\",\r\n                \"VerCodigo\": 0,\r\n                \"DataAdesaoSimplesNacional\": \"0000-00-00\",\r\n                \"EmpEnviaEmailCompactado\": \"N\",\r\n                \"ImprimeItensEspelho\": \"N\",\r\n                \"EmpClientId\": \"\",\r\n                \"EmpClientSecret\": \"\",\r\n                \"EmpImgRodapeBase64\": \"\",\r\n                \"FaixaRPSAutorizadoInicial\": 0,\r\n                \"FaixaRPSAutorizadoFinal\": 0,\r\n                \"ImprimirOutrasInformacoes\": \"\",\r\n                \"ImprimirRetencoesFederais\": \"\",\r\n                \"XMLRetornoRequisicao\": \"\",\r\n                \"NomeCanhoto\": \"\",\r\n                \"ReceberEmailEmpresaTrancada\": \"\",\r\n\t\t\t\t\"EmpRegEspTrib\": \"21\"\r\n            },\r\n            \"MDFe\": {\r\n                \"FormaRetornoPDFIntegracao\": 1,\r\n                \"FormaRetornoXMLIntegracao\": 2,\r\n                \"UltimoNSU\": \"000000000000754\",\r\n                \"JustCanc\": \"\",\r\n                \"JustCont\": \"\"\r\n            },\r\n            \"NFCe\": {\r\n                \"FormaRetornoPDFIntegracao\": 3,\r\n                \"FormaRetornoXMLIntegracao\": 3,\r\n                \"UltimoNSU\": \"000000000000000\",\r\n                \"IDTokenCscSEFAZ\": \"00001\",\r\n                \"CscSEFAZ\": \"1A451E99-0FE0-4C13-B97E-67D698FFBC37\",\r\n                \"PossuiLeituraX\": \"N\",\r\n                \"PossuiGeracaoRelatorios\": \"\",\r\n                \"JustInut\": \"\",\r\n                \"JustCanc\": \"\",\r\n                \"JustCont\": \"\",\r\n                \"TextoEmailAutorizado\": \"\",\r\n                \"TextoEmailCancelado\": \"\",\r\n                \"EmailTitulo\": \"\",\r\n                \"EnviarPDFEmail\": \"\",\r\n                \"EnviarXMLEmail\": \"\",\r\n                \"OrdemContingencia\": [\r\n                    {\r\n                        \"OrdemContingenciaNFCe\": 3\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFCe\": 0\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFCe\": 2\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFCe\": 1\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFCe\": 4\r\n                    }\r\n                ]\r\n            },\r\n            \"CTe\": {\r\n                \"FormaRetornoPDFIntegracao\": 3,\r\n                \"FormaRetornoXMLIntegracao\": 3,\r\n                \"UltimoNSU\": \"000000000039547\",\r\n                \"JustInut\": \"\",\r\n                \"JustCanc\": \"\",\r\n                \"JustCont\": \"\",\r\n                \"TextoEmailAutorizado\": \"\",\r\n                \"TextoEmailCancelado\": \"\",\r\n                \"TextoEmailCCe\": \"\",\r\n                \"EmailTitulo\": \"\",\r\n                \"EnviarPDFEmail\": \"\",\r\n                \"EnviarXMLEmail\": \"\",\r\n                \"OrdemContingencia\": [\r\n                    {\r\n                        \"OrdemContingenciaCTe\": 1\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaCTe\": 2\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaCTe\": 0\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaCTe\": 3\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaCTe\": 4\r\n                    }\r\n                ]\r\n            }\r\n        },\r\n        \"Licenciamento\": [\r\n            {\r\n                \"Modulo\": \"NF-e\",\r\n                \"Modelo\": \"1\",\r\n                \"Autor\": \"REST\"\r\n            },\r\n            {\r\n                \"Modulo\": \"NFC-e\",\r\n                \"Modelo\": \"1\",\r\n                \"Autor\": \"REST\"\r\n            },\r\n            {\r\n                \"Modulo\": \"MDF-e\",\r\n                \"Modelo\": \"1\",\r\n                \"Autor\": \"REST\"\r\n            },\r\n            {\r\n                \"Modulo\": \"NFS-e\",\r\n                \"Modelo\": \"1\",\r\n                \"Autor\": \"REST\"\r\n            },\r\n            {\r\n                \"Modulo\": \"CT-e\",\r\n                \"Modelo\": \"1\",\r\n                \"Autor\": \"REST\"\r\n            },\r\n            {\r\n                \"Modulo\": \"Digitacao NFe\",\r\n                \"Modelo\": \"1\",\r\n                \"Autor\": \"REST\"\r\n            }\r\n        ],\r\n        \"Impressao\": {\r\n            \"NFe\": {\r\n                \"OrientacaoNFe\": 1,\r\n                \"TamanhoFonte\": \"2\",\r\n                \"NumeroCopias\": 0,\r\n                \"ImprimirDataHoraImpressao\": \"S\",\r\n                \"MarcarComoReimpresso\": \"N\",\r\n                \"ImprimirTributos\": \"3\",\r\n                \"InformacaoComplementarNoVerso\": \"N\",\r\n                \"ExibirDescontoEIcmsNoItem\": \"N\",\r\n                \"LocalImpressaoCanhoto\": \"C\",\r\n                \"MolduraNaLogomarca\": \"S\",\r\n                \"ModeloDANFE\": 2,\r\n                \"ExibirEndEntregaInfoAdic\": \"N\",\r\n                \"ImprimirDuplicatas\": \"S\",\r\n                \"UtilizarImagemInfoEmitente\": \"N\",\r\n                \"ImprimirVolumesItensDestacados\": \"N\",\r\n                \"ImprimirTotalPISCOFINS\": \"N\",\r\n                \"InfoComplementar\": \"\",\r\n                \"TextoImpressaoTributos\": \"\",\r\n                \"DestaqueProdutosPerigosos\": \"\",\r\n                \"TamanhoFonteInfComplementares\": 0,\r\n                \"ImprimeUNTribDiferenteComercial\": \"\",\r\n                \"ImprimirEnderecoEntregaRetirada\": 0,\r\n                \"ImprimirCamposDIFAL\": \"\",\r\n                \"ImprimirICMSCST51\": \"\"\r\n            },\r\n            \"MDFe\": {\r\n                \"OrientacaoMDFe\": 1,\r\n                \"MolduraNaLogomarca\": \"N\"\r\n            },\r\n            \"CTe\": {\r\n                \"OrientacaoCTe\": 1,\r\n                \"MolduraNaLogomarca\": \"S\",\r\n                \"LocalImpressaoCanhoto\": \"C\",\r\n                \"TextoImpressaoTributos\": \"\"\r\n            },\r\n            \"NFCe\": {\r\n                \"ModeloDANFE\": \"1\",\r\n                \"ImpAcrescDescItens\": \"S\",\r\n                \"FormatoDANFE\": \"D\"\r\n            }\r\n        },\r\n        \"Extensao\": [\r\n            {\r\n                \"TipoExtensao\": 4,\r\n                \"Modulos\": [\r\n                    {\r\n                        \"Modulo\": \"NFe\"\r\n                    },\r\n                    {\r\n                        \"Modulo\": \"CTe\"\r\n                    },\r\n                    {\r\n                        \"Modulo\": \"MDFe\"\r\n                    }\r\n                ],\r\n                \"ParametrosExt\": {\r\n                    \"NumeroConsultasDiarias\": 12,\r\n                    \"ArmazenarEveTerceiros\": \"S\",\r\n                    \"BaixarDocsNaoArmazenados\": \"S\",\r\n                    \"CienciaAutomatica\": 0,\r\n                    \"Justificativa\": \"Operação não realizada.\",\r\n                    \"ValorMin\": \"0.00\",\r\n                    \"Email\": \"\",\r\n                    \"TipoNotificacao\": 0,\r\n                    \"BuscarDocsRetroativos\": \"\",\r\n                    \"MunicipiosAtivar\": \"\",\r\n                    \"MunicipiosDesativar\": \"\"\r\n                },\r\n                \"Ativar\": \"S\"\r\n            },\r\n            {\r\n                \"TipoExtensao\": 3,\r\n                \"Modulos\": [\r\n                    {\r\n                        \"Modulo\": \"NFe\"\r\n                    },\r\n                    {\r\n                        \"Modulo\": \"CTe\"\r\n                    },\r\n                    {\r\n                        \"Modulo\": \"MDFe\"\r\n                    },\r\n                    {\r\n                        \"Modulo\": \"NFCe\"\r\n                    },\r\n                    {\r\n                        \"Modulo\": \"NFSe\"\r\n                    }\r\n                ],\r\n                \"ParametrosExt\": {\r\n                    \"NumeroConsultasDiarias\": 0,\r\n                    \"ArmazenarEveTerceiros\": \"\",\r\n                    \"BaixarDocsNaoArmazenados\": \"\",\r\n                    \"CienciaAutomatica\": 0,\r\n                    \"Justificativa\": \"\",\r\n                    \"ValorMin\": \"0.00\",\r\n                    \"Email\": \"\",\r\n                    \"TipoNotificacao\": 0,\r\n                    \"BuscarDocsRetroativos\": \"\",\r\n                    \"MunicipiosAtivar\": \"\",\r\n                    \"MunicipiosDesativar\": \"\"\r\n                },\r\n                \"Ativar\": \"S\"\r\n            }\r\n        ]\r\n    }\r\n]";
-        public string EmpresaEditar = "[{\n  \"EmpNomFantasia\": \"Empresa de teste de cadastro 2\",\n  \"EmpApelido\": \"Empresa de teste de cadastro 2\",\n  \"EmpRazSocial\": \"Empresa de teste de cadastro 2\",\n  \"EmpCNPJ\": \"29487402000144\",\n  \"EmpIE\": \"342423324\",\n  \"EmpTelefone\": \"234234\",\n  \"EmpEndereco\": \"Rua legal\",\n  \"EmpNumero\": \"1\",\n  \"EmpBairro\": \"Centro\",\n  \"EmpCEP\": \"98918000\",\n  \"EmpComplemento\": \"Perto de outra rua\",\n  \"EmpIM\": \"664544654\",\n  \"MunCodigo\": 4321808,\n  \"EmpCNAE\": \"\",\n  \"EmpCRT\": 0,\n  \"EmpIEST\": \"\",\n  \"EmpEmail\": \"\",\n  \"EmpObservacao\": \"\",\n  \"EmpTpoEndereco\": \"\"\n}]";
-        public string EmpresaLicenciamento = "[\n    {\n        \"CnpjEmpresa\": \"70049563000150\",\n        \"tpAmb\": 2,\n        \"Acao\": 3,\n        \"Modulo\": \"CTe\",\n        \"Modelo\": 2,\n        \"Autor\": \"Empresa 1\"\n    }\n]";
+        private string EmpresaCadastro = "[\r\n    {\r\n        \"EmpNomFantasia\": \"Cadastro Empresa REST\",\r\n        \"EmpApelido\": \"Cadastro Empresa REST\",\r\n        \"EmpRazSocial\": \"Cadastro Empresa REST\",\r\n        \"EmpCNPJ\": \"36762131000153\",\r\n        \"EmpCPF\": \"\",\r\n        \"EmpIE\": \"9353071823\",\r\n        \"EmpTelefone\": \"5535354830\",\r\n        \"EmpEndereco\": \"RUA PE CACIQUE\",\r\n        \"EmpNumero\": \"985\",\r\n        \"EmpBairro\": \"CENTRO\",\r\n        \"EmpCEP\": \"98910000\",\r\n        \"EmpComplemento\": \"2 ANDAR\",\r\n        \"EmpIM\": \"136235781\",\r\n        \"MunCodigo\": 4321808,\r\n        \"EmpCNAE\": \"\",\r\n        \"EmpCRT\": 1,\r\n        \"EmpIEST\": \"\",\r\n        \"EmpMtrCNPJ\": \"\",\r\n        \"EmpMarca\": \"\",\r\n        \"EmpMarcaExtensao\": \"\",\r\n\t\t\"EmpArqApelido\":\"\",\r\n        \"EmpEmail\": \"\",\r\n        \"EmpObservacao\": \"\",\r\n        \"EmpTpoEndereco\": \"RUA\",\r\n        \"Certificado\": {\r\n            \"ArquivoPFX\": \"\",\r\n            \"Senha\": \"\",\r\n            \"TipoCertificado\": 0,\r\n            \"MetodoCriptografiaSenha\": \"\",\r\n            \"UsarCertPreCadastrado\": \"\"\r\n        },\r\n        \"Usuarios\": [],\r\n        \"CaixaEnvioEmail\": {\r\n            \"EmlValCertificado\": \"\",\r\n            \"EmlTmpEspera\": 0,\r\n            \"EmlAutenticacao\": \"\",\r\n            \"EmlSSL\": \"\",\r\n            \"EmlPorta\": 0,\r\n            \"EmlHost\": \"\",\r\n            \"EmlSenha\": \"\",\r\n            \"EmlUsuario\": \"\",\r\n            \"EmlNomeRemetente\": \"\",\r\n            \"EmlEmailRemetente\": \"\",\r\n            \"EmlStatus\": \"\"\r\n        },\r\n        \"CaixasLeituraEmail\": [],\r\n        \"Parametros\": {\r\n            \"NFe\": {\r\n                \"FormaRetornoPDFIntegracao\": 3,\r\n                \"FormaRetornoXMLIntegracao\": 3,\r\n                \"UltimoNSU\": \"000000000029764\",\r\n                \"JustInut\": \"\",\r\n                \"JustCanc\": \"\",\r\n                \"JustCont\": \"\",\r\n                \"TextoEmailAutorizado\": \"\",\r\n                \"TextoEmailCancelado\": \"\",\r\n                \"TextoEmailCCe\": \"\",\r\n                \"EmailTitulo\": \"\",\r\n                \"EnviarPDFEmail\": \"\",\r\n                \"EnviarXMLEmail\": \"\",\r\n                \"OrdemContingencia\": [\r\n                    {\r\n                        \"OrdemContingenciaNFe\": 4\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFe\": 1\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFe\": 2\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFe\": 0\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFe\": 3\r\n                    }\r\n                ]\r\n            },\r\n            \"NFSe\": {\r\n                \"CodCMC\": \"444\",\r\n                \"AmbienteEmissao\": 2,\r\n                \"MaxRPSLote\": 1,\r\n                \"TipoProcessamento\": \"S\",\r\n                \"UsuarioAutent\": \"usuario\",\r\n                \"SenhaAutent\": \"senha\",\r\n                \"ImprimeNFSeEfetivada\": \"S\",\r\n                \"EnviarNFSeTomador\": \"N\",\r\n                \"EmpRetImpPrefeitura\": \"N\",\r\n                \"EmpConcatOutrasInfo\": \"S\",\r\n                \"EmpImprimeCanhoto\": 2,\r\n                \"EmpChavePrimaria\": \"\",\r\n                \"FormaRetornoPDFIntegracao\": 1,\r\n                \"FormaRetornoXmlIntegracao\": 1,\r\n                \"EmpDiscriminacaoBetha\": \"N\",\r\n                \"EmpAutorizacaoRegimeRPS\": \"1234\",\r\n                \"EmpEspelhoNovo\": \"S\",\r\n                \"EmpEnviaCPFPrestador\": \"N\",\r\n                \"VerCodigo\": 0,\r\n                \"DataAdesaoSimplesNacional\": \"0000-00-00\",\r\n                \"EmpEnviaEmailCompactado\": \"N\",\r\n                \"ImprimeItensEspelho\": \"N\",\r\n                \"EmpClientId\": \"\",\r\n                \"EmpClientSecret\": \"\",\r\n                \"EmpImgRodapeBase64\": \"\",\r\n                \"FaixaRPSAutorizadoInicial\": 0,\r\n                \"FaixaRPSAutorizadoFinal\": 0,\r\n                \"ImprimirOutrasInformacoes\": \"\",\r\n                \"ImprimirRetencoesFederais\": \"\",\r\n                \"XMLRetornoRequisicao\": \"\",\r\n                \"NomeCanhoto\": \"\",\r\n                \"ReceberEmailEmpresaTrancada\": \"\",\r\n\t\t\t\t\"EmpRegEspTrib\": \"21\"\r\n            },\r\n            \"MDFe\": {\r\n                \"FormaRetornoPDFIntegracao\": 1,\r\n                \"FormaRetornoXMLIntegracao\": 2,\r\n                \"UltimoNSU\": \"000000000000754\",\r\n                \"JustCanc\": \"\",\r\n                \"JustCont\": \"\"\r\n            },\r\n            \"NFCe\": {\r\n                \"FormaRetornoPDFIntegracao\": 3,\r\n                \"FormaRetornoXMLIntegracao\": 3,\r\n                \"UltimoNSU\": \"000000000000000\",\r\n                \"IDTokenCscSEFAZ\": \"00001\",\r\n                \"CscSEFAZ\": \"1A451E99-0FE0-4C13-B97E-67D698FFBC37\",\r\n                \"PossuiLeituraX\": \"N\",\r\n                \"PossuiGeracaoRelatorios\": \"\",\r\n                \"JustInut\": \"\",\r\n                \"JustCanc\": \"\",\r\n                \"JustCont\": \"\",\r\n                \"TextoEmailAutorizado\": \"\",\r\n                \"TextoEmailCancelado\": \"\",\r\n                \"EmailTitulo\": \"\",\r\n                \"EnviarPDFEmail\": \"\",\r\n                \"EnviarXMLEmail\": \"\",\r\n                \"OrdemContingencia\": [\r\n                    {\r\n                        \"OrdemContingenciaNFCe\": 3\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFCe\": 0\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFCe\": 2\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFCe\": 1\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaNFCe\": 4\r\n                    }\r\n                ]\r\n            },\r\n            \"CTe\": {\r\n                \"FormaRetornoPDFIntegracao\": 3,\r\n                \"FormaRetornoXMLIntegracao\": 3,\r\n                \"UltimoNSU\": \"000000000039547\",\r\n                \"JustInut\": \"\",\r\n                \"JustCanc\": \"\",\r\n                \"JustCont\": \"\",\r\n                \"TextoEmailAutorizado\": \"\",\r\n                \"TextoEmailCancelado\": \"\",\r\n                \"TextoEmailCCe\": \"\",\r\n                \"EmailTitulo\": \"\",\r\n                \"EnviarPDFEmail\": \"\",\r\n                \"EnviarXMLEmail\": \"\",\r\n                \"OrdemContingencia\": [\r\n                    {\r\n                        \"OrdemContingenciaCTe\": 1\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaCTe\": 2\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaCTe\": 0\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaCTe\": 3\r\n                    },\r\n                    {\r\n                        \"OrdemContingenciaCTe\": 4\r\n                    }\r\n                ]\r\n            }\r\n        },\r\n        \"Licenciamento\": [\r\n            {\r\n                \"Modulo\": \"NF-e\",\r\n                \"Modelo\": \"1\",\r\n                \"Autor\": \"REST\"\r\n            },\r\n            {\r\n                \"Modulo\": \"NFC-e\",\r\n                \"Modelo\": \"1\",\r\n                \"Autor\": \"REST\"\r\n            },\r\n            {\r\n                \"Modulo\": \"MDF-e\",\r\n                \"Modelo\": \"1\",\r\n                \"Autor\": \"REST\"\r\n            },\r\n            {\r\n                \"Modulo\": \"NFS-e\",\r\n                \"Modelo\": \"1\",\r\n                \"Autor\": \"REST\"\r\n            },\r\n            {\r\n                \"Modulo\": \"CT-e\",\r\n                \"Modelo\": \"1\",\r\n                \"Autor\": \"REST\"\r\n            },\r\n            {\r\n                \"Modulo\": \"Digitacao NFe\",\r\n                \"Modelo\": \"1\",\r\n                \"Autor\": \"REST\"\r\n            }\r\n        ],\r\n        \"Impressao\": {\r\n            \"NFe\": {\r\n                \"OrientacaoNFe\": 1,\r\n                \"TamanhoFonte\": \"2\",\r\n                \"NumeroCopias\": 0,\r\n                \"ImprimirDataHoraImpressao\": \"S\",\r\n                \"MarcarComoReimpresso\": \"N\",\r\n                \"ImprimirTributos\": \"3\",\r\n                \"InformacaoComplementarNoVerso\": \"N\",\r\n                \"ExibirDescontoEIcmsNoItem\": \"N\",\r\n                \"LocalImpressaoCanhoto\": \"C\",\r\n                \"MolduraNaLogomarca\": \"S\",\r\n                \"ModeloDANFE\": 2,\r\n                \"ExibirEndEntregaInfoAdic\": \"N\",\r\n                \"ImprimirDuplicatas\": \"S\",\r\n                \"UtilizarImagemInfoEmitente\": \"N\",\r\n                \"ImprimirVolumesItensDestacados\": \"N\",\r\n                \"ImprimirTotalPISCOFINS\": \"N\",\r\n                \"InfoComplementar\": \"\",\r\n                \"TextoImpressaoTributos\": \"\",\r\n                \"DestaqueProdutosPerigosos\": \"\",\r\n                \"TamanhoFonteInfComplementares\": 0,\r\n                \"ImprimeUNTribDiferenteComercial\": \"\",\r\n                \"ImprimirEnderecoEntregaRetirada\": 0,\r\n                \"ImprimirCamposDIFAL\": \"\",\r\n                \"ImprimirICMSCST51\": \"\"\r\n            },\r\n            \"MDFe\": {\r\n                \"OrientacaoMDFe\": 1,\r\n                \"MolduraNaLogomarca\": \"N\"\r\n            },\r\n            \"CTe\": {\r\n                \"OrientacaoCTe\": 1,\r\n                \"MolduraNaLogomarca\": \"S\",\r\n                \"LocalImpressaoCanhoto\": \"C\",\r\n                \"TextoImpressaoTributos\": \"\"\r\n            },\r\n            \"NFCe\": {\r\n                \"ModeloDANFE\": \"1\",\r\n                \"ImpAcrescDescItens\": \"S\",\r\n                \"FormatoDANFE\": \"D\"\r\n            }\r\n        },\r\n        \"Extensao\": [\r\n            {\r\n                \"TipoExtensao\": 4,\r\n                \"Modulos\": [\r\n                    {\r\n                        \"Modulo\": \"NFe\"\r\n                    },\r\n                    {\r\n                        \"Modulo\": \"CTe\"\r\n                    },\r\n                    {\r\n                        \"Modulo\": \"MDFe\"\r\n                    }\r\n                ],\r\n                \"ParametrosExt\": {\r\n                    \"NumeroConsultasDiarias\": 12,\r\n                    \"ArmazenarEveTerceiros\": \"S\",\r\n                    \"BaixarDocsNaoArmazenados\": \"S\",\r\n                    \"CienciaAutomatica\": 0,\r\n                    \"Justificativa\": \"Operação não realizada.\",\r\n                    \"ValorMin\": \"0.00\",\r\n                    \"Email\": \"\",\r\n                    \"TipoNotificacao\": 0,\r\n                    \"BuscarDocsRetroativos\": \"\",\r\n                    \"MunicipiosAtivar\": \"\",\r\n                    \"MunicipiosDesativar\": \"\"\r\n                },\r\n                \"Ativar\": \"S\"\r\n            },\r\n            {\r\n                \"TipoExtensao\": 3,\r\n                \"Modulos\": [\r\n                    {\r\n                        \"Modulo\": \"NFe\"\r\n                    },\r\n                    {\r\n                        \"Modulo\": \"CTe\"\r\n                    },\r\n                    {\r\n                        \"Modulo\": \"MDFe\"\r\n                    },\r\n                    {\r\n                        \"Modulo\": \"NFCe\"\r\n                    },\r\n                    {\r\n                        \"Modulo\": \"NFSe\"\r\n                    }\r\n                ],\r\n                \"ParametrosExt\": {\r\n                    \"NumeroConsultasDiarias\": 0,\r\n                    \"ArmazenarEveTerceiros\": \"\",\r\n                    \"BaixarDocsNaoArmazenados\": \"\",\r\n                    \"CienciaAutomatica\": 0,\r\n                    \"Justificativa\": \"\",\r\n                    \"ValorMin\": \"0.00\",\r\n                    \"Email\": \"\",\r\n                    \"TipoNotificacao\": 0,\r\n                    \"BuscarDocsRetroativos\": \"\",\r\n                    \"MunicipiosAtivar\": \"\",\r\n                    \"MunicipiosDesativar\": \"\"\r\n                },\r\n                \"Ativar\": \"S\"\r\n            }\r\n        ]\r\n    }\r\n]";
+        private string EmpresaEditar = "[{\n  \"EmpNomFantasia\": \"Empresa de teste de cadastro 2\",\n  \"EmpApelido\": \"Empresa de teste de cadastro 2\",\n  \"EmpRazSocial\": \"Empresa de teste de cadastro 2\",\n  \"EmpCNPJ\": \"29487402000144\",\n  \"EmpIE\": \"342423324\",\n  \"EmpTelefone\": \"234234\",\n  \"EmpEndereco\": \"Rua legal\",\n  \"EmpNumero\": \"1\",\n  \"EmpBairro\": \"Centro\",\n  \"EmpCEP\": \"98918000\",\n  \"EmpComplemento\": \"Perto de outra rua\",\n  \"EmpIM\": \"664544654\",\n  \"MunCodigo\": 4321808,\n  \"EmpCNAE\": \"\",\n  \"EmpCRT\": 0,\n  \"EmpIEST\": \"\",\n  \"EmpEmail\": \"\",\n  \"EmpObservacao\": \"\",\n  \"EmpTpoEndereco\": \"\"\n}]";
+        private string EmpresaLicenciamento = "[\n    {\n        \"CnpjEmpresa\": \"70049563000150\",\n        \"tpAmb\": 2,\n        \"Acao\": 3,\n        \"Modulo\": \"CTe\",\n        \"Modelo\": 2,\n        \"Autor\": \"Empresa 1\"\n    }\n]";
 
         public Source()
         {   
@@ -87,9 +88,10 @@ namespace ConsoleUI.Services
                 { "Inutilizacao" , NFeInutilizacao },
                 { "exportdocuments" , NFeExport},
                 { "exportdocumentsv2" , NFeExportV2},
-                { "Substituicao" , NFeEnvio},
+                { "Substituicao" , NFeSubstituicao},
                 { "Evento" , NFeEvento},
-                { "Importacao" , NFeImportacao}
+                { "Importacao" , NFeImportacao},
+                { "Consulta" , null}
             };
             NFCe = new Dictionary<string, string>()
             {
@@ -101,7 +103,7 @@ namespace ConsoleUI.Services
                 { "Inutilizacao" , NFCeInutilizacao },
                 { "exportdocuments" , NFCeExport},
                 { "exportdocumentsv2" , NFCeExportV2},
-                { "Substituicao" , NFCeEnvio},
+                { "Substituicao" , NFCeSubstituicao},
                 { "Evento" , NFCeEvento},
                 { "Importacao" , NFCeImportacao}
             };
@@ -115,7 +117,7 @@ namespace ConsoleUI.Services
                 { "Inutilizacao" , MDFeInutilizacao },
                 { "exportdocuments" , MDFeExport},
                 { "exportdocumentsv2" , MDFeExportV2},
-                { "Substituicao" , MDFeEnvio},
+                { "Substituicao" , MDFeSubstituicao},
                 { "Evento" , MDFeEvento},
                 { "Importacao" , MDFeImportacao}
             };
@@ -129,7 +131,7 @@ namespace ConsoleUI.Services
                 { "Inutilizacao" , NFSeInutilizacao },
                 { "exportdocuments" , NFSeExport},
                 { "exportdocumentsv2" , NFSeExportV2},
-                { "Substituicao" , NFSeEnvio},
+                { "Substituicao" , NFSeSubstituicao},
                 { "Evento" , NFSeEvento},
                 { "Importacao" , NFSeImportacao}
             };
@@ -143,9 +145,18 @@ namespace ConsoleUI.Services
                 { "Inutilizacao" , CTeInutilizacao },
                 { "exportdocuments" , CTeExport},
                 { "exportdocumentsv2" , CTeExportV2},
-                { "Substituicao" , CTeEnvio},
+                { "Substituicao" , CTeSubstituicao},
                 { "Evento" , CTeEvento},
                 { "Importacao" , CTeImportacao}
+            };
+
+            Dados = new Dictionary<string, Dictionary<string, string>>()
+            {
+                {"NFe", NFe },
+                {"NFCe", NFCe },
+                {"MDFe", MDFe },
+                {"NFSe", NFSe },
+                {"CTe", CTe }
             };
 
             Empresa = new Dictionary<string, string>()
