+++ Exemplos de Codigo-fonte/CSharp/IntegracaoInvoicy/ConsoleUI/packages.config
@@ -5,6 +5,7 @@
   <package id="Microsoft.IdentityModel.JsonWebTokens" version="6.30.1" targetFramework="net48" />
   <package id="Microsoft.IdentityModel.Logging" version="6.30.1" targetFramework="net48" />
   <package id="Microsoft.IdentityModel.Tokens" version="6.30.1" targetFramework="net48" />
+  <package id="Newtonsoft.Json" version="13.0.3" targetFramework="net48" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net48" />
   <package id="System.IdentityModel.Tokens.Jwt" version="6.30.1" targetFramework="net48" />
   <package id="System.Memory" version="4.5.5" targetFramework="net48" />
